refactor(bindings): combine helpers.ts and security.ts in bindings.ts
