+++ docs/assets/main.js
@@ -1,18 +1,16 @@
-var Q6=Object.getPrototypeOf(async()=>{});function s(f){return typeof f==="function"}function i0(f){return s(f)&&Object.getPrototypeOf(f)===Q6}function S5(f){return s(f)&&Object.getPrototypeOf(f)!==Q6}function Z0(f,m){return f!=null&&f[Symbol.toStringTag]===m}function W0(f){return f!==null&&typeof f==="object"&&Object.getPrototypeOf(f)===Object.prototype}function B5(f){return typeof f==="string"?`"${f}"`:!!f&&typeof f==="object"?JSON.stringify(f):String(f)}class d5 extends Error{constructor(f){super(`[${f}] Circular dependency detected`);this.name="CircularDependencyError"}}class bf extends TypeError{constructor(f){super(`[${f}] Signal value cannot be null or undefined`);this.name="NullishSignalValueError"}}class h0 extends Error{constructor(f){super(`[${f}] Signal value is unset`);this.name="UnsetSignalValueError"}}class k5 extends TypeError{constructor(f,m){super(`[${f}] Signal value ${B5(m)} is invalid`);this.name="InvalidSignalValueError"}}class Qf extends TypeError{constructor(f,m){super(`[${f}] Callback ${B5(m)} is invalid`);this.name="InvalidCallbackError"}}class v5 extends Error{constructor(f){super(`[${f}] Signal is read-only`);this.name="ReadonlySignalError"}}class o5 extends Error{constructor(f){super(`[${f}] Active owner is required`);this.name="RequiredOwnerError"}}class X5 extends Error{constructor(f,m,M){super(`[${f}] Could not add key "${m}"${M?` with value ${JSON.stringify(M)}`:""} because it already exists`);this.name="DuplicateKeyError"}}function n(f,m,M){if(m==null)throw new bf(f);if(M&&!M(m))throw new k5(f,m)}function L5(f,m){if(m==null)throw new h0(f)}function x0(f,m,M=s){if(!M(m))throw new Qf(f,m)}var S0="State",d0="Memo",k0="Task",t0="Sensor",b0="List",r5="Collection",v0="Store",o0="Slot",L0=0,J5=1,g=2,n5=4,P0=8,S=null,G0=null,Uf=[],h=0,m1=!1,q0=(f,m)=>f===m;var M1=(f,m)=>{if(Object.is(f,m))return!0;if(typeof f!==typeof m)return!1;if(f==null||typeof f!=="object"||m==null||typeof m!=="object")return!1;let M=Array.isArray(f);if(M!==Array.isArray(m))return!1;if(M){let $=f,H=m;if($.length!==H.length)return!1;for(let Z=0;Z<$.length;Z++)if(!M1($[Z],H[Z]))return!1;return!0}if(W0(f)&&W0(m)){let $=Object.keys(f);if($.length!==Object.keys(m).length)return!1;for(let H of $){if(!(H in m))return!1;if(!M1(f[H],m[H]))return!1}return!0}return!1},r0=(f,m)=>M1(f,m);function k2(f,m){let M=m.sourcesTail;if(M){let $=m.sources;while($){if($===f)return!0;if($===M)break;$=$.nextSource}}return!1}function N0(f,m){let M=m.sourcesTail;if(M?.source===f)return;let $=null,H=m.flags&n5;if(H){if($=M?M.nextSource:m.sources,$?.source===f){m.sourcesTail=$;return}}let Z=f.sinksTail;if(Z?.sink===m&&(!H||k2(Z,m)))return;let z={source:f,sink:m,nextSource:$,prevSink:Z,nextSink:null};if(m.sourcesTail=f.sinksTail=z,M)M.nextSource=z;else m.sources=z;if(Z)Z.nextSink=z;else f.sinks=z}function v2(f){let{source:m,nextSource:M,nextSink:$,prevSink:H}=f;if($)$.prevSink=H;else m.sinksTail=H;if(H)H.nextSink=$;else m.sinks=$;if(!m.sinks){if(m.stop)m.stop(),m.stop=void 0;if("sources"in m&&m.sources){let Z=m;Z.sourcesTail=null,R5(Z),Z.flags|=g}}return M}function R5(f){let m=f.sourcesTail,M=m?m.nextSource:f.sources;while(M)M=v2(M);if(m)m.nextSource=null;else f.sources=null}function c(f,m=g){let M=f.flags;if("sinks"in f){if((M&(g|J5))>=m)return;if(f.flags=M|m,"controller"in f&&f.controller)f.controller.abort(),f.controller=void 0;for(let $=f.sinks;$;$=$.nextSink)c($.sink,J5)}else{if((M&(g|J5))>=m)return;let $=M&(g|J5);if(f.flags=m,!$)Uf.push(f)}}function O0(f,m){if(f.equals(f.value,m))return;f.value=m;for(let M=f.sinks;M;M=M.nextSink)c(M.sink);if(h===0)a()}function W5(f,m){if(!f.cleanup)f.cleanup=m;else if(Array.isArray(f.cleanup))f.cleanup.push(m);else f.cleanup=[f.cleanup,m]}function jf(f){if(!f.cleanup)return;if(Array.isArray(f.cleanup))for(let m=0;m<f.cleanup.length;m++)f.cleanup[m]();else f.cleanup();f.cleanup=null}function o2(f){let m=S;S=f,f.sourcesTail=null,f.flags=n5;let M=!1;try{let $=f.fn(f.value);if(f.error||!f.equals($,f.value))f.value=$,f.error=void 0,M=!0}catch($){M=!0,f.error=$ instanceof Error?$:Error(String($))}finally{S=m,R5(f)}if(M){for(let $=f.sinks;$;$=$.nextSink)if($.sink.flags&J5)$.sink.flags|=g}f.flags=L0}function r2(f){f.controller?.abort();let m=new AbortController;f.controller=m,f.error=void 0;let M=S;S=f,f.sourcesTail=null,f.flags=n5;let $;try{$=f.fn(f.value,m.signal)}catch(H){f.controller=void 0,f.error=H instanceof Error?H:Error(String(H));return}finally{S=M,R5(f)}O0(f.pendingNode,!0),$.then((H)=>{if(m.signal.aborted)return;f.controller=void 0,i(()=>{if(f.error||!f.equals(H,f.value)){f.value=H,f.error=void 0;for(let Z=f.sinks;Z;Z=Z.nextSink)c(Z.sink)}O0(f.pendingNode,!1)})},(H)=>{if(m.signal.aborted)return;f.controller=void 0;let Z=H instanceof Error?H:Error(String(H));i(()=>{if(!f.error||Z.name!==f.error.name||Z.message!==f.error.message){f.error=Z;for(let z=f.sinks;z;z=z.nextSink)c(z.sink)}O0(f.pendingNode,!1)})}),f.flags=L0}function $1(f){jf(f);let m=S,M=G0;S=G0=f,f.sourcesTail=null,f.flags=n5;try{let $=f.fn();if(typeof $==="function")W5(f,$)}finally{S=m,G0=M,R5(f)}f.flags=L0}function $0(f){if(f.flags&J5)for(let m=f.sources;m;m=m.nextSource){if("fn"in m.source)$0(m.source);if(f.flags&g)break}if(f.flags&n5)throw new d5("controller"in f?k0:("value"in f)?d0:"Effect");if(f.flags&g)if("controller"in f)r2(f);else if("value"in f)o2(f);else $1(f);else f.flags=L0}function a(){if(m1)return;m1=!0;try{for(let f=0;f<Uf.length;f++){let m=Uf[f];if(m.flags&(g|J5))$0(m)}Uf.length=0}finally{m1=!1}}function i(f){h++;try{f()}finally{if(h--,h===0)a()}}function Q0(f){let m=S;S=null;try{return f()}finally{S=m}}function U0(f,m){let M=G0,$={cleanup:null};G0=$;let H=()=>jf($);try{let Z=f();if(typeof Z==="function")W5($,Z);return H}finally{if(G0=M,!m?.root&&M)W5(M,H)}}function A0(f,m){return m?()=>{if(S){if(!f.sinks)f.stop=m();N0(f,S)}}:()=>{if(S)N0(f,S)}}function k(f,m){n(S0,f,m?.guard);let M={value:f,sinks:null,sinksTail:null,equals:m?.equals??q0,guard:m?.guard};return{[Symbol.toStringTag]:S0,get(){if(S)N0(M,S);return M.value},set($){n(S0,$,M.guard),O0(M,$)},update($){x0(S0,$);let H=$(M.value);n(S0,H,M.guard),O0(M,H)}}}function l5(f){return Z0(f,S0)}function Pf(f,m){if(f.length!==m.length)return!1;for(let M=0;M<f.length;M++)if(f[M]!==m[M])return!1;return!0}function U6(f){let m=0,M=typeof f==="function";return[typeof f==="string"?()=>`${f}${m++}`:M?($)=>f($)||String(m++):()=>String(m++),M]}function n2(f,m,M,$,H,Z){let z={},X={},J={},B=[],W=!1,Q=new Map;for(let b=0;b<f.length;b++){let U=M[b],p=f[b];if(U&&p!==void 0)Q.set(U,p)}let j=new Set;for(let b=0;b<m.length;b++){let U=m[b];if(U===void 0)continue;let p=H?$(U):M[b]??$(U);if(j.has(p))throw new X5(b0,p,U);if(B.push(p),j.add(p),!Q.has(p))z[p]=U,W=!0;else if(!Z(Q.get(p),U))X[p]=U,W=!0}for(let[b]of Q)if(!j.has(b))J[b]=null,W=!0;if(!W&&!Pf(M,B))W=!0;return{add:z,change:X,remove:J,newKeys:B,changed:W}}function n0(f,m){n(b0,f,Array.isArray);let M=new Map,$=[],[H,Z]=U6(m?.keyConfig),z=m?.itemEquals??r0,X=m?.createItem??((b)=>k(b,{equals:z})),J=()=>$.map((b)=>M.get(b)?.get()).filter((b)=>b!==void 0),B={fn:J,value:f,flags:g,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:r0,error:void 0},W=(b)=>{let U=!1;for(let p in b.add){let N=b.add[p];n(`${b0} item for key "${p}"`,N),M.set(p,X(N)),U=!0}if(Object.keys(b.change).length)i(()=>{for(let p in b.change){let N=b.change[p];n(`${b0} item for key "${p}"`,N);let K=M.get(p);if(K)K.set(N)}});for(let p in b.remove){M.delete(p);let N=$.indexOf(p);if(N!==-1)$.splice(N,1);U=!0}if(U)B.flags|=P0;return b.changed},Q=A0(B,m?.watched);for(let b=0;b<f.length;b++){let U=f[b];if(U===void 0)continue;let p=$[b];if(!p)p=H(U),$[b]=p;n(`${b0} item for key "${p}"`,U),M.set(p,X(U))}B.value=f,B.flags=0;let j={[Symbol.toStringTag]:b0,[Symbol.isConcatSpreadable]:!0,*[Symbol.iterator](){for(let b of $){let U=M.get(b);if(U)yield U}},get length(){return Q(),$.length},get(){if(Q(),B.sources){if(B.flags){let b=B.flags&P0;if(B.value=Q0(J),b){if(B.flags=g,$0(B),B.error)throw B.error}else B.flags=L0}}else if($0(B),B.error)throw B.error;return B.value},set(b){let U=B.flags&g?J():B.value,p=n2(U,b,$,H,Z,z);if(p.changed){$=p.newKeys,W(p),B.flags|=g;for(let N=B.sinks;N;N=N.nextSink)c(N.sink);if(h===0)a()}},update(b){j.set(b(j.get()))},at(b){let U=$[b];return U!==void 0?M.get(U):void 0},keys(){return Q(),$.values()},byKey(b){return M.get(b)},keyAt(b){return $[b]},indexOfKey(b){return $.indexOf(b)},add(b){let U=H(b);if(M.has(U))throw new X5(b0,U,b);if(!$.includes(U))$.push(U);n(`${b0} item for key "${U}"`,b),M.set(U,X(b)),B.flags|=g|P0;for(let p=B.sinks;p;p=p.nextSink)c(p.sink);if(h===0)a();return U},remove(b){let U=typeof b==="number"?$[b]:b;if(U===void 0)return;if(M.delete(U)){let N=typeof b==="number"?b:$.indexOf(U);if(N>=0)$.splice(N,1);B.flags|=g|P0;for(let K=B.sinks;K;K=K.nextSink)c(K.sink);if(h===0)a()}},replace(b,U){let p=M.get(b);if(!p)return;if(n(`${b0} item for key "${b}"`,U),z(Q0(()=>p.get()),U))return;p.set(U),B.flags|=g;for(let N=B.sinks;N;N=N.nextSink)c(N.sink);if(h===0)a()},sort(b){let p=$.map((N)=>[N,M.get(N)?.get()]).sort(s(b)?(N,K)=>b(N[1],K[1]):(N,K)=>String(N[1]).localeCompare(String(K[1]))).map(([N])=>N);if(!Pf($,p)){$=p,B.flags|=g;for(let N=B.sinks;N;N=N.nextSink)c(N.sink);if(h===0)a()}},splice(b,U,...p){let N=$.length,K=b<0?Math.max(0,N+b):Math.min(b,N),u=Math.max(0,Math.min(U??Math.max(0,N-Math.max(0,K)),N-K)),o={},D={};for(let I=0;I<u;I++){let r=K+I,d=$[r];if(d){let l=M.get(d);if(l)D[d]=l.get()}}let Y=$.slice(0,K),C={};for(let I of p){let r=H(I);if(r in D)delete D[r],C[r]=I;else if(M.has(r))throw new X5(b0,r,I);else o[r]=I;Y.push(r)}Y.push(...$.slice(K+u));let G=!!(Object.keys(o).length||Object.keys(D).length||Object.keys(C).length);if(G){W({add:o,change:C,remove:D,changed:G}),$=Y,B.flags|=g;for(let I=B.sinks;I;I=I.nextSink)c(I.sink);if(h===0)a()}return Object.values(D)},deriveCollection(b){return H1(j,b)}};return j}function qf(f){return Z0(f,b0)}function v(f,m){if(x0(d0,f,S5),m?.value!==void 0)n(d0,m.value,m?.guard);let M={fn:f,value:m?.value,flags:g,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:m?.equals??q0,error:void 0,stop:void 0},$=m?.watched,H=A0(M,$?()=>$(()=>{if(c(M),h===0)a()}):void 0);return{[Symbol.toStringTag]:d0,get(){if(H(),$0(M),M.error)throw M.error;return L5(d0,M.value),M.value}}}function b5(f){return Z0(f,d0)}function I0(f,m){if(x0(k0,f,i0),m?.value!==void 0)n(k0,m.value,m?.guard);let M={value:!1,sinks:null,sinksTail:null,equals:q0},$={fn:f,value:m?.value,sources:null,sourcesTail:null,sinks:null,sinksTail:null,flags:g,equals:m?.equals??q0,controller:void 0,error:void 0,stop:void 0,pendingNode:M},H=m?.watched,Z=A0($,H?()=>H(()=>{if(c($),h===0)a()}):void 0),z=A0(M);return{[Symbol.toStringTag]:k0,get(){if(Z(),$0($),$.error)throw $.error;return L5(k0,$.value),$.value},isPending(){return z(),$.pendingNode.value},abort(){$.controller?.abort(),$.controller=void 0,O0($.pendingNode,!1)}}}function D5(f){return Z0(f,k0)}function H1(f,m){x0(r5,m);let M=i0(m),$=new Map,H=[],Z=(b)=>{let U=M?I0(async(p,N)=>{let K=f.byKey(b)?.get();if(K==null)return p;return m(K,N)}):v(()=>{let p=f.byKey(b)?.get();if(p==null)return;return m(p)});$.set(b,U)};function z(b){if(!Pf(H,b)){let U=new Set(H),p=new Set(b);for(let N of H)if(!p.has(N))$.delete(N);for(let N of b)if(!U.has(N))Z(N);H=b,B.flags|=P0}}function X(){z(Array.from(f.keys()));let b=[];for(let U of H)try{let p=$.get(U)?.get();if(p!=null)b.push(p)}catch(p){if(!(p instanceof h0))throw p}return b}let B={fn:X,value:[],flags:g,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:(b,U)=>{if(b.length!==U.length)return!1;for(let p=0;p<b.length;p++)if(b[p]!==U[p])return!1;return!0},error:void 0};function W(){if(B.sources){if(B.flags)if(B.value=Q0(X),B.flags&P0){if(B.flags=g,$0(B),B.error)throw B.error}else B.flags=L0}else if(B.sinks){if($0(B),B.error)throw B.error}else B.value=Q0(X)}let Q=Array.from(Q0(()=>f.keys()));for(let b of Q)Z(b);H=Q;let j={[Symbol.toStringTag]:r5,[Symbol.isConcatSpreadable]:!0,*[Symbol.iterator](){for(let b of H){let U=$.get(b);if(U)yield U}},get length(){if(S)N0(B,S);return W(),H.length},keys(){if(S)N0(B,S);return W(),H.values()},get(){if(S)N0(B,S);return W(),B.value},at(b){let U=H[b];return U!==void 0?$.get(U):void 0},byKey(b){return $.get(b)},keyAt(b){return H[b]},indexOfKey(b){return H.indexOf(b)},deriveCollection(b){return H1(j,b)}};return j}function R0(f){x0("Effect",f);let m={fn:f,flags:g,sources:null,sourcesTail:null,cleanup:null},M=()=>{jf(m),m.fn=void 0,m.flags=L0,m.sourcesTail=null,R5(m)};if(G0)W5(G0,M);return $1(m),M}function V5(f,m){if(!G0)throw new o5("match");let M=!Array.isArray(f),$=M?[f]:f,{nil:H,stale:Z}=m,z=M?(j)=>m.ok(j[0]):(j)=>m.ok(j),X=M&&m.err?(j)=>m.err(j[0]):m.err??console.error,J,B=!1,W=Array($.length);for(let j=0;j<$.length;j++)try{W[j]=$[j].get()}catch(b){if(b instanceof h0){B=!0;continue}if(!J)J=[];J.push(b instanceof Error?b:Error(String(b)))}let Q;try{if(B)Q=H?.();else if(J)Q=X(J);else if(Z&&(M?D5($[0])&&$[0].isPending():$.some((j)=>D5(j)&&j.isPending())))Q=Z();else Q=z(W)}catch(j){Q=X([j instanceof Error?j:Error(String(j))])}if(typeof Q==="function")return Q;if(Q instanceof Promise){let j=G0,b=new AbortController;W5(j,()=>b.abort()),Q.then((U)=>{if(!b.signal.aborted&&typeof U==="function")W5(j,U)}).catch((U)=>{X([U instanceof Error?U:Error(String(U))])})}}function Q5(f,m){if(x0(t0,f,S5),m?.value!==void 0)n(t0,m.value,m?.guard);let M={value:m?.value,sinks:null,sinksTail:null,equals:m?.equals??q0,guard:m?.guard,stop:void 0};return{[Symbol.toStringTag]:t0,get(){if(S){if(!M.sinks)M.stop=f(($)=>{n(t0,$,M.guard),O0(M,$)});N0(M,S)}return L5(t0,M.value),M.value}}}function l2(f,m){let M={},$={},H={},Z=!1,z=Object.keys(f),X=Object.keys(m);for(let J of X)if(J in f){if(!r0(f[J],m[J]))$[J]=m[J],Z=!0}else M[J]=m[J],Z=!0;for(let J of z)if(!(J in m))H[J]=void 0,Z=!0;return{add:M,change:$,remove:H,changed:Z}}function Y5(f,m){n(v0,f,W0);let M=new Map,$=(B,W)=>{if(n(`${v0} for key "${B}"`,W),Array.isArray(W))M.set(B,n0(W));else if(W0(W))M.set(B,Y5(W));else M.set(B,k(W))},H=()=>{let B={};return M.forEach((W,Q)=>{B[Q]=W.get()}),B},Z={fn:H,value:f,flags:g,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:r0,error:void 0},z=(B)=>{let W=!1;for(let Q in B.add)$(Q,B.add[Q]),W=!0;if(Object.keys(B.change).length)i(()=>{for(let Q in B.change){let j=B.change[Q];n(`${v0} for key "${Q}"`,j);let b=M.get(Q);if(b)if(W0(j)!==c5(b))$(Q,j),W=!0;else b.set(j)}});for(let Q in B.remove)M.delete(Q),W=!0;if(W)Z.flags|=P0;return B.changed},X=A0(Z,m?.watched);for(let B of Object.keys(f))$(B,f[B]);let J={[Symbol.toStringTag]:v0,[Symbol.isConcatSpreadable]:!1,*[Symbol.iterator](){for(let B of Array.from(M.keys())){let W=M.get(B);if(W)yield[B,W]}},keys(){return X(),M.keys()},byKey(B){return M.get(B)},get(){if(X(),Z.sources){if(Z.flags){let B=Z.flags&P0;if(Z.value=Q0(H),B){if(Z.flags=g,$0(Z),Z.error)throw Z.error}else Z.flags=L0}}else if($0(Z),Z.error)throw Z.error;return Z.value},set(B){let W=Z.flags&g?H():Z.value,Q=l2(W,B);if(z(Q)){Z.flags|=g;for(let j=Z.sinks;j;j=j.nextSink)c(j.sink);if(h===0)a()}},update(B){J.set(B(J.get()))},add(B,W){if(M.has(B))throw new X5(v0,B,W);$(B,W),Z.flags|=g|P0;for(let Q=Z.sinks;Q;Q=Q.nextSink)c(Q.sink);if(h===0)a();return B},remove(B){if(M.delete(B)){Z.flags|=g|P0;for(let Q=Z.sinks;Q;Q=Q.nextSink)c(Q.sink);if(h===0)a()}}};return new Proxy(J,{get(B,W){if(W in B)return Reflect.get(B,W);if(typeof W!=="symbol")return B.byKey(W)},has(B,W){if(W in B)return!0;return B.byKey(String(W))!==void 0},ownKeys(B){return Array.from(B.keys())},getOwnPropertyDescriptor(B,W){if(W in B)return Reflect.getOwnPropertyDescriptor(B,W);if(typeof W==="symbol")return;let Q=B.byKey(String(W));return Q?{enumerable:!0,configurable:!0,writable:!0,value:Q}:void 0}})}function c5(f){return Z0(f,v0)}var c2=new Set([S0,d0,k0,t0,o0,b0,r5,v0]);function G5(f,m){return i0(f)?I0(f,m):v(f,m)}function F5(f){return f!=null&&c2.has(f[Symbol.toStringTag])}function pf(f){return l5(f)||c5(f)||qf(f)}function P6(f){if(F5(f))return!0;return f!==null&&typeof f==="object"&&"get"in f&&typeof f.get==="function"}function xf(f,m){n(o0,f,P6);let M=f,$=m?.guard,H={fn:()=>M.get(),value:void 0,flags:g,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:m?.equals??q0,error:void 0},Z=()=>{if(S)N0(H,S);if($0(H),H.error)throw H.error;return H.value},z=(J)=>{if(U5(M))return M.set(J);if("set"in M&&typeof M.set==="function")n(o0,J,$),M.set(J);else throw new v5(o0)},X=(J)=>{n(o0,J,P6),M=J,H.flags|=g;for(let B=H.sinks;B;B=B.nextSink)c(B.sink);if(h===0)a()};return{[Symbol.toStringTag]:o0,configurable:!0,enumerable:!0,get:Z,set:z,replace:X,current:()=>M}}function U5(f){return Z0(f,o0)}var Nf="context-request";class Z1 extends Event{context;callback;subscribe;constructor(f,m,M=!1){super(Nf,{bubbles:!0,composed:!0});this.context=f,this.callback=m,this.subscribe=M}}var q6=(f)=>(m)=>()=>U0(()=>{let M=($)=>{let{context:H,callback:Z}=$;if(typeof H==="string"&&m.includes(H)&&s(Z))$.stopImmediatePropagation(),Z(()=>f[H])};return f.addEventListener(Nf,M),()=>f.removeEventListener(Nf,M)}),p6=(f)=>(m,M)=>{let $=()=>M;return f.dispatchEvent(new Z1(m,(H)=>{$=H})),v($)};var j5=typeof process<"u"&&!1,P5="warn",m8=(f)=>f?`#${f}`:"",M8=(f)=>f?.length?`.${Array.from(f).join(".")}`:"",z1=(f)=>f.localName.includes("-"),B1=(f)=>z1(f)&&f.matches(":not(:defined)"),F0=(f)=>f?`<${f.localName}${m8(f.id)}${M8(f.classList)}>`:"<unknown>";class Lf extends TypeError{constructor(f){super(`Invalid component name "${f}". Custom element names must contain a hyphen, start with a lowercase letter, and contain only lowercase letters, numbers, and hyphens.`);this.name="InvalidComponentNameError"}}class i5 extends Error{constructor(f,m,M){super(`Missing required element <${m}> in component ${F0(f)}. ${M}`);this.name="MissingElementError"}}class Rf extends Error{constructor(f,m){super(`Timeout waiting for: [${m.join(", ")}] in component ${F0(f)}.`);this.name="DependencyTimeoutError"}}class Df extends TypeError{constructor(f,m,M){super(`Expected reactives passed from ${F0(f)} to ${F0(m)} to be a record of signals, reactive property names or functions. Got ${B5(M)}.`);this.name="InvalidReactivesError"}}class Vf extends TypeError{constructor(f,m){super(`Target ${F0(f)} is not a custom element in ${m}.`);this.name="InvalidCustomElementError"}}var x6=new WeakMap,h5=(f)=>{let m=x6.get(f);if(!m)m={},x6.set(f,m);return m};var Yf=(f)=>{for(let m of f)if(Array.isArray(m))Yf(m);else if(m)m()},X1=(f,m)=>{if(s(m))return G5(m);if(typeof m==="string"){let M=h5(f)[m];if(M)return M;return v(()=>f[m])}if(m&&typeof m==="object"&&"get"in m&&!(Symbol.toStringTag in m))return m;return m},N6=(f)=>{function m(M,$){return()=>{if(Array.isArray(M)){let Z=M.map((X)=>X1(f,X)),z=$;return R0(()=>V5(Z,{ok:(X)=>Q0(()=>z(X))}))}let H=X1(f,M);if(typeof $==="function")return R0(()=>V5(H,{ok:(Z)=>Q0(()=>$(Z))}));return R0(()=>V5(H,$))}}return m},L6=(f)=>{let m=($,H)=>U0(()=>{if(!z1($))throw new Vf($,`pass from ${F0(f)}`);if(!W0(H))throw new Df(f,$,H);let Z=h5($),z=F0($),X=[];for(let[J,B]of Object.entries(H)){if(B==null)continue;if(!(J in $)){if(j5)console[P5](`pass(): property '${J}' does not exist on ${z}`);continue}let W=X1(f,B);if(!W)continue;let Q=Z[J];if(U5(Q)){let j=Q.current();Q.replace(W),X.push(()=>Q.replace(j));continue}if(j5)console[P5](`pass(): property '${J}' on ${z} is not Slot-backed — use setProperty() for non-Le Truc elements`)}if(X.length)return()=>{for(let J of X)J()}});function M($,H){return()=>{if(!$)return;if(b5($))R0(()=>{for(let Z of $.get())U0(()=>m(Z,H))});else m($,H)}}return M};function t(f,m){return()=>{R0(()=>{for(let M of f.get())U0(()=>{let $=m(M);if(Array.isArray($))Yf($);else if(typeof $==="function")$()})})}}var J1=new Set,R6=new WeakMap,Gf=new Set,W1,$8=()=>{W1=void 0;let f=Array.from(J1);J1.clear();for(let M of f)R6.get(M)?.();let m=Array.from(Gf);Gf.clear();for(let M of m)M()},D6=()=>{if(!W1)W1=requestAnimationFrame($8)},t5=(f,m)=>{R6.set(f,m),J1.add(f),D6()},a0=(f,m)=>{let M=!1,$,H=()=>{M=!1,f(...$)},Z=(...z)=>{if($=z,M)return;M=!0,Gf.add(H),D6()};return Z.cancel=()=>{if(M)Gf.delete(H),M=!1},m?.addEventListener("abort",Z.cancel,{once:!0}),Z};var H8=new Set(["scroll","resize","mousewheel","touchstart","touchmove","wheel"]),Z8=new Set(["focus","blur","scroll","resize","load","unload","error","toggle","mouseenter","mouseleave","pointerenter","pointerleave","abort","canplay","canplaythrough","durationchange","emptied","ended","loadeddata","loadedmetadata","loadstart","pause","play","playing","progress","ratechange","seeked","seeking","stalled","suspend","timeupdate","volumechange","waiting"]),V6=(f,m,M,$,H)=>{let Z=(X)=>{let J=$(X,m);if(!W0(J))return;i(()=>{for(let[B,W]of Object.entries(J))f[B]=W})},z=H.passive?a0(Z):Z;return m.addEventListener(M,z,H),()=>{m.removeEventListener(M,z),z.cancel?.()}},Y6=(f)=>{function m(M,$,H,Z={}){return()=>{if(!M)return;if(!("passive"in Z))Z={...Z,passive:H8.has($)};if(b5(M)){if(Z8.has($)){if(j5)console[P5](`on(): '${$}' does not bubble — prefer each() + on() for per-element listeners in ${F0(f)}`);return R0(()=>{for(let B of M.get())U0(()=>{return V6(f,B,$,H,Z)})})}let z=f.shadowRoot??f,X=(B)=>{let W=B.composedPath();for(let Q of M.get())if(W.includes(Q)){let j=H(B,Q);if(!W0(j))break;i(()=>{for(let[b,U]of Object.entries(j))f[b]=U});break}},J=Z.passive?a0(X):X;U0(()=>{return z.addEventListener($,J,Z),()=>{z.removeEventListener($,J),J.cancel?.()}});return}U0(()=>V6(f,M,$,H,Z))}}return m};var z8=200,B8=(f)=>{let m=new Set,M="",$=0;for(let H of f)if(H==="[")$++;else if(H==="]"){if($>0)$--}else if($===0)M+=H;if(M.includes("."))m.add("class");if(M.includes("#"))m.add("id");if(f.includes("[")){let H=f.split("[");for(let Z=1;Z<H.length;Z++){let z=H[Z];if(!z||!z.includes("]"))continue;let X=z.split("=")[0].split("]")[0].trim().replace(/[^a-zA-Z0-9_-]/g,"");if(X)m.add(X)}}return[...m]};function Ff(f,m){return v(()=>Array.from(f.querySelectorAll(m)),{value:[],equals:(M,$)=>M.length===$.length&&M.every((H,Z)=>H===$[Z]),watched:(M)=>{let $={childList:!0,subtree:!0},H=B8(m);if(H.length)$.attributes=!0,$.attributeFilter=H;let Z=(J)=>J instanceof Element&&(J.matches(m)||J.querySelector(m)),z=(J)=>{if(J.type==="attributes")return!0;if(J.type==="childList")return Array.from(J.addedNodes).some(Z)||Array.from(J.removedNodes).some(Z);return!1},X=new MutationObserver((J)=>{for(let B of J)if(z(B)){M();return}});return X.observe(f,$),()=>X.disconnect()}})}var G6=(f)=>{let m=f.shadowRoot??f,M=new Set;function $(z,X){let J=m.querySelector(z);if(X!=null&&!J)throw new i5(f,z,X);if(J&&B1(J))M.add(J.localName);return J??void 0}function H(z,X){let J=Ff(m,z),B=J.get();if(X!=null&&!B.length)throw new i5(f,z,X);if(B.length){for(let W of B)if(B1(W))M.add(W.localName)}return J}return[{first:$,all:H},(z)=>{if(M.size)queueMicrotask(()=>{let X=Array.from(M).filter((J)=>!customElements.get(J));if(!X.length){z();return}Promise.race([Promise.all(X.map((J)=>customElements.whenDefined(J))),new Promise((J,B)=>{setTimeout(()=>{B(new Rf(f,X.filter((W)=>!customElements.get(W))))},z8)})]).then(z).catch((J)=>{if(j5)console[P5](J);z()})});else z()}]};var F6=Symbol("parser"),C6=Symbol("method"),K6=(f)=>s(f)&&(F6 in f),O6=(f)=>s(f)&&(C6 in f),D0=(f)=>Object.assign(f,{[F6]:!0}),s0=(f)=>Object.assign(f,{[C6]:!0});function R(f,m){if(!f.includes("-")||!f.match(/^[a-z][a-z0-9-]*$/))throw new Lf(f);class M extends HTMLElement{debug;#M=!1;#f=[];#m;connectedCallback(){let $=()=>{this.#m=U0(()=>Yf(this.#f),{root:!0})};if(this.#M)$();else{let H=this,[Z,z]=G6(this),X={expose:this.#H.bind(this),host:H,...Z,watch:N6(H),on:Y6(H),pass:L6(H),provideContexts:q6(H),requestContext:p6(H)},J=m(X);if(J)this.#f=J;if(this.#M=!0,!this.#f.length)return;z($)}}disconnectedCallback(){if(s(this.#m))this.#m()}#H($){let H=(Z,z)=>{if(K6(z)){let X=z(this.getAttribute(Z));if(X!=null)this.#$(Z,X)}else if(O6(z))this[Z]=z;else{let X=z;if(X!=null)this.#$(Z,X)}};for(let[Z,z]of Object.entries($)){if(z==null||Z in this)continue;H(Z,z)}}#$($,H){let Z=F5(H)?H:s(H)?G5(H):k(H),z=h5(this),X=$,J=z[X];if(U5(J))J.replace(Z);else if(pf(Z)){let B=xf(Z);z[X]=B,Object.defineProperty(this,$,B)}else z[X]=Z,Object.defineProperty(this,$,{get:Z.get,enumerable:!0})}}return customElements.define(f,M),customElements.get(f)}var X8=(f)=>{if(/^(javascript|data|vbscript):/i.test(f))return!1;if(/^(mailto|tel):/i.test(f))return!0;if(f.includes("://"))try{let m=new URL(f);return["http:","https:","ftp:"].includes(m.protocol)}catch{return!1}return!0},C5=(f,m,M)=>{if(/^on/i.test(m))throw Error(`setAttribute: blocked unsafe attribute name '${m}' on ${f.localName} — event handler attributes are not allowed`);if(M=String(M).trim(),!X8(M))throw Error(`setAttribute: blocked unsafe value for '${m}' on <${f.localName}>: '${M}'`);f.setAttribute(m,M)},V0=(f)=>f.replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;").replace(/"/g,"&quot;").replace(/'/g,"&#39;"),b1=(f,m)=>{Array.from(f.childNodes).filter((M)=>M.nodeType!==Node.COMMENT_NODE).forEach((M)=>{M.remove()}),f.append(document.createTextNode(m))};var J8=["type","src","async","defer","nomodule","crossorigin","integrity","referrerpolicy","fetchpriority"],w=(f,m=!1)=>m?(M)=>b1(f,String(M)):(M)=>{f.textContent=String(M)},H0=(f,m)=>(M)=>{f[m]=M},a5=(f,m)=>(M)=>{f.classList.toggle(m,Boolean(M))},z0=(f)=>(m)=>{f.hidden=!m},q5=(f,m,M=!1)=>({ok:($)=>{if(typeof $==="boolean")f.toggleAttribute(m,$);else if(M)f.setAttribute(m,$);else C5(f,m,$)},nil:()=>{f.removeAttribute(m)}}),e0=(f,m)=>({ok:(M)=>{f.style.setProperty(m,M)},nil:()=>{f.style.removeProperty(m)}}),Q1=(f,m={})=>{let M=()=>{if(f.shadowRoot)f.shadowRoot.innerHTML="<slot></slot>";else f.innerHTML=""};return{ok:($)=>{if(!$){M();return}let{shadowRootMode:H,allowScripts:Z}=m;if(H&&!f.shadowRoot)f.attachShadow({mode:H});let z=f.shadowRoot||f;t5(f,()=>{if(z.innerHTML=$,Z)z.querySelectorAll("script").forEach((X)=>{let J=document.createElement("script");for(let B of J8){let W=X.getAttribute(B);if(W!==null)J.setAttribute(B,W)}if(!X.hasAttribute("src"))J.appendChild(document.createTextNode(X.textContent??""));z.appendChild(J),X.remove()})})},nil:M}};var s5=()=>D0((f)=>f!=null&&f!=="false");var U1=(f)=>D0((m)=>{if((m??f)==null)throw TypeError("asJSON: Value and fallback are both null or undefined");if(m==null)return f;if(m==="")throw SyntaxError("Empty string is not valid JSON");let M;try{M=JSON.parse(m)}catch($){throw SyntaxError(`Failed to parse JSON: ${String($)}`,{cause:$})}return M??f});var j1=(f,m)=>{if(m==null)return;let M=f(m);return Number.isFinite(M)?M:void 0};var p5=(f=0)=>D0((m)=>j1(parseFloat,m)??f),K5=(f=0,m=Number.MAX_SAFE_INTEGER)=>D0((M)=>{if(M==null)return f;let $=M.trim(),H=$.toLowerCase().startsWith("0x")?j1((z)=>parseInt(z,16),$):j1(parseFloat,M),Z=H!=null?Math.trunc(H):f;return Math.max(f,Math.min(Z,m))});var y0=(f="")=>D0((m)=>m??f);var rm=R("basic-button",({expose:f,first:m,watch:M})=>{let $=m("button","Add a native button as descendant."),H=m("span.label"),Z=m("span.badge");return f({disabled:$.disabled,label:H?.textContent??$.textContent??"",badge:Z?.textContent??""}),[M("disabled",H0($,"disabled")),H&&M("label",w(H)),Z&&M("badge",w(Z))]});var cm=R("basic-counter",({expose:f,first:m,host:M,on:$,watch:H})=>{let Z=m("button","Add a native button element to increment the count."),z=m("span","Add a span to display the count.");return f({count:Number.parseInt(z.textContent||"0")}),[$(Z,"click",()=>({count:M.count+1})),H("count",w(z))]});var tm=R("basic-gauge",({expose:f,first:m,host:M,pass:$,watch:H})=>{let Z=m("meter","Add a <meter> element to display the level"),z=m("basic-number","Add a <basic-number> element to display the value"),X=m(".label","Add an element to display the qualification label");f({value:Z.value});let J=U1([])(M.getAttribute("thresholds"));return[$(z,{value:()=>M.value}),H("value",(B)=>{Z.value=B,M.style.setProperty("--basic-gauge-degree",`${240*B/Z.max}deg`)}),H(()=>J.find((B)=>M.value>=B.min)||{label:"",color:"var(--color-primary)"},(B)=>{X.textContent=B.label,M.style.setProperty("--basic-gauge-color",B.color)})]});var em=R("basic-hello",({expose:f,first:m,on:M,watch:$})=>{let H=m("input","Needed to enter the name."),Z=m("output","Needed to display the name."),z=Z.textContent||"";return f({name:Z.textContent??""}),[M(H,"input",()=>({name:H.value||z})),$("name",w(Z))]});function O5(f){return f.closest("[lang]")?.getAttribute("lang")||"en"}function W8(f,m,M={onWarn:console.warn,onError:console.error}){let $=()=>new Intl.NumberFormat(f);if(!m)return $();let{onWarn:H,onError:Z}=M,z={};try{z=JSON.parse(m)}catch(U){return Z?.(`Invalid JSON: ${U}`),$()}let X=z.style??"decimal",J=[];if(X==="currency"){if(!z.currency||typeof z.currency!=="string"||z.currency.length!==3)return Z?.('style="currency" requires a 3-letter ISO currency (e.g. "CHF").'),$()}else J.push("currency","currencyDisplay","currencySign");if(X==="unit"){if(!z.unit||typeof z.unit!=="string")return Z?.('style="unit" requires a "unit" (e.g. "liter", "kilometer-per-hour").'),$()}else J.push("unit","unitDisplay");if(z.notation&&z.notation!=="compact")J.push("compactDisplay");let B={};for(let[U,p]of Object.entries(z))if(!J.includes(U))B[U]=p;else H?.(`Option "${U}" is ignored for style="${X}".`);let{minimumFractionDigits:W,maximumFractionDigits:Q}=B;if(W!=null&&Q!=null&&W>Q)H?.(`minimumFractionDigits (${W}) > maximumFractionDigits (${Q}); swapping.`),B.minimumFractionDigits=Q,B.maximumFractionDigits=W;let{minimumSignificantDigits:j,maximumSignificantDigits:b}=B;if(j!=null&&b!=null&&j>b)H?.(`minimumSignificantDigits (${j}) > maximumSignificantDigits (${b}); swapping.`),B.minimumSignificantDigits=b,B.maximumSignificantDigits=j;try{let U=new Intl.NumberFormat(f,B);if(U.resolvedOptions().locale!==f)H(`Fall back to locale ${U.resolvedOptions().locale} instead of ${f}`);return U}catch(U){return Z?.(`Options rejected by Intl.NumberFormat: ${U instanceof Error?U.message:String(U)}`),$()}}var H7=R("basic-number",({expose:f,host:m,watch:M})=>{let $=W8(O5(m),m.getAttribute("options"));return f({value:p5()}),[M(()=>$.format(m.value),w(m))]});var X7=R("basic-pluralize",({expose:f,first:m,host:M,watch:$})=>{let H=m(".count"),Z=m(".none"),z=m(".some"),X=m(".zero"),J=m(".one"),B=m(".two"),W=m(".few"),Q=m(".many"),j=m(".other"),b=new Intl.PluralRules(O5(M),M.hasAttribute("ordinal")?{type:"ordinal"}:void 0);f({count:K5()});let U={zero:X,one:J,two:B,few:W,many:Q,other:j},p=b.resolvedOptions().pluralCategories;return[H&&$("count",w(H)),Z&&$(()=>M.count===0,z0(Z)),z&&$(()=>M.count!==0,z0(z)),...p.map((N)=>{let K=U[N];return K&&$(()=>b.select(M.count)===N,z0(K))})]});var P1="invalid date",b8="unknown date";function A6(f,m,{dateStyle:M="long"}={}){let[$,H,Z]=m.split("-").map(Number);if(!$||Number.isNaN($)||!H||Number.isNaN(H)||Number.isNaN(Z))return P1;let z=new Date($,H-1,Z);return new Intl.DateTimeFormat(f,{dateStyle:M}).format(z)}var Q7=R("card-blogmeta",({host:f,first:m})=>{let M=m("time.published","Add a <time> element to display the publication date."),$=m("span.modified"),H=m(".modified time"),Z=O5(f);if(M.textContent=M.dateTime?A6(Z,M.dateTime):b8,H){let z=H.dateTime?A6(Z,H.dateTime):P1;if($&&z===P1)$.remove();else H.textContent=z}});var Q8=(f,m)=>{if(typeof f!=="number")return;if(m===3)return{mode:"rgb",r:(f>>8&15|f>>4&240)/255,g:(f>>4&15|f&240)/255,b:(f&15|f<<4&240)/255};if(m===4)return{mode:"rgb",r:(f>>12&15|f>>8&240)/255,g:(f>>8&15|f>>4&240)/255,b:(f>>4&15|f&240)/255,alpha:(f&15|f<<4&240)/255};if(m===6)return{mode:"rgb",r:(f>>16&255)/255,g:(f>>8&255)/255,b:(f&255)/255};if(m===8)return{mode:"rgb",r:(f>>24&255)/255,g:(f>>16&255)/255,b:(f>>8&255)/255,alpha:(f&255)/255}},Cf=Q8;var U8={aliceblue:15792383,antiquewhite:16444375,aqua:65535,aquamarine:8388564,azure:15794175,beige:16119260,bisque:16770244,black:0,blanchedalmond:16772045,blue:255,blueviolet:9055202,brown:10824234,burlywood:14596231,cadetblue:6266528,chartreuse:8388352,chocolate:13789470,coral:16744272,cornflowerblue:6591981,cornsilk:16775388,crimson:14423100,cyan:65535,darkblue:139,darkcyan:35723,darkgoldenrod:12092939,darkgray:11119017,darkgreen:25600,darkgrey:11119017,darkkhaki:12433259,darkmagenta:9109643,darkolivegreen:5597999,darkorange:16747520,darkorchid:10040012,darkred:9109504,darksalmon:15308410,darkseagreen:9419919,darkslateblue:4734347,darkslategray:3100495,darkslategrey:3100495,darkturquoise:52945,darkviolet:9699539,deeppink:16716947,deepskyblue:49151,dimgray:6908265,dimgrey:6908265,dodgerblue:2003199,firebrick:11674146,floralwhite:16775920,forestgreen:2263842,fuchsia:16711935,gainsboro:14474460,ghostwhite:16316671,gold:16766720,goldenrod:14329120,gray:8421504,green:32768,greenyellow:11403055,grey:8421504,honeydew:15794160,hotpink:16738740,indianred:13458524,indigo:4915330,ivory:16777200,khaki:15787660,lavender:15132410,lavenderblush:16773365,lawngreen:8190976,lemonchiffon:16775885,lightblue:11393254,lightcoral:15761536,lightcyan:14745599,lightgoldenrodyellow:16448210,lightgray:13882323,lightgreen:9498256,lightgrey:13882323,lightpink:16758465,lightsalmon:16752762,lightseagreen:2142890,lightskyblue:8900346,lightslategray:7833753,lightslategrey:7833753,lightsteelblue:11584734,lightyellow:16777184,lime:65280,limegreen:3329330,linen:16445670,magenta:16711935,maroon:8388608,mediumaquamarine:6737322,mediumblue:205,mediumorchid:12211667,mediumpurple:9662683,mediumseagreen:3978097,mediumslateblue:8087790,mediumspringgreen:64154,mediumturquoise:4772300,mediumvioletred:13047173,midnightblue:1644912,mintcream:16121850,mistyrose:16770273,moccasin:16770229,navajowhite:16768685,navy:128,oldlace:16643558,olive:8421376,olivedrab:7048739,orange:16753920,orangered:16729344,orchid:14315734,palegoldenrod:15657130,palegreen:10025880,paleturquoise:11529966,palevioletred:14381203,papayawhip:16773077,peachpuff:16767673,peru:13468991,pink:16761035,plum:14524637,powderblue:11591910,purple:8388736,rebeccapurple:6697881,red:16711680,rosybrown:12357519,royalblue:4286945,saddlebrown:9127187,salmon:16416882,sandybrown:16032864,seagreen:3050327,seashell:16774638,sienna:10506797,silver:12632256,skyblue:8900331,slateblue:6970061,slategray:7372944,slategrey:7372944,snow:16775930,springgreen:65407,steelblue:4620980,tan:13808780,teal:32896,thistle:14204888,tomato:16737095,turquoise:4251856,violet:15631086,wheat:16113331,white:16777215,whitesmoke:16119285,yellow:16776960,yellowgreen:10145074},e5=U8;var j8=(f)=>{return Cf(e5[f.toLowerCase()],6)},I6=j8;var P8=/^#?([0-9a-f]{8}|[0-9a-f]{6}|[0-9a-f]{4}|[0-9a-f]{3})$/i,q8=(f)=>{let m;return(m=f.match(P8))?Cf(parseInt(m[1],16),m[1].length):void 0},y6=q8;var Kf="([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?)";var x5="([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?)%";var ff="(?:([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?)%|([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?))";var w6="(?:([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?)(deg|grad|rad|turn)|([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?))";var f5="\\s*,\\s*";var R7=new RegExp("^(?:([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?)%|([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?)|none)$");var p8=new RegExp(`^rgba?\\(\\s*${Kf}${f5}${Kf}${f5}${Kf}\\s*(?:,\\s*${ff}\\s*)?\\)$`),x8=new RegExp(`^rgba?\\(\\s*${x5}${f5}${x5}${f5}${x5}\\s*(?:,\\s*${ff}\\s*)?\\)$`),N8=(f)=>{let m={mode:"rgb"},M;if(M=f.match(p8)){if(M[1]!==void 0)m.r=M[1]/255;if(M[2]!==void 0)m.g=M[2]/255;if(M[3]!==void 0)m.b=M[3]/255}else if(M=f.match(x8)){if(M[1]!==void 0)m.r=M[1]/100;if(M[2]!==void 0)m.g=M[2]/100;if(M[3]!==void 0)m.b=M[3]/100}else return;if(M[4]!==void 0)m.alpha=Math.max(0,Math.min(1,M[4]/100));else if(M[5]!==void 0)m.alpha=Math.max(0,Math.min(1,+M[5]));return m},u6=N8;var L8=(f,m)=>f===void 0?void 0:typeof f!=="object"?T6(f):f.mode!==void 0?f:m?{...f,mode:m}:void 0,m5=L8;var R8=(f="rgb")=>(m)=>(m=m5(m,f))!==void 0?m.mode===f?m:C0[m.mode][f]?C0[m.mode][f](m):f==="rgb"?C0[m.mode].rgb(m):C0.rgb[f](C0[m.mode].rgb(m)):void 0,e=R8;var C0={},g6={},mf=[],q1={},D8=(f)=>f,f0=(f)=>{if(C0[f.mode]={...C0[f.mode],...f.toMode},Object.keys(f.fromMode||{}).forEach((m)=>{if(!C0[m])C0[m]={};C0[m][f.mode]=f.fromMode[m]}),!f.ranges)f.ranges={};if(!f.difference)f.difference={};return f.channels.forEach((m)=>{if(f.ranges[m]===void 0)f.ranges[m]=[0,1];if(!f.interpolate[m])throw Error(`Missing interpolator for: ${m}`);if(typeof f.interpolate[m]==="function")f.interpolate[m]={use:f.interpolate[m]};if(!f.interpolate[m].fixup)f.interpolate[m].fixup=D8}),g6[f.mode]=f,(f.parse||[]).forEach((m)=>{V8(m,f.mode)}),e(f.mode)},w0=(f)=>g6[f],V8=(f,m)=>{if(typeof f==="string"){if(!m)throw Error("'mode' required when 'parser' is a string");q1[f]=m}else if(typeof f==="function"){if(mf.indexOf(f)<0)mf.push(f)}};var p1=/[^\x00-\x7F]|[a-zA-Z_]/,Y8=/[^\x00-\x7F]|[-\w]/,x={Function:"function",Ident:"ident",Number:"number",Percentage:"percentage",ParenClose:")",None:"none",Hue:"hue",Alpha:"alpha"},A=0;function Of(f){let m=f[A],M=f[A+1];if(m==="-"||m==="+")return/\d/.test(M)||M==="."&&/\d/.test(f[A+2]);if(m===".")return/\d/.test(M);return/\d/.test(m)}function x1(f){if(A>=f.length)return!1;let m=f[A];if(p1.test(m))return!0;if(m==="-"){if(f.length-A<2)return!1;let M=f[A+1];if(M==="-"||p1.test(M))return!0;return!1}return!1}var G8={deg:1,rad:180/Math.PI,grad:0.9,turn:360};function Mf(f){let m="";if(f[A]==="-"||f[A]==="+")m+=f[A++];if(m+=Af(f),f[A]==="."&&/\d/.test(f[A+1]))m+=f[A++]+Af(f);if(f[A]==="e"||f[A]==="E"){if((f[A+1]==="-"||f[A+1]==="+")&&/\d/.test(f[A+2]))m+=f[A++]+f[A++]+Af(f);else if(/\d/.test(f[A+1]))m+=f[A++]+Af(f)}if(x1(f)){let M=If(f);if(M==="deg"||M==="rad"||M==="turn"||M==="grad")return{type:x.Hue,value:m*G8[M]};return}if(f[A]==="%")return A++,{type:x.Percentage,value:+m};return{type:x.Number,value:+m}}function Af(f){let m="";while(/\d/.test(f[A]))m+=f[A++];return m}function If(f){let m="";while(A<f.length&&Y8.test(f[A]))m+=f[A++];return m}function F8(f){let m=If(f);if(f[A]==="(")return A++,{type:x.Function,value:m};if(m==="none")return{type:x.None,value:void 0};return{type:x.Ident,value:m}}function C8(f=""){let m=f.trim(),M=[],$;A=0;while(A<m.length){if($=m[A++],$===`
+var b6=Object.getPrototypeOf(async()=>{});function s(f){return typeof f==="function"}function i0(f){return s(f)&&Object.getPrototypeOf(f)===b6}function _5(f){return s(f)&&Object.getPrototypeOf(f)!==b6}function Z0(f,m){return f!=null&&f[Symbol.toStringTag]===m}function W0(f){return f!==null&&typeof f==="object"&&Object.getPrototypeOf(f)===Object.prototype}function B5(f){return typeof f==="string"?`"${f}"`:!!f&&typeof f==="object"?JSON.stringify(f):String(f)}class S5 extends Error{constructor(f){super(`[${f}] Circular dependency detected`);this.name="CircularDependencyError"}}class bf extends TypeError{constructor(f){super(`[${f}] Signal value cannot be null or undefined`);this.name="NullishSignalValueError"}}class h0 extends Error{constructor(f){super(`[${f}] Signal value is unset`);this.name="UnsetSignalValueError"}}class d5 extends TypeError{constructor(f,m){super(`[${f}] Signal value ${B5(m)} is invalid`);this.name="InvalidSignalValueError"}}class Qf extends TypeError{constructor(f,m){super(`[${f}] Callback ${B5(m)} is invalid`);this.name="InvalidCallbackError"}}class k5 extends Error{constructor(f){super(`[${f}] Signal is read-only`);this.name="ReadonlySignalError"}}class v5 extends Error{constructor(f){super(`[${f}] Active owner is required`);this.name="RequiredOwnerError"}}class X5 extends Error{constructor(f,m,M){super(`[${f}] Could not add key "${m}"${M?` with value ${JSON.stringify(M)}`:""} because it already exists`);this.name="DuplicateKeyError"}}function n(f,m,M){if(m==null)throw new bf(f);if(M&&!M(m))throw new d5(f,m)}function L5(f,m){if(m==null)throw new h0(f)}function x0(f,m,M=s){if(!M(m))throw new Qf(f,m)}var S0="State",d0="Memo",k0="Task",t0="Sensor",b0="List",o5="Collection",v0="Store",o0="Slot",L0=0,J5=1,g=2,r5=4,P0=8,S=null,G0=null,Uf=[],h=0,m1=!1,q0=(f,m)=>f===m;var M1=(f,m)=>{if(Object.is(f,m))return!0;if(typeof f!==typeof m)return!1;if(f==null||typeof f!=="object"||m==null||typeof m!=="object")return!1;let M=Array.isArray(f);if(M!==Array.isArray(m))return!1;if(M){let $=f,H=m;if($.length!==H.length)return!1;for(let Z=0;Z<$.length;Z++)if(!M1($[Z],H[Z]))return!1;return!0}if(W0(f)&&W0(m)){let $=Object.keys(f);if($.length!==Object.keys(m).length)return!1;for(let H of $){if(!(H in m))return!1;if(!M1(f[H],m[H]))return!1}return!0}return!1},r0=(f,m)=>M1(f,m);function k2(f,m){let M=m.sourcesTail;if(M){let $=m.sources;while($){if($===f)return!0;if($===M)break;$=$.nextSource}}return!1}function N0(f,m){let M=m.sourcesTail;if(M?.source===f)return;let $=null,H=m.flags&r5;if(H){if($=M?M.nextSource:m.sources,$?.source===f){m.sourcesTail=$;return}}let Z=f.sinksTail;if(Z?.sink===m&&(!H||k2(Z,m)))return;let z={source:f,sink:m,nextSource:$,prevSink:Z,nextSink:null};if(m.sourcesTail=f.sinksTail=z,M)M.nextSource=z;else m.sources=z;if(Z)Z.nextSink=z;else f.sinks=z}function v2(f){let{source:m,nextSource:M,nextSink:$,prevSink:H}=f;if($)$.prevSink=H;else m.sinksTail=H;if(H)H.nextSink=$;else m.sinks=$;if(!m.sinks){if(m.stop)m.stop(),m.stop=void 0;if("sources"in m&&m.sources){let Z=m;Z.sourcesTail=null,R5(Z),Z.flags|=g}}return M}function R5(f){let m=f.sourcesTail,M=m?m.nextSource:f.sources;while(M)M=v2(M);if(m)m.nextSource=null;else f.sources=null}function c(f,m=g){let M=f.flags;if("sinks"in f){if((M&(g|J5))>=m)return;if(f.flags=M|m,"controller"in f&&f.controller)f.controller.abort(),f.controller=void 0;for(let $=f.sinks;$;$=$.nextSink)c($.sink,J5)}else{if((M&(g|J5))>=m)return;let $=M&(g|J5);if(f.flags=m,!$)Uf.push(f)}}function O0(f,m){if(f.equals(f.value,m))return;f.value=m;for(let M=f.sinks;M;M=M.nextSink)c(M.sink);if(h===0)a()}function W5(f,m){if(!f.cleanup)f.cleanup=m;else if(Array.isArray(f.cleanup))f.cleanup.push(m);else f.cleanup=[f.cleanup,m]}function jf(f){if(!f.cleanup)return;if(Array.isArray(f.cleanup))for(let m=0;m<f.cleanup.length;m++)f.cleanup[m]();else f.cleanup();f.cleanup=null}function o2(f){let m=S;S=f,f.sourcesTail=null,f.flags=r5;let M=!1;try{let $=f.fn(f.value);if(f.error||!f.equals($,f.value))f.value=$,f.error=void 0,M=!0}catch($){M=!0,f.error=$ instanceof Error?$:Error(String($))}finally{S=m,R5(f)}if(M){for(let $=f.sinks;$;$=$.nextSink)if($.sink.flags&J5)$.sink.flags|=g}f.flags=L0}function r2(f){f.controller?.abort();let m=new AbortController;f.controller=m,f.error=void 0;let M=S;S=f,f.sourcesTail=null,f.flags=r5;let $;try{$=f.fn(f.value,m.signal)}catch(H){f.controller=void 0,f.error=H instanceof Error?H:Error(String(H));return}finally{S=M,R5(f)}O0(f.pendingNode,!0),$.then((H)=>{if(m.signal.aborted)return;f.controller=void 0,i(()=>{if(f.error||!f.equals(H,f.value)){f.value=H,f.error=void 0;for(let Z=f.sinks;Z;Z=Z.nextSink)c(Z.sink)}O0(f.pendingNode,!1)})},(H)=>{if(m.signal.aborted)return;f.controller=void 0;let Z=H instanceof Error?H:Error(String(H));i(()=>{if(!f.error||Z.name!==f.error.name||Z.message!==f.error.message){f.error=Z;for(let z=f.sinks;z;z=z.nextSink)c(z.sink)}O0(f.pendingNode,!1)})}),f.flags=L0}function $1(f){jf(f);let m=S,M=G0;S=G0=f,f.sourcesTail=null,f.flags=r5;try{let $=f.fn();if(typeof $==="function")W5(f,$)}finally{S=m,G0=M,R5(f)}f.flags=L0}function $0(f){if(f.flags&J5)for(let m=f.sources;m;m=m.nextSource){if("fn"in m.source)$0(m.source);if(f.flags&g)break}if(f.flags&r5)throw new S5("controller"in f?k0:("value"in f)?d0:"Effect");if(f.flags&g)if("controller"in f)r2(f);else if("value"in f)o2(f);else $1(f);else f.flags=L0}function a(){if(m1)return;m1=!0;try{for(let f=0;f<Uf.length;f++){let m=Uf[f];if(m.flags&(g|J5))$0(m)}Uf.length=0}finally{m1=!1}}function i(f){h++;try{f()}finally{if(h--,h===0)a()}}function Q0(f){let m=S;S=null;try{return f()}finally{S=m}}function U0(f,m){let M=G0,$={cleanup:null};G0=$;let H=()=>jf($);try{let Z=f();if(typeof Z==="function")W5($,Z);return H}finally{if(G0=M,!m?.root&&M)W5(M,H)}}function A0(f,m){return m?()=>{if(S){if(!f.sinks)f.stop=m();N0(f,S)}}:()=>{if(S)N0(f,S)}}function k(f,m){n(S0,f,m?.guard);let M={value:f,sinks:null,sinksTail:null,equals:m?.equals??q0,guard:m?.guard};return{[Symbol.toStringTag]:S0,get(){if(S)N0(M,S);return M.value},set($){n(S0,$,M.guard),O0(M,$)},update($){x0(S0,$);let H=$(M.value);n(S0,H,M.guard),O0(M,H)}}}function n5(f){return Z0(f,S0)}function Pf(f,m){if(f.length!==m.length)return!1;for(let M=0;M<f.length;M++)if(f[M]!==m[M])return!1;return!0}function Q6(f){let m=0,M=typeof f==="function";return[typeof f==="string"?()=>`${f}${m++}`:M?($)=>f($)||String(m++):()=>String(m++),M]}function n2(f,m,M,$,H,Z){let z={},X={},J={},B=[],W=!1,Q=new Map;for(let b=0;b<f.length;b++){let U=M[b],p=f[b];if(U&&p!==void 0)Q.set(U,p)}let j=new Set;for(let b=0;b<m.length;b++){let U=m[b];if(U===void 0)continue;let p=H?$(U):M[b]??$(U);if(j.has(p))throw new X5(b0,p,U);if(B.push(p),j.add(p),!Q.has(p))z[p]=U,W=!0;else if(!Z(Q.get(p),U))X[p]=U,W=!0}for(let[b]of Q)if(!j.has(b))J[b]=null,W=!0;if(!W&&!Pf(M,B))W=!0;return{add:z,change:X,remove:J,newKeys:B,changed:W}}function n0(f,m){n(b0,f,Array.isArray);let M=new Map,$=[],[H,Z]=Q6(m?.keyConfig),z=m?.itemEquals??r0,X=m?.createItem??((b)=>k(b,{equals:z})),J=()=>$.map((b)=>M.get(b)?.get()).filter((b)=>b!==void 0),B={fn:J,value:f,flags:g,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:r0,error:void 0},W=(b)=>{let U=!1;for(let p in b.add){let N=b.add[p];n(`${b0} item for key "${p}"`,N),M.set(p,X(N)),U=!0}if(Object.keys(b.change).length)i(()=>{for(let p in b.change){let N=b.change[p];n(`${b0} item for key "${p}"`,N);let K=M.get(p);if(K)K.set(N)}});for(let p in b.remove){M.delete(p);let N=$.indexOf(p);if(N!==-1)$.splice(N,1);U=!0}if(U)B.flags|=P0;return b.changed},Q=A0(B,m?.watched);for(let b=0;b<f.length;b++){let U=f[b];if(U===void 0)continue;let p=$[b];if(!p)p=H(U),$[b]=p;n(`${b0} item for key "${p}"`,U),M.set(p,X(U))}B.value=f,B.flags=0;let j={[Symbol.toStringTag]:b0,[Symbol.isConcatSpreadable]:!0,*[Symbol.iterator](){for(let b of $){let U=M.get(b);if(U)yield U}},get length(){return Q(),$.length},get(){if(Q(),B.sources){if(B.flags){let b=B.flags&P0;if(B.value=Q0(J),b){if(B.flags=g,$0(B),B.error)throw B.error}else B.flags=L0}}else if($0(B),B.error)throw B.error;return B.value},set(b){let U=B.flags&g?J():B.value,p=n2(U,b,$,H,Z,z);if(p.changed){$=p.newKeys,W(p),B.flags|=g;for(let N=B.sinks;N;N=N.nextSink)c(N.sink);if(h===0)a()}},update(b){j.set(b(j.get()))},at(b){let U=$[b];return U!==void 0?M.get(U):void 0},keys(){return Q(),$.values()},byKey(b){return M.get(b)},keyAt(b){return $[b]},indexOfKey(b){return $.indexOf(b)},add(b){let U=H(b);if(M.has(U))throw new X5(b0,U,b);if(!$.includes(U))$.push(U);n(`${b0} item for key "${U}"`,b),M.set(U,X(b)),B.flags|=g|P0;for(let p=B.sinks;p;p=p.nextSink)c(p.sink);if(h===0)a();return U},remove(b){let U=typeof b==="number"?$[b]:b;if(U===void 0)return;if(M.delete(U)){let N=typeof b==="number"?b:$.indexOf(U);if(N>=0)$.splice(N,1);B.flags|=g|P0;for(let K=B.sinks;K;K=K.nextSink)c(K.sink);if(h===0)a()}},replace(b,U){let p=M.get(b);if(!p)return;if(n(`${b0} item for key "${b}"`,U),z(Q0(()=>p.get()),U))return;p.set(U),B.flags|=g;for(let N=B.sinks;N;N=N.nextSink)c(N.sink);if(h===0)a()},sort(b){let p=$.map((N)=>[N,M.get(N)?.get()]).sort(s(b)?(N,K)=>b(N[1],K[1]):(N,K)=>String(N[1]).localeCompare(String(K[1]))).map(([N])=>N);if(!Pf($,p)){$=p,B.flags|=g;for(let N=B.sinks;N;N=N.nextSink)c(N.sink);if(h===0)a()}},splice(b,U,...p){let N=$.length,K=b<0?Math.max(0,N+b):Math.min(b,N),u=Math.max(0,Math.min(U??Math.max(0,N-Math.max(0,K)),N-K)),o={},D={};for(let I=0;I<u;I++){let r=K+I,d=$[r];if(d){let l=M.get(d);if(l)D[d]=l.get()}}let Y=$.slice(0,K),C={};for(let I of p){let r=H(I);if(r in D)delete D[r],C[r]=I;else if(M.has(r))throw new X5(b0,r,I);else o[r]=I;Y.push(r)}Y.push(...$.slice(K+u));let G=!!(Object.keys(o).length||Object.keys(D).length||Object.keys(C).length);if(G){W({add:o,change:C,remove:D,changed:G}),$=Y,B.flags|=g;for(let I=B.sinks;I;I=I.nextSink)c(I.sink);if(h===0)a()}return Object.values(D)},deriveCollection(b){return H1(j,b)}};return j}function qf(f){return Z0(f,b0)}function v(f,m){if(x0(d0,f,_5),m?.value!==void 0)n(d0,m.value,m?.guard);let M={fn:f,value:m?.value,flags:g,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:m?.equals??q0,error:void 0,stop:void 0},$=m?.watched,H=A0(M,$?()=>$(()=>{if(c(M),h===0)a()}):void 0);return{[Symbol.toStringTag]:d0,get(){if(H(),$0(M),M.error)throw M.error;return L5(d0,M.value),M.value}}}function b5(f){return Z0(f,d0)}function I0(f,m){if(x0(k0,f,i0),m?.value!==void 0)n(k0,m.value,m?.guard);let M={value:!1,sinks:null,sinksTail:null,equals:q0},$={fn:f,value:m?.value,sources:null,sourcesTail:null,sinks:null,sinksTail:null,flags:g,equals:m?.equals??q0,controller:void 0,error:void 0,stop:void 0,pendingNode:M},H=m?.watched,Z=A0($,H?()=>H(()=>{if(c($),h===0)a()}):void 0),z=A0(M);return{[Symbol.toStringTag]:k0,get(){if(Z(),$0($),$.error)throw $.error;return L5(k0,$.value),$.value},isPending(){return z(),$.pendingNode.value},abort(){$.controller?.abort(),$.controller=void 0,O0($.pendingNode,!1)}}}function D5(f){return Z0(f,k0)}function H1(f,m){x0(o5,m);let M=i0(m),$=new Map,H=[],Z=(b)=>{let U=M?I0(async(p,N)=>{let K=f.byKey(b)?.get();if(K==null)return p;return m(K,N)}):v(()=>{let p=f.byKey(b)?.get();if(p==null)return;return m(p)});$.set(b,U)};function z(b){if(!Pf(H,b)){let U=new Set(H),p=new Set(b);for(let N of H)if(!p.has(N))$.delete(N);for(let N of b)if(!U.has(N))Z(N);H=b,B.flags|=P0}}function X(){z(Array.from(f.keys()));let b=[];for(let U of H)try{let p=$.get(U)?.get();if(p!=null)b.push(p)}catch(p){if(!(p instanceof h0))throw p}return b}let B={fn:X,value:[],flags:g,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:(b,U)=>{if(b.length!==U.length)return!1;for(let p=0;p<b.length;p++)if(b[p]!==U[p])return!1;return!0},error:void 0};function W(){if(B.sources){if(B.flags)if(B.value=Q0(X),B.flags&P0){if(B.flags=g,$0(B),B.error)throw B.error}else B.flags=L0}else if(B.sinks){if($0(B),B.error)throw B.error}else B.value=Q0(X)}let Q=Array.from(Q0(()=>f.keys()));for(let b of Q)Z(b);H=Q;let j={[Symbol.toStringTag]:o5,[Symbol.isConcatSpreadable]:!0,*[Symbol.iterator](){for(let b of H){let U=$.get(b);if(U)yield U}},get length(){if(S)N0(B,S);return W(),H.length},keys(){if(S)N0(B,S);return W(),H.values()},get(){if(S)N0(B,S);return W(),B.value},at(b){let U=H[b];return U!==void 0?$.get(U):void 0},byKey(b){return $.get(b)},keyAt(b){return H[b]},indexOfKey(b){return H.indexOf(b)},deriveCollection(b){return H1(j,b)}};return j}function R0(f){x0("Effect",f);let m={fn:f,flags:g,sources:null,sourcesTail:null,cleanup:null},M=()=>{jf(m),m.fn=void 0,m.flags=L0,m.sourcesTail=null,R5(m)};if(G0)W5(G0,M);return $1(m),M}function V5(f,m){if(!G0)throw new v5("match");let M=!Array.isArray(f),$=M?[f]:f,{nil:H,stale:Z}=m,z=M?(j)=>m.ok(j[0]):(j)=>m.ok(j),X=M&&m.err?(j)=>m.err(j[0]):m.err??console.error,J,B=!1,W=Array($.length);for(let j=0;j<$.length;j++)try{W[j]=$[j].get()}catch(b){if(b instanceof h0){B=!0;continue}if(!J)J=[];J.push(b instanceof Error?b:Error(String(b)))}let Q;try{if(B)Q=H?.();else if(J)Q=X(J);else if(Z&&(M?D5($[0])&&$[0].isPending():$.some((j)=>D5(j)&&j.isPending())))Q=Z();else Q=z(W)}catch(j){Q=X([j instanceof Error?j:Error(String(j))])}if(typeof Q==="function")return Q;if(Q instanceof Promise){let j=G0,b=new AbortController;W5(j,()=>b.abort()),Q.then((U)=>{if(!b.signal.aborted&&typeof U==="function")W5(j,U)}).catch((U)=>{X([U instanceof Error?U:Error(String(U))])})}}function Q5(f,m){if(x0(t0,f,_5),m?.value!==void 0)n(t0,m.value,m?.guard);let M={value:m?.value,sinks:null,sinksTail:null,equals:m?.equals??q0,guard:m?.guard,stop:void 0};return{[Symbol.toStringTag]:t0,get(){if(S){if(!M.sinks)M.stop=f(($)=>{n(t0,$,M.guard),O0(M,$)});N0(M,S)}return L5(t0,M.value),M.value}}}function l2(f,m){let M={},$={},H={},Z=!1,z=Object.keys(f),X=Object.keys(m);for(let J of X)if(J in f){if(!r0(f[J],m[J]))$[J]=m[J],Z=!0}else M[J]=m[J],Z=!0;for(let J of z)if(!(J in m))H[J]=void 0,Z=!0;return{add:M,change:$,remove:H,changed:Z}}function Y5(f,m){n(v0,f,W0);let M=new Map,$=(B,W)=>{if(n(`${v0} for key "${B}"`,W),Array.isArray(W))M.set(B,n0(W));else if(W0(W))M.set(B,Y5(W));else M.set(B,k(W))},H=()=>{let B={};return M.forEach((W,Q)=>{B[Q]=W.get()}),B},Z={fn:H,value:f,flags:g,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:r0,error:void 0},z=(B)=>{let W=!1;for(let Q in B.add)$(Q,B.add[Q]),W=!0;if(Object.keys(B.change).length)i(()=>{for(let Q in B.change){let j=B.change[Q];n(`${v0} for key "${Q}"`,j);let b=M.get(Q);if(b)if(W0(j)!==l5(b))$(Q,j),W=!0;else b.set(j)}});for(let Q in B.remove)M.delete(Q),W=!0;if(W)Z.flags|=P0;return B.changed},X=A0(Z,m?.watched);for(let B of Object.keys(f))$(B,f[B]);let J={[Symbol.toStringTag]:v0,[Symbol.isConcatSpreadable]:!1,*[Symbol.iterator](){for(let B of Array.from(M.keys())){let W=M.get(B);if(W)yield[B,W]}},keys(){return X(),M.keys()},byKey(B){return M.get(B)},get(){if(X(),Z.sources){if(Z.flags){let B=Z.flags&P0;if(Z.value=Q0(H),B){if(Z.flags=g,$0(Z),Z.error)throw Z.error}else Z.flags=L0}}else if($0(Z),Z.error)throw Z.error;return Z.value},set(B){let W=Z.flags&g?H():Z.value,Q=l2(W,B);if(z(Q)){Z.flags|=g;for(let j=Z.sinks;j;j=j.nextSink)c(j.sink);if(h===0)a()}},update(B){J.set(B(J.get()))},add(B,W){if(M.has(B))throw new X5(v0,B,W);$(B,W),Z.flags|=g|P0;for(let Q=Z.sinks;Q;Q=Q.nextSink)c(Q.sink);if(h===0)a();return B},remove(B){if(M.delete(B)){Z.flags|=g|P0;for(let Q=Z.sinks;Q;Q=Q.nextSink)c(Q.sink);if(h===0)a()}}};return new Proxy(J,{get(B,W){if(W in B)return Reflect.get(B,W);if(typeof W!=="symbol")return B.byKey(W)},has(B,W){if(W in B)return!0;return B.byKey(String(W))!==void 0},ownKeys(B){return Array.from(B.keys())},getOwnPropertyDescriptor(B,W){if(W in B)return Reflect.getOwnPropertyDescriptor(B,W);if(typeof W==="symbol")return;let Q=B.byKey(String(W));return Q?{enumerable:!0,configurable:!0,writable:!0,value:Q}:void 0}})}function l5(f){return Z0(f,v0)}var c2=new Set([S0,d0,k0,t0,o0,b0,o5,v0]);function G5(f,m){return i0(f)?I0(f,m):v(f,m)}function F5(f){return f!=null&&c2.has(f[Symbol.toStringTag])}function pf(f){return n5(f)||l5(f)||qf(f)}function j6(f){if(F5(f))return!0;return f!==null&&typeof f==="object"&&"get"in f&&typeof f.get==="function"}function xf(f,m){n(o0,f,j6);let M=f,$=m?.guard,H={fn:()=>M.get(),value:void 0,flags:g,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:m?.equals??q0,error:void 0},Z=()=>{if(S)N0(H,S);if($0(H),H.error)throw H.error;return H.value},z=(J)=>{if(U5(M))return M.set(J);if("set"in M&&typeof M.set==="function")n(o0,J,$),M.set(J);else throw new k5(o0)},X=(J)=>{n(o0,J,j6),M=J,H.flags|=g;for(let B=H.sinks;B;B=B.nextSink)c(B.sink);if(h===0)a()};return{[Symbol.toStringTag]:o0,configurable:!0,enumerable:!0,get:Z,set:z,replace:X,current:()=>M}}function U5(f){return Z0(f,o0)}var Nf="context-request";class Z1 extends Event{context;callback;subscribe;constructor(f,m,M=!1){super(Nf,{bubbles:!0,composed:!0});this.context=f,this.callback=m,this.subscribe=M}}var P6=(f)=>(m)=>()=>U0(()=>{let M=($)=>{let{context:H,callback:Z}=$;if(typeof H==="string"&&m.includes(H)&&s(Z))$.stopImmediatePropagation(),Z(()=>f[H])};return f.addEventListener(Nf,M),()=>f.removeEventListener(Nf,M)}),q6=(f)=>(m,M)=>{let $=()=>M;return f.dispatchEvent(new Z1(m,(H)=>{$=H})),v($)};var j5=typeof process<"u"&&!0,P5="warn",m8=(f)=>f?`#${f}`:"",M8=(f)=>f?.length?`.${Array.from(f).join(".")}`:"",z1=(f)=>f.localName.includes("-"),B1=(f)=>z1(f)&&f.matches(":not(:defined)"),F0=(f)=>f?`<${f.localName}${m8(f.id)}${M8(f.classList)}>`:"<unknown>";class Lf extends TypeError{constructor(f){super(`Invalid component name "${f}". Custom element names must contain a hyphen, start with a lowercase letter, and contain only lowercase letters, numbers, and hyphens.`);this.name="InvalidComponentNameError"}}class c5 extends Error{constructor(f,m,M){super(`Missing required element <${m}> in component ${F0(f)}. ${M}`);this.name="MissingElementError"}}class Rf extends Error{constructor(f,m){super(`Timeout waiting for: [${m.join(", ")}] in component ${F0(f)}.`);this.name="DependencyTimeoutError"}}class Df extends TypeError{constructor(f,m,M){super(`Expected reactives passed from ${F0(f)} to ${F0(m)} to be a record of signals, reactive property names or functions. Got ${B5(M)}.`);this.name="InvalidReactivesError"}}class Vf extends TypeError{constructor(f,m){super(`Target ${F0(f)} is not a custom element in ${m}.`);this.name="InvalidCustomElementError"}}var p6=new WeakMap,i5=(f)=>{let m=p6.get(f);if(!m)m={},p6.set(f,m);return m};var Yf=(f)=>{for(let m of f)if(Array.isArray(m))Yf(m);else if(m)m()},X1=(f,m)=>{if(s(m))return G5(m);if(typeof m==="string"){let M=i5(f)[m];if(M)return M;return v(()=>f[m])}if(m&&typeof m==="object"&&"get"in m&&!(Symbol.toStringTag in m))return m;return m},x6=(f)=>{function m(M,$){return()=>{if(Array.isArray(M)){let Z=M.map((X)=>X1(f,X)),z=$;return R0(()=>V5(Z,{ok:(X)=>Q0(()=>z(X))}))}let H=X1(f,M);if(typeof $==="function")return R0(()=>V5(H,{ok:(Z)=>Q0(()=>$(Z))}));return R0(()=>V5(H,$))}}return m},N6=(f)=>{let m=($,H)=>U0(()=>{if(!z1($))throw new Vf($,`pass from ${F0(f)}`);if(!W0(H))throw new Df(f,$,H);let Z=i5($),z=F0($),X=[];for(let[J,B]of Object.entries(H)){if(B==null)continue;if(!(J in $)){if(j5)console[P5](`pass(): property '${J}' does not exist on ${z}`);continue}let W=X1(f,B);if(!W)continue;let Q=Z[J];if(U5(Q)){let j=Q.current();Q.replace(W),X.push(()=>Q.replace(j));continue}if(j5)console[P5](`pass(): property '${J}' on ${z} is not Slot-backed — use setProperty() for non-Le Truc elements`)}if(X.length)return()=>{for(let J of X)J()}});function M($,H){return()=>{if(!$)return;if(b5($))R0(()=>{for(let Z of $.get())U0(()=>m(Z,H))});else m($,H)}}return M};function t(f,m){return()=>{R0(()=>{for(let M of f.get())U0(()=>{let $=m(M);if(Array.isArray($))Yf($);else if(typeof $==="function")$()})})}}var J1=new Set,L6=new WeakMap,Gf=new Set,W1,$8=()=>{W1=void 0;let f=Array.from(J1);J1.clear();for(let M of f)L6.get(M)?.();let m=Array.from(Gf);Gf.clear();for(let M of m)M()},R6=()=>{if(!W1)W1=requestAnimationFrame($8)},h5=(f,m)=>{L6.set(f,m),J1.add(f),R6()},a0=(f,m)=>{let M=!1,$,H=()=>{M=!1,f(...$)},Z=(...z)=>{if($=z,M)return;M=!0,Gf.add(H),R6()};return Z.cancel=()=>{if(M)Gf.delete(H),M=!1},m?.addEventListener("abort",Z.cancel,{once:!0}),Z};var H8=new Set(["scroll","resize","mousewheel","touchstart","touchmove","wheel"]),Z8=new Set(["focus","blur","scroll","resize","load","unload","error","toggle","mouseenter","mouseleave","pointerenter","pointerleave","abort","canplay","canplaythrough","durationchange","emptied","ended","loadeddata","loadedmetadata","loadstart","pause","play","playing","progress","ratechange","seeked","seeking","stalled","suspend","timeupdate","volumechange","waiting"]),D6=(f,m,M,$,H)=>{let Z=(X)=>{let J=$(X,m);if(!W0(J))return;i(()=>{for(let[B,W]of Object.entries(J))f[B]=W})},z=H.passive?a0(Z):Z;return m.addEventListener(M,z,H),()=>{m.removeEventListener(M,z),z.cancel?.()}},V6=(f)=>{function m(M,$,H,Z={}){return()=>{if(!M)return;if(!("passive"in Z))Z={...Z,passive:H8.has($)};if(b5(M)){if(Z8.has($)){if(j5)console[P5](`on(): '${$}' does not bubble — prefer each() + on() for per-element listeners in ${F0(f)}`);return R0(()=>{for(let B of M.get())U0(()=>{return D6(f,B,$,H,Z)})})}let z=f.shadowRoot??f,X=(B)=>{let W=B.composedPath();for(let Q of M.get())if(W.includes(Q)){let j=H(B,Q);if(!W0(j))break;i(()=>{for(let[b,U]of Object.entries(j))f[b]=U});break}},J=Z.passive?a0(X):X;U0(()=>{return z.addEventListener($,J,Z),()=>{z.removeEventListener($,J),J.cancel?.()}});return}U0(()=>D6(f,M,$,H,Z))}}return m};var z8=200,B8=(f)=>{let m=new Set,M="",$=0;for(let H of f)if(H==="[")$++;else if(H==="]"){if($>0)$--}else if($===0)M+=H;if(M.includes("."))m.add("class");if(M.includes("#"))m.add("id");if(f.includes("[")){let H=f.split("[");for(let Z=1;Z<H.length;Z++){let z=H[Z];if(!z||!z.includes("]"))continue;let X=z.split("=")[0].split("]")[0].trim().replace(/[^a-zA-Z0-9_-]/g,"");if(X)m.add(X)}}return[...m]};function Ff(f,m){return v(()=>Array.from(f.querySelectorAll(m)),{value:[],equals:(M,$)=>M.length===$.length&&M.every((H,Z)=>H===$[Z]),watched:(M)=>{let $={childList:!0,subtree:!0},H=B8(m);if(H.length)$.attributes=!0,$.attributeFilter=H;let Z=(J)=>J instanceof Element&&(J.matches(m)||J.querySelector(m)),z=(J)=>{if(J.type==="attributes")return!0;if(J.type==="childList")return Array.from(J.addedNodes).some(Z)||Array.from(J.removedNodes).some(Z);return!1},X=new MutationObserver((J)=>{for(let B of J)if(z(B)){M();return}});return X.observe(f,$),()=>X.disconnect()}})}var Y6=(f)=>{let m=f.shadowRoot??f,M=new Set;function $(z,X){let J=m.querySelector(z);if(X!=null&&!J)throw new c5(f,z,X);if(J&&B1(J))M.add(J.localName);return J??void 0}function H(z,X){let J=Ff(m,z),B=J.get();if(X!=null&&!B.length)throw new c5(f,z,X);if(B.length){for(let W of B)if(B1(W))M.add(W.localName)}return J}return[{first:$,all:H},(z)=>{if(M.size)queueMicrotask(()=>{let X=Array.from(M).filter((J)=>!customElements.get(J));if(!X.length){z();return}Promise.race([Promise.all(X.map((J)=>customElements.whenDefined(J))),new Promise((J,B)=>{setTimeout(()=>{B(new Rf(f,X.filter((W)=>!customElements.get(W))))},z8)})]).then(z).catch((J)=>{if(j5)console[P5](J);z()})});else z()}]};var G6=Symbol("parser"),F6=Symbol("method"),C6=(f)=>s(f)&&(G6 in f),K6=(f)=>s(f)&&(F6 in f),D0=(f)=>Object.assign(f,{[G6]:!0}),s0=(f)=>Object.assign(f,{[F6]:!0});function R(f,m){if(!f.includes("-")||!f.match(/^[a-z][a-z0-9-]*$/))throw new Lf(f);class M extends HTMLElement{debug;#M=!1;#f=[];#m;connectedCallback(){let $=()=>{this.#m=U0(()=>Yf(this.#f),{root:!0})};if(this.#M)$();else{let H=this,[Z,z]=Y6(this),X={expose:this.#H.bind(this),host:H,...Z,watch:x6(H),on:V6(H),pass:N6(H),provideContexts:P6(H),requestContext:q6(H)},J=m(X);if(J)this.#f=J;if(this.#M=!0,!this.#f.length)return;z($)}}disconnectedCallback(){if(s(this.#m))this.#m()}#H($){let H=(Z,z)=>{if(C6(z)){let X=z(this.getAttribute(Z));if(X!=null)this.#$(Z,X)}else if(K6(z))this[Z]=z;else{let X=z;if(X!=null)this.#$(Z,X)}};for(let[Z,z]of Object.entries($)){if(z==null||Z in this)continue;H(Z,z)}}#$($,H){let Z=F5(H)?H:s(H)?G5(H):k(H),z=i5(this),X=$,J=z[X];if(U5(J))J.replace(Z);else if(pf(Z)){let B=xf(Z);z[X]=B,Object.defineProperty(this,$,B)}else z[X]=Z,Object.defineProperty(this,$,{get:Z.get,enumerable:!0})}}return customElements.define(f,M),customElements.get(f)}var X8=["type","src","async","defer","nomodule","crossorigin","integrity","referrerpolicy","fetchpriority"],J8=(f)=>{if(/^(javascript|data|vbscript):/i.test(f))return!1;if(/^(mailto|tel):/i.test(f))return!0;if(f.includes("://"))try{let m=new URL(f);return["http:","https:","ftp:"].includes(m.protocol)}catch{return!1}return!0},t5=(f,m,M)=>{if(/^on/i.test(m))throw Error(`setAttribute: blocked unsafe attribute name '${m}' on ${f.localName} — event handler attributes are not allowed`);if(M=String(M).trim(),!J8(M))throw Error(`setAttribute: blocked unsafe value for '${m}' on <${f.localName}>: '${M}'`);f.setAttribute(m,M)},V0=(f)=>f.replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;").replace(/"/g,"&quot;").replace(/'/g,"&#39;"),O6=(f,m)=>{Array.from(f.childNodes).filter((M)=>M.nodeType!==Node.COMMENT_NODE).forEach((M)=>{M.remove()}),f.append(document.createTextNode(m))},w=(f,m=!1)=>m?(M)=>O6(f,String(M)):(M)=>{f.textContent=String(M)},H0=(f,m)=>(M)=>{f[m]=M},a5=(f,m)=>(M)=>{f.classList.toggle(m,Boolean(M))},z0=(f)=>(m)=>{f.hidden=!m},q5=(f,m,M=!1)=>({ok:($)=>{if(typeof $==="boolean")f.toggleAttribute(m,$);else if(M)f.setAttribute(m,$);else t5(f,m,$)},nil:()=>{f.removeAttribute(m)}}),e0=(f,m)=>({ok:(M)=>{f.style.setProperty(m,M)},nil:()=>{f.style.removeProperty(m)}}),b1=(f,m={})=>{let M=()=>{if(f.shadowRoot)f.shadowRoot.innerHTML="<slot></slot>";else f.innerHTML=""};return{ok:($)=>{if(!$){M();return}let{shadowRootMode:H,allowScripts:Z}=m;if(H&&!f.shadowRoot)f.attachShadow({mode:H});let z=f.shadowRoot||f;h5(f,()=>{if(z.innerHTML=$,Z)z.querySelectorAll("script").forEach((X)=>{let J=document.createElement("script");for(let B of X8){let W=X.getAttribute(B);if(W!==null)J.setAttribute(B,W)}if(!X.hasAttribute("src"))J.appendChild(document.createTextNode(X.textContent??""));z.appendChild(J),X.remove()})})},nil:M}};var s5=()=>D0((f)=>f!=null&&f!=="false");var Q1=(f)=>D0((m)=>{if((m??f)==null)throw TypeError("asJSON: Value and fallback are both null or undefined");if(m==null)return f;if(m==="")throw SyntaxError("Empty string is not valid JSON");let M;try{M=JSON.parse(m)}catch($){throw SyntaxError(`Failed to parse JSON: ${String($)}`,{cause:$})}return M??f});var U1=(f,m)=>{if(m==null)return;let M=f(m);return Number.isFinite(M)?M:void 0};var p5=(f=0)=>D0((m)=>U1(parseFloat,m)??f),C5=(f=0,m=Number.MAX_SAFE_INTEGER)=>D0((M)=>{if(M==null)return f;let $=M.trim(),H=$.toLowerCase().startsWith("0x")?U1((z)=>parseInt(z,16),$):U1(parseFloat,M),Z=H!=null?Math.trunc(H):f;return Math.max(f,Math.min(Z,m))});var y0=(f="")=>D0((m)=>m??f);var km=R("basic-button",({expose:f,first:m,watch:M})=>{let $=m("button","Add a native button as descendant."),H=m("span.label"),Z=m("span.badge");return f({disabled:$.disabled,label:H?.textContent??$.textContent??"",badge:Z?.textContent??""}),[M("disabled",H0($,"disabled")),H&&M("label",w(H)),Z&&M("badge",w(Z))]});var rm=R("basic-counter",({expose:f,first:m,host:M,on:$,watch:H})=>{let Z=m("button","Add a native button element to increment the count."),z=m("span","Add a span to display the count.");return f({count:Number.parseInt(z.textContent||"0")}),[$(Z,"click",()=>({count:M.count+1})),H("count",w(z))]});var cm=R("basic-gauge",({expose:f,first:m,host:M,pass:$,watch:H})=>{let Z=m("meter","Add a <meter> element to display the level"),z=m("basic-number","Add a <basic-number> element to display the value"),X=m(".label","Add an element to display the qualification label");f({value:Z.value});let J=Q1([])(M.getAttribute("thresholds"));return[$(z,{value:()=>M.value}),H("value",(B)=>{Z.value=B,M.style.setProperty("--basic-gauge-degree",`${240*B/Z.max}deg`)}),H(()=>J.find((B)=>M.value>=B.min)||{label:"",color:"var(--color-primary)"},(B)=>{X.textContent=B.label,M.style.setProperty("--basic-gauge-color",B.color)})]});var tm=R("basic-hello",({expose:f,first:m,on:M,watch:$})=>{let H=m("input","Needed to enter the name."),Z=m("output","Needed to display the name."),z=Z.textContent||"";return f({name:Z.textContent??""}),[M(H,"input",()=>({name:H.value||z})),$("name",w(Z))]});function K5(f){return f.closest("[lang]")?.getAttribute("lang")||"en"}function W8(f,m,M={onWarn:console.warn,onError:console.error}){let $=()=>new Intl.NumberFormat(f);if(!m)return $();let{onWarn:H,onError:Z}=M,z={};try{z=JSON.parse(m)}catch(U){return Z?.(`Invalid JSON: ${U}`),$()}let X=z.style??"decimal",J=[];if(X==="currency"){if(!z.currency||typeof z.currency!=="string"||z.currency.length!==3)return Z?.('style="currency" requires a 3-letter ISO currency (e.g. "CHF").'),$()}else J.push("currency","currencyDisplay","currencySign");if(X==="unit"){if(!z.unit||typeof z.unit!=="string")return Z?.('style="unit" requires a "unit" (e.g. "liter", "kilometer-per-hour").'),$()}else J.push("unit","unitDisplay");if(z.notation&&z.notation!=="compact")J.push("compactDisplay");let B={};for(let[U,p]of Object.entries(z))if(!J.includes(U))B[U]=p;else H?.(`Option "${U}" is ignored for style="${X}".`);let{minimumFractionDigits:W,maximumFractionDigits:Q}=B;if(W!=null&&Q!=null&&W>Q)H?.(`minimumFractionDigits (${W}) > maximumFractionDigits (${Q}); swapping.`),B.minimumFractionDigits=Q,B.maximumFractionDigits=W;let{minimumSignificantDigits:j,maximumSignificantDigits:b}=B;if(j!=null&&b!=null&&j>b)H?.(`minimumSignificantDigits (${j}) > maximumSignificantDigits (${b}); swapping.`),B.minimumSignificantDigits=b,B.maximumSignificantDigits=j;try{let U=new Intl.NumberFormat(f,B);if(U.resolvedOptions().locale!==f)H(`Fall back to locale ${U.resolvedOptions().locale} instead of ${f}`);return U}catch(U){return Z?.(`Options rejected by Intl.NumberFormat: ${U instanceof Error?U.message:String(U)}`),$()}}var m7=R("basic-number",({expose:f,host:m,watch:M})=>{let $=W8(K5(m),m.getAttribute("options"));return f({value:p5()}),[M(()=>$.format(m.value),w(m))]});var Z7=R("basic-pluralize",({expose:f,first:m,host:M,watch:$})=>{let H=m(".count"),Z=m(".none"),z=m(".some"),X=m(".zero"),J=m(".one"),B=m(".two"),W=m(".few"),Q=m(".many"),j=m(".other"),b=new Intl.PluralRules(K5(M),M.hasAttribute("ordinal")?{type:"ordinal"}:void 0);f({count:C5()});let U={zero:X,one:J,two:B,few:W,many:Q,other:j},p=b.resolvedOptions().pluralCategories;return[H&&$("count",w(H)),Z&&$(()=>M.count===0,z0(Z)),z&&$(()=>M.count!==0,z0(z)),...p.map((N)=>{let K=U[N];return K&&$(()=>b.select(M.count)===N,z0(K))})]});var j1="invalid date",b8="unknown date";function A6(f,m,{dateStyle:M="long"}={}){let[$,H,Z]=m.split("-").map(Number);if(!$||Number.isNaN($)||!H||Number.isNaN(H)||Number.isNaN(Z))return j1;let z=new Date($,H-1,Z);return new Intl.DateTimeFormat(f,{dateStyle:M}).format(z)}var J7=R("card-blogmeta",({host:f,first:m})=>{let M=m("time.published","Add a <time> element to display the publication date."),$=m("span.modified"),H=m(".modified time"),Z=K5(f);if(M.textContent=M.dateTime?A6(Z,M.dateTime):b8,H){let z=H.dateTime?A6(Z,H.dateTime):j1;if($&&z===j1)$.remove();else H.textContent=z}});var Q8=(f,m)=>{if(typeof f!=="number")return;if(m===3)return{mode:"rgb",r:(f>>8&15|f>>4&240)/255,g:(f>>4&15|f&240)/255,b:(f&15|f<<4&240)/255};if(m===4)return{mode:"rgb",r:(f>>12&15|f>>8&240)/255,g:(f>>8&15|f>>4&240)/255,b:(f>>4&15|f&240)/255,alpha:(f&15|f<<4&240)/255};if(m===6)return{mode:"rgb",r:(f>>16&255)/255,g:(f>>8&255)/255,b:(f&255)/255};if(m===8)return{mode:"rgb",r:(f>>24&255)/255,g:(f>>16&255)/255,b:(f>>8&255)/255,alpha:(f&255)/255}},Cf=Q8;var U8={aliceblue:15792383,antiquewhite:16444375,aqua:65535,aquamarine:8388564,azure:15794175,beige:16119260,bisque:16770244,black:0,blanchedalmond:16772045,blue:255,blueviolet:9055202,brown:10824234,burlywood:14596231,cadetblue:6266528,chartreuse:8388352,chocolate:13789470,coral:16744272,cornflowerblue:6591981,cornsilk:16775388,crimson:14423100,cyan:65535,darkblue:139,darkcyan:35723,darkgoldenrod:12092939,darkgray:11119017,darkgreen:25600,darkgrey:11119017,darkkhaki:12433259,darkmagenta:9109643,darkolivegreen:5597999,darkorange:16747520,darkorchid:10040012,darkred:9109504,darksalmon:15308410,darkseagreen:9419919,darkslateblue:4734347,darkslategray:3100495,darkslategrey:3100495,darkturquoise:52945,darkviolet:9699539,deeppink:16716947,deepskyblue:49151,dimgray:6908265,dimgrey:6908265,dodgerblue:2003199,firebrick:11674146,floralwhite:16775920,forestgreen:2263842,fuchsia:16711935,gainsboro:14474460,ghostwhite:16316671,gold:16766720,goldenrod:14329120,gray:8421504,green:32768,greenyellow:11403055,grey:8421504,honeydew:15794160,hotpink:16738740,indianred:13458524,indigo:4915330,ivory:16777200,khaki:15787660,lavender:15132410,lavenderblush:16773365,lawngreen:8190976,lemonchiffon:16775885,lightblue:11393254,lightcoral:15761536,lightcyan:14745599,lightgoldenrodyellow:16448210,lightgray:13882323,lightgreen:9498256,lightgrey:13882323,lightpink:16758465,lightsalmon:16752762,lightseagreen:2142890,lightskyblue:8900346,lightslategray:7833753,lightslategrey:7833753,lightsteelblue:11584734,lightyellow:16777184,lime:65280,limegreen:3329330,linen:16445670,magenta:16711935,maroon:8388608,mediumaquamarine:6737322,mediumblue:205,mediumorchid:12211667,mediumpurple:9662683,mediumseagreen:3978097,mediumslateblue:8087790,mediumspringgreen:64154,mediumturquoise:4772300,mediumvioletred:13047173,midnightblue:1644912,mintcream:16121850,mistyrose:16770273,moccasin:16770229,navajowhite:16768685,navy:128,oldlace:16643558,olive:8421376,olivedrab:7048739,orange:16753920,orangered:16729344,orchid:14315734,palegoldenrod:15657130,palegreen:10025880,paleturquoise:11529966,palevioletred:14381203,papayawhip:16773077,peachpuff:16767673,peru:13468991,pink:16761035,plum:14524637,powderblue:11591910,purple:8388736,rebeccapurple:6697881,red:16711680,rosybrown:12357519,royalblue:4286945,saddlebrown:9127187,salmon:16416882,sandybrown:16032864,seagreen:3050327,seashell:16774638,sienna:10506797,silver:12632256,skyblue:8900331,slateblue:6970061,slategray:7372944,slategrey:7372944,snow:16775930,springgreen:65407,steelblue:4620980,tan:13808780,teal:32896,thistle:14204888,tomato:16737095,turquoise:4251856,violet:15631086,wheat:16113331,white:16777215,whitesmoke:16119285,yellow:16776960,yellowgreen:10145074},e5=U8;var j8=(f)=>{return Cf(e5[f.toLowerCase()],6)},I6=j8;var P8=/^#?([0-9a-f]{8}|[0-9a-f]{6}|[0-9a-f]{4}|[0-9a-f]{3})$/i,q8=(f)=>{let m;return(m=f.match(P8))?Cf(parseInt(m[1],16),m[1].length):void 0},y6=q8;var Kf="([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?)";var x5="([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?)%";var ff="(?:([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?)%|([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?))";var w6="(?:([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?)(deg|grad|rad|turn)|([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?))";var f5="\\s*,\\s*";var x7=new RegExp("^(?:([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?)%|([+-]?\\d*\\.?\\d+(?:[eE][+-]?\\d+)?)|none)$");var p8=new RegExp(`^rgba?\\(\\s*${Kf}${f5}${Kf}${f5}${Kf}\\s*(?:,\\s*${ff}\\s*)?\\)$`),x8=new RegExp(`^rgba?\\(\\s*${x5}${f5}${x5}${f5}${x5}\\s*(?:,\\s*${ff}\\s*)?\\)$`),N8=(f)=>{let m={mode:"rgb"},M;if(M=f.match(p8)){if(M[1]!==void 0)m.r=M[1]/255;if(M[2]!==void 0)m.g=M[2]/255;if(M[3]!==void 0)m.b=M[3]/255}else if(M=f.match(x8)){if(M[1]!==void 0)m.r=M[1]/100;if(M[2]!==void 0)m.g=M[2]/100;if(M[3]!==void 0)m.b=M[3]/100}else return;if(M[4]!==void 0)m.alpha=Math.max(0,Math.min(1,M[4]/100));else if(M[5]!==void 0)m.alpha=Math.max(0,Math.min(1,+M[5]));return m},u6=N8;var L8=(f,m)=>f===void 0?void 0:typeof f!=="object"?T6(f):f.mode!==void 0?f:m?{...f,mode:m}:void 0,m5=L8;var R8=(f="rgb")=>(m)=>(m=m5(m,f))!==void 0?m.mode===f?m:C0[m.mode][f]?C0[m.mode][f](m):f==="rgb"?C0[m.mode].rgb(m):C0.rgb[f](C0[m.mode].rgb(m)):void 0,e=R8;var C0={},g6={},mf=[],P1={},D8=(f)=>f,f0=(f)=>{if(C0[f.mode]={...C0[f.mode],...f.toMode},Object.keys(f.fromMode||{}).forEach((m)=>{if(!C0[m])C0[m]={};C0[m][f.mode]=f.fromMode[m]}),!f.ranges)f.ranges={};if(!f.difference)f.difference={};return f.channels.forEach((m)=>{if(f.ranges[m]===void 0)f.ranges[m]=[0,1];if(!f.interpolate[m])throw Error(`Missing interpolator for: ${m}`);if(typeof f.interpolate[m]==="function")f.interpolate[m]={use:f.interpolate[m]};if(!f.interpolate[m].fixup)f.interpolate[m].fixup=D8}),g6[f.mode]=f,(f.parse||[]).forEach((m)=>{V8(m,f.mode)}),e(f.mode)},w0=(f)=>g6[f],V8=(f,m)=>{if(typeof f==="string"){if(!m)throw Error("'mode' required when 'parser' is a string");P1[f]=m}else if(typeof f==="function"){if(mf.indexOf(f)<0)mf.push(f)}};var q1=/[^\x00-\x7F]|[a-zA-Z_]/,Y8=/[^\x00-\x7F]|[-\w]/,x={Function:"function",Ident:"ident",Number:"number",Percentage:"percentage",ParenClose:")",None:"none",Hue:"hue",Alpha:"alpha"},A=0;function Of(f){let m=f[A],M=f[A+1];if(m==="-"||m==="+")return/\d/.test(M)||M==="."&&/\d/.test(f[A+2]);if(m===".")return/\d/.test(M);return/\d/.test(m)}function p1(f){if(A>=f.length)return!1;let m=f[A];if(q1.test(m))return!0;if(m==="-"){if(f.length-A<2)return!1;let M=f[A+1];if(M==="-"||q1.test(M))return!0;return!1}return!1}var G8={deg:1,rad:180/Math.PI,grad:0.9,turn:360};function Mf(f){let m="";if(f[A]==="-"||f[A]==="+")m+=f[A++];if(m+=Af(f),f[A]==="."&&/\d/.test(f[A+1]))m+=f[A++]+Af(f);if(f[A]==="e"||f[A]==="E"){if((f[A+1]==="-"||f[A+1]==="+")&&/\d/.test(f[A+2]))m+=f[A++]+f[A++]+Af(f);else if(/\d/.test(f[A+1]))m+=f[A++]+Af(f)}if(p1(f)){let M=If(f);if(M==="deg"||M==="rad"||M==="turn"||M==="grad")return{type:x.Hue,value:m*G8[M]};return}if(f[A]==="%")return A++,{type:x.Percentage,value:+m};return{type:x.Number,value:+m}}function Af(f){let m="";while(/\d/.test(f[A]))m+=f[A++];return m}function If(f){let m="";while(A<f.length&&Y8.test(f[A]))m+=f[A++];return m}function F8(f){let m=If(f);if(f[A]==="(")return A++,{type:x.Function,value:m};if(m==="none")return{type:x.None,value:void 0};return{type:x.Ident,value:m}}function C8(f=""){let m=f.trim(),M=[],$;A=0;while(A<m.length){if($=m[A++],$===`
 `||$==="\t"||$===" "){while(A<m.length&&(m[A]===`
-`||m[A]==="\t"||m[A]===" "))A++;continue}if($===",")return;if($===")"){M.push({type:x.ParenClose});continue}if($==="+"){if(A--,Of(m)){M.push(Mf(m));continue}return}if($==="-"){if(A--,Of(m)){M.push(Mf(m));continue}if(x1(m)){M.push({type:x.Ident,value:If(m)});continue}return}if($==="."){if(A--,Of(m)){M.push(Mf(m));continue}return}if($==="/"){while(A<m.length&&(m[A]===`
-`||m[A]==="\t"||m[A]===" "))A++;let H;if(Of(m)){if(H=Mf(m),H.type!==x.Hue){M.push({type:x.Alpha,value:H});continue}}if(x1(m)){if(If(m)==="none"){M.push({type:x.Alpha,value:{type:x.None,value:void 0}});continue}}return}if(/\d/.test($)){A--,M.push(Mf(m));continue}if(p1.test($)){A--,M.push(F8(m));continue}return}return M}function K8(f){f._i=0;let m=f[f._i++];if(!m||m.type!==x.Function||m.value!=="color")return;if(m=f[f._i++],m.type!==x.Ident)return;let M=q1[m.value];if(!M)return;let $={mode:M},H=E6(f,!1);if(!H)return;let Z=w0(M).channels;for(let z=0,X,J;z<Z.length;z++)if(X=H[z],J=Z[z],X.type!==x.None){if($[J]=X.type===x.Number?X.value:X.value/100,J==="alpha")$[J]=Math.max(0,Math.min(1,$[J]))}return $}function E6(f,m){let M=[],$;while(f._i<f.length){if($=f[f._i++],$.type===x.None||$.type===x.Number||$.type===x.Alpha||$.type===x.Percentage||m&&$.type===x.Hue){M.push($);continue}if($.type===x.ParenClose){if(f._i<f.length)return;continue}return}if(M.length<3||M.length>4)return;if(M.length===4){if(M[3].type!==x.Alpha)return;M[3]=M[3].value}if(M.length===3)M.push({type:x.None,value:void 0});return M.every((H)=>H.type!==x.Alpha)?M:void 0}function O8(f,m){f._i=0;let M=f[f._i++];if(!M||M.type!==x.Function)return;let $=E6(f,m);if(!$)return;return $.unshift(M.value),$}var A8=(f)=>{if(typeof f!=="string")return;let m=C8(f),M=m?O8(m,!0):void 0,$=void 0,H=0,Z=mf.length;while(H<Z)if(($=mf[H++](f,M))!==void 0)return $;return m?K8(m):void 0},T6=A8;function I8(f,m){if(!m||m[0]!=="rgb"&&m[0]!=="rgba")return;let M={mode:"rgb"},[,$,H,Z,z]=m;if($.type===x.Hue||H.type===x.Hue||Z.type===x.Hue)return;if($.type!==x.None)M.r=$.type===x.Number?$.value/255:$.value/100;if(H.type!==x.None)M.g=H.type===x.Number?H.value/255:H.value/100;if(Z.type!==x.None)M.b=Z.type===x.Number?Z.value/255:Z.value/100;if(z.type!==x.None)M.alpha=Math.min(1,Math.max(0,z.type===x.Number?z.value:z.value/100));return M}var _6=I8;var y8=(f)=>f==="transparent"?{mode:"rgb",r:0,g:0,b:0,alpha:0}:void 0,S6=y8;var d6=(f,m,M)=>f+M*(m-f);var w8=(f)=>{let m=[];for(let M=0;M<f.length-1;M++){let $=f[M],H=f[M+1];if($===void 0&&H===void 0)m.push(void 0);else if($!==void 0&&H!==void 0)m.push([$,H]);else m.push($!==void 0?[$,$]:[H,H])}return m},k6=(f)=>(m)=>{let M=w8(m);return($)=>{let H=$*M.length,Z=$>=1?M.length-1:Math.max(Math.floor(H),0),z=M[Z];return z===void 0?void 0:f(z[0],z[1],H-Z)}};var T=k6(d6);var B0=(f)=>{let m=!1,M=f.map(($)=>{if($!==void 0)return m=!0,$;return 1});return m?M:f};var u8={mode:"rgb",channels:["r","g","b","alpha"],parse:[_6,y6,u6,I6,S6,"srgb"],serialize:"srgb",interpolate:{r:T,g:T,b:T,alpha:{use:T,fixup:B0}},gamut:!0,white:{r:1,g:1,b:1},black:{r:0,g:0,b:0}},Y0=u8;var N1=(f=0)=>Math.pow(Math.abs(f),2.19921875)*Math.sign(f),T8=(f)=>{let m=N1(f.r),M=N1(f.g),$=N1(f.b),H={mode:"xyz65",x:0.5766690429101305*m+0.1855582379065463*M+0.1882286462349947*$,y:0.297344975250536*m+0.6273635662554661*M+0.0752914584939979*$,z:0.0270313613864123*m+0.0706888525358272*M+0.9913375368376386*$};if(f.alpha!==void 0)H.alpha=f.alpha;return H},L1=T8;var R1=(f)=>Math.pow(Math.abs(f),0.4547069271758437)*Math.sign(f),g8=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H={mode:"a98",r:R1(f*2.0415879038107465-m*0.5650069742788597-0.3447313507783297*M),g:R1(f*-0.9692436362808798+m*1.8759675015077206+0.0415550574071756*M),b:R1(f*0.0134442806320312-m*0.1183623922310184+1.0151749943912058*M)};if($!==void 0)H.alpha=$;return H},D1=g8;var V1=(f=0)=>{let m=Math.abs(f);if(m<=0.04045)return f/12.92;return(Math.sign(f)||1)*Math.pow((m+0.055)/1.055,2.4)},E8=({r:f,g:m,b:M,alpha:$})=>{let H={mode:"lrgb",r:V1(f),g:V1(m),b:V1(M)};if($!==void 0)H.alpha=$;return H},u0=E8;var _8=(f)=>{let{r:m,g:M,b:$,alpha:H}=u0(f),Z={mode:"xyz65",x:0.4123907992659593*m+0.357584339383878*M+0.1804807884018343*$,y:0.2126390058715102*m+0.715168678767756*M+0.0721923153607337*$,z:0.0193308187155918*m+0.119194779794626*M+0.9505321522496607*$};if(H!==void 0)Z.alpha=H;return Z},T0=_8;var Y1=(f=0)=>{let m=Math.abs(f);if(m>0.0031308)return(Math.sign(f)||1)*(1.055*Math.pow(m,0.4166666666666667)-0.055);return f*12.92},S8=({r:f,g:m,b:M,alpha:$},H="rgb")=>{let Z={mode:H,r:Y1(f),g:Y1(m),b:Y1(M)};if($!==void 0)Z.alpha=$;return Z},g0=S8;var d8=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=g0({r:f*3.2409699419045226-m*1.537383177570094-0.4986107602930034*M,g:f*-0.9692436362808796+m*1.8759675015077204+0.0415550574071756*M,b:f*0.0556300796969936-m*0.2039769588889765+1.0569715142428784*M});if($!==void 0)H.alpha=$;return H},E0=d8;var k8={...Y0,mode:"a98",parse:["a98-rgb"],serialize:"a98-rgb",fromMode:{rgb:(f)=>D1(T0(f)),xyz65:D1},toMode:{rgb:(f)=>E0(L1(f)),xyz65:L1}},v6=k8;var v8=(f)=>(f=f%360)<0?f+360:f,X0=v8;function G1({h:f,s:m,l:M,alpha:$}){if(f=X0(f!==void 0?f:0),m===void 0)m=0;if(M===void 0)M=0;let H=M+m*(M<0.5?M:1-M),Z=H-(H-M)*2*Math.abs(f/60%2-1),z;switch(Math.floor(f/60)){case 0:z={r:H,g:Z,b:2*M-H};break;case 1:z={r:Z,g:H,b:2*M-H};break;case 2:z={r:2*M-H,g:H,b:Z};break;case 3:z={r:2*M-H,g:Z,b:H};break;case 4:z={r:Z,g:2*M-H,b:H};break;case 5:z={r:H,g:2*M-H,b:Z};break;default:z={r:2*M-H,g:2*M-H,b:2*M-H}}if(z.mode="rgb",$!==void 0)z.alpha=$;return z}function F1({r:f,g:m,b:M,alpha:$}){if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=Math.max(f,m,M),Z=Math.min(f,m,M),z={mode:"hsl",s:H===Z?0:(H-Z)/(1-Math.abs(H+Z-1)),l:0.5*(H+Z)};if(H-Z!==0)z.h=(H===f?(m-M)/(H-Z)+(m<M)*6:H===m?(M-f)/(H-Z)+2:(f-m)/(H-Z)+4)*60;if($!==void 0)z.alpha=$;return z}var o8=(f,m)=>{switch(m){case"deg":return+f;case"rad":return f/Math.PI*180;case"grad":return f/10*9;case"turn":return f*360}},o6=o8;var r8=new RegExp(`^hsla?\\(\\s*${w6}${f5}${x5}${f5}${x5}\\s*(?:,\\s*${ff}\\s*)?\\)$`),n8=(f)=>{let m=f.match(r8);if(!m)return;let M={mode:"hsl"};if(m[3]!==void 0)M.h=+m[3];else if(m[1]!==void 0&&m[2]!==void 0)M.h=o6(m[1],m[2]);if(m[4]!==void 0)M.s=Math.min(Math.max(0,m[4]/100),1);if(m[5]!==void 0)M.l=Math.min(Math.max(0,m[5]/100),1);if(m[6]!==void 0)M.alpha=Math.max(0,Math.min(1,m[6]/100));else if(m[7]!==void 0)M.alpha=Math.max(0,Math.min(1,+m[7]));return M},r6=n8;function l8(f,m){if(!m||m[0]!=="hsl"&&m[0]!=="hsla")return;let M={mode:"hsl"},[,$,H,Z,z]=m;if($.type!==x.None){if($.type===x.Percentage)return;M.h=$.value}if(H.type!==x.None){if(H.type===x.Hue)return;M.s=H.value/100}if(Z.type!==x.None){if(Z.type===x.Hue)return;M.l=Z.value/100}if(z.type!==x.None)M.alpha=Math.min(1,Math.max(0,z.type===x.Number?z.value:z.value/100));return M}var n6=l8;var c8=(f,m)=>{return f.map((M,$,H)=>{if(M===void 0)return M;let Z=X0(M);if($===0||f[$-1]===void 0)return Z;return m(Z-X0(H[$-1]))}).reduce((M,$)=>{if(!M.length||$===void 0||M[M.length-1]===void 0)return M.push($),M;return M.push($+M[M.length-1]),M},[])},M5=(f)=>c8(f,(m)=>Math.abs(m)<=180?m:m-360*Math.sign(m));var $f=(f,m)=>{if(f.h===void 0||m.h===void 0||!f.s||!m.s)return 0;let M=X0(f.h),$=X0(m.h),H=Math.sin(($-M+360)/2*Math.PI/180);return 2*Math.sqrt(f.s*m.s)*H},C1=(f,m)=>{if(f.h===void 0||m.h===void 0)return 0;let M=X0(f.h),$=X0(m.h);if(Math.abs($-M)>180)return M-($-360*Math.sign($-M));return $-M},K1=(f,m)=>{if(f.h===void 0||m.h===void 0||!f.c||!m.c)return 0;let M=X0(f.h),$=X0(m.h),H=Math.sin(($-M+360)/2*Math.PI/180);return 2*Math.sqrt(f.c*m.c)*H},O1=(f="rgb",m=[1,1,1,0])=>{let M=w0(f),$=M.channels,H=M.difference,Z=e(f);return(z,X)=>{let J=Z(z),B=Z(X);return Math.sqrt($.reduce((W,Q,j)=>{let b=H[Q]?H[Q](J,B):J[Q]-B[Q];return W+(m[j]||0)*Math.pow(isNaN(b)?0:b,2)},0))}};var A1=(f=1,m=1,M=1)=>{let $=e("lab65");return(H,Z)=>{let z=$(H),X=$(Z),J=z.l,B=z.a,W=z.b,Q=Math.sqrt(B*B+W*W),j=X.l,b=X.a,U=X.b,p=Math.sqrt(b*b+U*U),N=(Q+p)/2,K=0.5*(1-Math.sqrt(Math.pow(N,7)/(Math.pow(N,7)+Math.pow(25,7)))),u=B*(1+K),o=b*(1+K),D=Math.sqrt(u*u+W*W),Y=Math.sqrt(o*o+U*U),C=Math.abs(u)+Math.abs(W)===0?0:Math.atan2(W,u);C+=(C<0)*2*Math.PI;let G=Math.abs(o)+Math.abs(U)===0?0:Math.atan2(U,o);G+=(G<0)*2*Math.PI;let I=j-J,r=Y-D,d=D*Y===0?0:G-C;d-=(d>Math.PI)*2*Math.PI,d+=(d<-Math.PI)*2*Math.PI;let l=2*Math.sqrt(D*Y)*Math.sin(d/2),J0=(J+j)/2,j0=(D+Y)/2,m0;if(D*Y===0)m0=C+G;else m0=(C+G)/2,m0-=(Math.abs(C-G)>Math.PI)*Math.PI,m0+=(m0<0)*2*Math.PI;let P=Math.pow(J0-50,2),q=1-0.17*Math.cos(m0-Math.PI/6)+0.24*Math.cos(2*m0)+0.32*Math.cos(3*m0+Math.PI/30)-0.2*Math.cos(4*m0-63*Math.PI/180),V=1+0.015*P/Math.sqrt(20+P),_=1+0.045*j0,L=1+0.015*j0*q,F=30*Math.PI/180*Math.exp(-1*Math.pow((180/Math.PI*m0-275)/25,2)),y=2*Math.sqrt(Math.pow(j0,7)/(Math.pow(j0,7)+Math.pow(25,7))),O=-1*Math.sin(2*F)*y;return Math.sqrt(Math.pow(I/(f*V),2)+Math.pow(r/(m*_),2)+Math.pow(l/(M*L),2)+O*r/(m*_)*l/(M*L))}};var $5=(f)=>{let m=f.reduce(($,H)=>{if(H!==void 0){let Z=H*Math.PI/180;$.sin+=Math.sin(Z),$.cos+=Math.cos(Z)}return $},{sin:0,cos:0}),M=Math.atan2(m.sin,m.cos)*180/Math.PI;return M<0?360+M:M};var i8={mode:"hsl",toMode:{rgb:G1},fromMode:{rgb:F1},channels:["h","s","l","alpha"],ranges:{h:[0,360]},gamut:"rgb",parse:[n6,r6],serialize:(f)=>`hsl(${f.h!==void 0?f.h:"none"} ${f.s!==void 0?f.s*100+"%":"none"} ${f.l!==void 0?f.l*100+"%":"none"}${f.alpha<1?` / ${f.alpha}`:""})`,interpolate:{h:{use:T,fixup:M5},s:T,l:T,alpha:{use:T,fixup:B0}},difference:{h:$f},average:{h:$5}},l6=i8;function Hf({h:f,s:m,v:M,alpha:$}){if(f=X0(f!==void 0?f:0),m===void 0)m=0;if(M===void 0)M=0;let H=Math.abs(f/60%2-1),Z;switch(Math.floor(f/60)){case 0:Z={r:M,g:M*(1-m*H),b:M*(1-m)};break;case 1:Z={r:M*(1-m*H),g:M,b:M*(1-m)};break;case 2:Z={r:M*(1-m),g:M,b:M*(1-m*H)};break;case 3:Z={r:M*(1-m),g:M*(1-m*H),b:M};break;case 4:Z={r:M*(1-m*H),g:M*(1-m),b:M};break;case 5:Z={r:M,g:M*(1-m),b:M*(1-m*H)};break;default:Z={r:M*(1-m),g:M*(1-m),b:M*(1-m)}}if(Z.mode="rgb",$!==void 0)Z.alpha=$;return Z}function Zf({r:f,g:m,b:M,alpha:$}){if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=Math.max(f,m,M),Z=Math.min(f,m,M),z={mode:"hsv",s:H===0?0:1-Z/H,v:H};if(H-Z!==0)z.h=(H===f?(m-M)/(H-Z)+(m<M)*6:H===m?(M-f)/(H-Z)+2:(f-m)/(H-Z)+4)*60;if($!==void 0)z.alpha=$;return z}var h8={mode:"hsv",toMode:{rgb:Hf},parse:["--hsv"],serialize:"--hsv",fromMode:{rgb:Zf},channels:["h","s","v","alpha"],ranges:{h:[0,360]},gamut:"rgb",interpolate:{h:{use:T,fixup:M5},s:T,v:T,alpha:{use:T,fixup:B0}},difference:{h:$f},average:{h:$5}},c6=h8;function I1({h:f,w:m,b:M,alpha:$}){if(m===void 0)m=0;if(M===void 0)M=0;if(m+M>1){let H=m+M;m/=H,M/=H}return Hf({h:f,s:M===1?1:1-m/(1-M),v:1-M,alpha:$})}function y1(f){let m=Zf(f);if(m===void 0)return;let M=m.s!==void 0?m.s:0,$=m.v!==void 0?m.v:0,H={mode:"hwb",w:(1-M)*$,b:1-$};if(m.h!==void 0)H.h=m.h;if(m.alpha!==void 0)H.alpha=m.alpha;return H}function t8(f,m){if(!m||m[0]!=="hwb")return;let M={mode:"hwb"},[,$,H,Z,z]=m;if($.type!==x.None){if($.type===x.Percentage)return;M.h=$.value}if(H.type!==x.None){if(H.type===x.Hue)return;M.w=H.value/100}if(Z.type!==x.None){if(Z.type===x.Hue)return;M.b=Z.value/100}if(z.type!==x.None)M.alpha=Math.min(1,Math.max(0,z.type===x.Number?z.value:z.value/100));return M}var i6=t8;var a8={mode:"hwb",toMode:{rgb:I1},fromMode:{rgb:y1},channels:["h","w","b","alpha"],ranges:{h:[0,360]},gamut:"rgb",parse:[i6],serialize:(f)=>`hwb(${f.h!==void 0?f.h:"none"} ${f.w!==void 0?f.w*100+"%":"none"} ${f.b!==void 0?f.b*100+"%":"none"}${f.alpha<1?` / ${f.alpha}`:""})`,interpolate:{h:{use:T,fixup:M5},w:T,b:T,alpha:{use:T,fixup:B0}},difference:{h:C1},average:{h:$5}},h6=a8;var yf=Math.pow(29,3)/Math.pow(3,3),wf=Math.pow(6,3)/Math.pow(29,3);var H5={X:0.9642956764295677,Y:1,Z:0.8251046025104602},Z5={X:0.9504559270516716,Y:1,Z:1.0890577507598784},b$=Math.pow(29,3)/Math.pow(3,3),Q$=Math.pow(6,3)/Math.pow(29,3);var w1=(f)=>Math.pow(f,3)>wf?Math.pow(f,3):(116*f-16)/yf,s8=({l:f,a:m,b:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=(f+16)/116,Z=m/500+H,z=H-M/200,X={mode:"xyz50",x:w1(Z)*H5.X,y:w1(H)*H5.Y,z:w1(z)*H5.Z};if($!==void 0)X.alpha=$;return X},A5=s8;var e8=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=g0({r:f*3.1341359569958707-m*1.6173863321612538-0.4906619460083532*M,g:f*-0.978795502912089+m*1.916254567259524+0.03344273116131949*M,b:f*0.07195537988411677-m*0.2289768264158322+1.405386058324125*M});if($!==void 0)H.alpha=$;return H},I5=e8;var f9=(f)=>I5(A5(f)),uf=f9;var m9=(f)=>{let{r:m,g:M,b:$,alpha:H}=u0(f),Z={mode:"xyz50",x:0.436065742824811*m+0.3851514688337912*M+0.14307845442264197*$,y:0.22249319175623702*m+0.7168870538238823*M+0.06061979053616537*$,z:0.013923904500943465*m+0.09708128566574634*M+0.7140993584005155*$};if(H!==void 0)Z.alpha=H;return Z},y5=m9;var u1=(f)=>f>wf?Math.cbrt(f):(yf*f+16)/116,M9=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=u1(f/H5.X),Z=u1(m/H5.Y),z=u1(M/H5.Z),X={mode:"lab",l:116*Z-16,a:500*(H-Z),b:200*(Z-z)};if($!==void 0)X.alpha=$;return X},w5=M9;var $9=(f)=>{let m=w5(y5(f));if(f.r===f.b&&f.b===f.g)m.a=m.b=0;return m},Tf=$9;function H9(f,m){if(!m||m[0]!=="lab")return;let M={mode:"lab"},[,$,H,Z,z]=m;if($.type===x.Hue||H.type===x.Hue||Z.type===x.Hue)return;if($.type!==x.None)M.l=Math.min(Math.max(0,$.value),100);if(H.type!==x.None)M.a=H.type===x.Number?H.value:H.value*125/100;if(Z.type!==x.None)M.b=Z.type===x.Number?Z.value:Z.value*125/100;if(z.type!==x.None)M.alpha=Math.min(1,Math.max(0,z.type===x.Number?z.value:z.value/100));return M}var t6=H9;var Z9={mode:"lab",toMode:{xyz50:A5,rgb:uf},fromMode:{xyz50:w5,rgb:Tf},channels:["l","a","b","alpha"],ranges:{l:[0,100],a:[-125,125],b:[-125,125]},parse:[t6],serialize:(f)=>`lab(${f.l!==void 0?f.l:"none"} ${f.a!==void 0?f.a:"none"} ${f.b!==void 0?f.b:"none"}${f.alpha<1?` / ${f.alpha}`:""})`,interpolate:{l:T,a:T,b:T,alpha:{use:T,fixup:B0}}},u5=Z9;var gf=Math.pow(29,3)/Math.pow(3,3),Ef=Math.pow(6,3)/Math.pow(29,3);var T1=(f)=>Math.pow(f,3)>Ef?Math.pow(f,3):(116*f-16)/gf,z9=({l:f,a:m,b:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=(f+16)/116,Z=m/500+H,z=H-M/200,X={mode:"xyz65",x:T1(Z)*Z5.X,y:T1(H)*Z5.Y,z:T1(z)*Z5.Z};if($!==void 0)X.alpha=$;return X},_f=z9;var B9=(f)=>E0(_f(f)),Sf=B9;var g1=(f)=>f>Ef?Math.cbrt(f):(gf*f+16)/116,X9=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=g1(f/Z5.X),Z=g1(m/Z5.Y),z=g1(M/Z5.Z),X={mode:"lab65",l:116*Z-16,a:500*(H-Z),b:200*(Z-z)};if($!==void 0)X.alpha=$;return X},df=X9;var J9=(f)=>{let m=df(T0(f));if(f.r===f.b&&f.b===f.g)m.a=m.b=0;return m},kf=J9;var W9={...u5,mode:"lab65",parse:["--lab-d65"],serialize:"--lab-d65",toMode:{xyz65:_f,rgb:Sf},fromMode:{xyz65:df,rgb:kf},ranges:{l:[0,100],a:[-125,125],b:[-125,125]}},a6=W9;var b9=({l:f,a:m,b:M,alpha:$},H="lch")=>{if(m===void 0)m=0;if(M===void 0)M=0;let Z=Math.sqrt(m*m+M*M),z={mode:H,l:f,c:Z};if(Z)z.h=X0(Math.atan2(M,m)*180/Math.PI);if($!==void 0)z.alpha=$;return z},l0=b9;var Q9=({l:f,c:m,h:M,alpha:$},H="lab")=>{if(M===void 0)M=0;let Z={mode:H,l:f,a:m?m*Math.cos(M/180*Math.PI):0,b:m?m*Math.sin(M/180*Math.PI):0};if($!==void 0)Z.alpha=$;return Z},c0=Q9;function U9(f,m){if(!m||m[0]!=="lch")return;let M={mode:"lch"},[,$,H,Z,z]=m;if($.type!==x.None){if($.type===x.Hue)return;M.l=Math.min(Math.max(0,$.value),100)}if(H.type!==x.None)M.c=Math.max(0,H.type===x.Number?H.value:H.value*150/100);if(Z.type!==x.None){if(Z.type===x.Percentage)return;M.h=Z.value}if(z.type!==x.None)M.alpha=Math.min(1,Math.max(0,z.type===x.Number?z.value:z.value/100));return M}var s6=U9;var j9={mode:"lch",toMode:{lab:c0,rgb:(f)=>uf(c0(f))},fromMode:{rgb:(f)=>l0(Tf(f)),lab:l0},channels:["l","c","h","alpha"],ranges:{l:[0,100],c:[0,150],h:[0,360]},parse:[s6],serialize:(f)=>`lch(${f.l!==void 0?f.l:"none"} ${f.c!==void 0?f.c:"none"} ${f.h!==void 0?f.h:"none"}${f.alpha<1?` / ${f.alpha}`:""})`,interpolate:{h:{use:T,fixup:M5},c:T,l:T,alpha:{use:T,fixup:B0}},difference:{h:K1},average:{h:$5}},T5=j9;var P9={...T5,mode:"lch65",parse:["--lch-d65"],serialize:"--lch-d65",toMode:{lab65:(f)=>c0(f,"lab65"),rgb:(f)=>Sf(c0(f,"lab65"))},fromMode:{rgb:(f)=>l0(kf(f),"lch65"),lab65:(f)=>l0(f,"lch65")},ranges:{l:[0,100],c:[0,150],h:[0,360]}},e6=P9;var q9={...Y0,mode:"lrgb",toMode:{rgb:g0},fromMode:{rgb:u0},parse:["srgb-linear"],serialize:"srgb-linear"},f2=q9;var p9=({l:f,a:m,b:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=Math.pow(f+0.3963377773761749*m+0.2158037573099136*M,3),Z=Math.pow(f-0.1055613458156586*m-0.0638541728258133*M,3),z=Math.pow(f-0.0894841775298119*m-1.2914855480194092*M,3),X={mode:"lrgb",r:4.076741636075957*H-3.3077115392580616*Z+0.2309699031821044*z,g:-1.2684379732850317*H+2.6097573492876887*Z-0.3413193760026573*z,b:-0.0041960761386756*H-0.7034186179359362*Z+1.7076146940746117*z};if($!==void 0)X.alpha=$;return X},vf=p9;var x9=({r:f,g:m,b:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=Math.cbrt(0.412221469470763*f+0.5363325372617348*m+0.0514459932675022*M),Z=Math.cbrt(0.2119034958178252*f+0.6806995506452344*m+0.1073969535369406*M),z=Math.cbrt(0.0883024591900564*f+0.2817188391361215*m+0.6299787016738222*M),X={mode:"oklab",l:0.210454268309314*H+0.7936177747023054*Z-0.0040720430116193*z,a:1.9779985324311684*H-2.42859224204858*Z+0.450593709617411*z,b:0.0259040424655478*H+0.7827717124575296*Z-0.8086757549230774*z};if($!==void 0)X.alpha=$;return X},of=x9;var N9=(f)=>{let m=of(u0(f));if(f.r===f.b&&f.b===f.g)m.a=m.b=0;return m},rf=N9;var L9=(f)=>g0(vf(f)),nf=L9;function R9(f,m){if(!m||m[0]!=="oklab")return;let M={mode:"oklab"},[,$,H,Z,z]=m;if($.type===x.Hue||H.type===x.Hue||Z.type===x.Hue)return;if($.type!==x.None)M.l=Math.min(Math.max(0,$.type===x.Number?$.value:$.value/100),1);if(H.type!==x.None)M.a=H.type===x.Number?H.value:H.value*0.4/100;if(Z.type!==x.None)M.b=Z.type===x.Number?Z.value:Z.value*0.4/100;if(z.type!==x.None)M.alpha=Math.min(1,Math.max(0,z.type===x.Number?z.value:z.value/100));return M}var m2=R9;var D9={...u5,mode:"oklab",toMode:{lrgb:vf,rgb:nf},fromMode:{lrgb:of,rgb:rf},ranges:{l:[0,1],a:[-0.4,0.4],b:[-0.4,0.4]},parse:[m2],serialize:(f)=>`oklab(${f.l!==void 0?f.l:"none"} ${f.a!==void 0?f.a:"none"} ${f.b!==void 0?f.b:"none"}${f.alpha<1?` / ${f.alpha}`:""})`},M2=D9;function V9(f,m){if(!m||m[0]!=="oklch")return;let M={mode:"oklch"},[,$,H,Z,z]=m;if($.type!==x.None){if($.type===x.Hue)return;M.l=Math.min(Math.max(0,$.type===x.Number?$.value:$.value/100),1)}if(H.type!==x.None)M.c=Math.max(0,H.type===x.Number?H.value:H.value*0.4/100);if(Z.type!==x.None){if(Z.type===x.Percentage)return;M.h=Z.value}if(z.type!==x.None)M.alpha=Math.min(1,Math.max(0,z.type===x.Number?z.value:z.value/100));return M}var $2=V9;var Y9={...T5,mode:"oklch",toMode:{oklab:(f)=>c0(f,"oklab"),rgb:(f)=>nf(c0(f,"oklab"))},fromMode:{rgb:(f)=>l0(rf(f),"oklch"),oklab:(f)=>l0(f,"oklch")},parse:[$2],serialize:(f)=>`oklch(${f.l!==void 0?f.l:"none"} ${f.c!==void 0?f.c:"none"} ${f.h!==void 0?f.h:"none"}${f.alpha<1?` / ${f.alpha}`:""})`,ranges:{l:[0,1],c:[0,0.4],h:[0,360]}},H2=Y9;var G9=(f)=>{let{r:m,g:M,b:$,alpha:H}=u0(f),Z={mode:"xyz65",x:0.486570948648216*m+0.265667693169093*M+0.1982172852343625*$,y:0.2289745640697487*m+0.6917385218365062*M+0.079286914093745*$,z:0*m+0.0451133818589026*M+1.043944368900976*$};if(H!==void 0)Z.alpha=H;return Z},E1=G9;var F9=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=g0({r:f*2.4934969119414263-m*0.9313836179191242-0.402710784450717*M,g:f*-0.8294889695615749+m*1.7626640603183465+0.0236246858419436*M,b:f*0.0358458302437845-m*0.0761723892680418+0.9568845240076871*M},"p3");if($!==void 0)H.alpha=$;return H},_1=F9;var C9={...Y0,mode:"p3",parse:["display-p3"],serialize:"display-p3",fromMode:{rgb:(f)=>_1(T0(f)),xyz65:_1},toMode:{rgb:(f)=>E0(E1(f)),xyz65:E1}},Z2=C9;var S1=(f)=>{let m=Math.abs(f);if(m>=0.001953125)return Math.sign(f)*Math.pow(m,0.5555555555555556);return 16*f},K9=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H={mode:"prophoto",r:S1(f*1.3457868816471585-m*0.2555720873797946-0.0511018649755453*M),g:S1(f*-0.5446307051249019+m*1.5082477428451466+0.0205274474364214*M),b:S1(f*0+m*0+1.2119675456389452*M)};if($!==void 0)H.alpha=$;return H},d1=K9;var k1=(f=0)=>{let m=Math.abs(f);if(m>=0.03125)return Math.sign(f)*Math.pow(m,1.8);return f/16},O9=(f)=>{let m=k1(f.r),M=k1(f.g),$=k1(f.b),H={mode:"xyz50",x:0.7977666449006423*m+0.1351812974005331*M+0.0313477341283922*$,y:0.2880748288194013*m+0.7118352342418731*M+0.0000899369387256*$,z:0*m+0*M+0.8251046025104602*$};if(f.alpha!==void 0)H.alpha=f.alpha;return H},v1=O9;var A9={...Y0,mode:"prophoto",parse:["prophoto-rgb"],serialize:"prophoto-rgb",fromMode:{xyz50:d1,rgb:(f)=>d1(y5(f))},toMode:{xyz50:v1,rgb:(f)=>I5(v1(f))}},z2=A9;var o1=(f)=>{let m=Math.abs(f);if(m>0.018053968510807)return(Math.sign(f)||1)*(1.09929682680944*Math.pow(m,0.45)-0.09929682680944008);return 4.5*f},I9=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H={mode:"rec2020",r:o1(f*1.7166511879712683-m*0.3556707837763925-0.2533662813736599*M),g:o1(f*-0.6666843518324893+m*1.6164812366349395+0.0157685458139111*M),b:o1(f*0.0176398574453108-m*0.0427706132578085+0.9421031212354739*M)};if($!==void 0)H.alpha=$;return H},r1=I9;var n1=(f=0)=>{let m=Math.abs(f);if(m<0.08124285829863151)return f/4.5;return(Math.sign(f)||1)*Math.pow((m+1.09929682680944-1)/1.09929682680944,2.2222222222222223)},y9=(f)=>{let m=n1(f.r),M=n1(f.g),$=n1(f.b),H={mode:"xyz65",x:0.6369580483012911*m+0.1446169035862083*M+0.1688809751641721*$,y:0.262700212011267*m+0.6779980715188708*M+0.059301716469862*$,z:0*m+0.0280726930490874*M+1.0609850577107909*$};if(f.alpha!==void 0)H.alpha=f.alpha;return H},l1=y9;var w9={...Y0,mode:"rec2020",fromMode:{xyz65:r1,rgb:(f)=>r1(T0(f))},toMode:{xyz65:l1,rgb:(f)=>E0(l1(f))},parse:["rec2020"],serialize:"rec2020"},B2=w9;var u9={mode:"xyz50",parse:["xyz-d50"],serialize:"xyz-d50",toMode:{rgb:I5,lab:w5},fromMode:{rgb:y5,lab:A5},channels:["x","y","z","alpha"],ranges:{x:[0,0.964],y:[0,0.999],z:[0,0.825]},interpolate:{x:T,y:T,z:T,alpha:{use:T,fixup:B0}}},X2=u9;var T9=(f)=>{let{x:m,y:M,z:$,alpha:H}=f;if(m===void 0)m=0;if(M===void 0)M=0;if($===void 0)$=0;let Z={mode:"xyz50",x:1.0479298208405488*m+0.0229467933410191*M-0.0501922295431356*$,y:0.0296278156881593*m+0.990434484573249*M-0.0170738250293851*$,z:-0.0092430581525912*m+0.0150551448965779*M+0.7518742899580008*$};if(H!==void 0)Z.alpha=H;return Z},J2=T9;var g9=(f)=>{let{x:m,y:M,z:$,alpha:H}=f;if(m===void 0)m=0;if(M===void 0)M=0;if($===void 0)$=0;let Z={mode:"xyz65",x:0.9554734527042182*m-0.0230985368742614*M+0.0632593086610217*$,y:-0.0283697069632081*m+1.0099954580058226*M+0.021041398966943*$,z:0.0123140016883199*m-0.0205076964334779*M+1.3303659366080753*$};if(H!==void 0)Z.alpha=H;return Z},W2=g9;var E9={mode:"xyz65",toMode:{rgb:E0,xyz50:J2},fromMode:{rgb:T0,xyz50:W2},ranges:{x:[0,0.95],y:[0,1],z:[0,1.088]},channels:["x","y","z","alpha"],parse:["xyz","xyz-d65"],serialize:"xyz-d65",interpolate:{x:T,y:T,z:T,alpha:{use:T,fixup:B0}}},b2=E9;var Jz=f0(v6),Wz=f0(l6),bz=f0(c6),Qz=f0(h6),Uz=f0(u5),jz=f0(a6),Pz=f0(T5),qz=f0(e6),pz=f0(f2),xz=f0(M2),Nz=f0(H2),Lz=f0(Z2),Rz=f0(z2),Dz=f0(B2),Vz=f0(Y0),Yz=f0(X2),Gz=f0(b2);var _9=(f,m)=>Math.round(f*(m=Math.pow(10,m)))/m,S9=(f=4)=>(m)=>typeof m==="number"?_9(m,f):m,Q2=S9;var zf=Q2(2),Bf=(f)=>Math.max(0,Math.min(1,f||0)),g5=(f)=>Math.round(Bf(f)*255),U2=e("rgb"),d9=e("hsl"),j2=(f)=>{if(f===void 0)return;let m=g5(f.r),M=g5(f.g),$=g5(f.b);return"#"+(16777216|m<<16|M<<8|$).toString(16).slice(1)};var P2=(f)=>{if(f===void 0)return;let m=g5(f.r),M=g5(f.g),$=g5(f.b);if(f.alpha===void 0||f.alpha===1)return`rgb(${m}, ${M}, ${$})`;else return`rgba(${m}, ${M}, ${$}, ${zf(Bf(f.alpha))})`},q2=(f)=>{if(f===void 0)return;let m=zf(f.h||0),M=zf(Bf(f.s)*100)+"%",$=zf(Bf(f.l)*100)+"%";if(f.alpha===void 0||f.alpha===1)return`hsl(${m}, ${M}, ${$})`;else return`hsla(${m}, ${M}, ${$}, ${zf(Bf(f.alpha))})`},p0=(f)=>{let m=m5(f);if(!m)return;let M=w0(m.mode);if(!M.serialize||typeof M.serialize==="string"){let $=`color(${M.serialize||`--${m.mode}`} `;if(M.channels.forEach((H,Z)=>{if(H!=="alpha")$+=(Z?" ":"")+(m[H]!==void 0?m[H]:"none")}),m.alpha!==void 0&&m.alpha<1)$+=` / ${m.alpha}`;return $+")"}if(typeof M.serialize==="function")return M.serialize(m);return},Xf=(f)=>j2(U2(f));var c1=(f)=>P2(U2(f)),i1=(f)=>q2(d9(f));var p2=e("rgb"),x2=(f)=>{let m={mode:f.mode,r:Math.max(0,Math.min(f.r!==void 0?f.r:0,1)),g:Math.max(0,Math.min(f.g!==void 0?f.g:0,1)),b:Math.max(0,Math.min(f.b!==void 0?f.b:0,1))};if(f.alpha!==void 0)m.alpha=f.alpha;return m},k9=(f)=>x2(p2(f)),N2=(f)=>{return f!==void 0&&(f.r===void 0||f.r>=0&&f.r<=1)&&(f.g===void 0||f.g>=0&&f.g<=1)&&(f.b===void 0||f.b>=0&&f.b<=1)};function L2(f){return N2(p2(f))}function E5(f="rgb"){let{gamut:m}=w0(f);if(!m)return($)=>!0;let M=e(typeof m==="string"?m:f);return($)=>N2(M($))}function R2(f="rgb"){let{gamut:m}=w0(f);if(!m)return(Z)=>m5(Z);let M=typeof m==="string"?m:f,$=e(M),H=E5(M);return(Z)=>{let z=m5(Z);if(!z)return;let X=$(z);if(H(X))return z;let J=x2(X);if(z.mode===J.mode)return J;return e(z.mode)(J)}}function h1(f,m="lch",M="rgb"){f=m5(f);let $=M==="rgb"?L2:E5(M),H=M==="rgb"?k9:R2(M);if(f===void 0||$(f))return f;let Z=e(f.mode);f=e(m)(f);let z={...f,c:0};if(!$(z))return Z(H(z));let X=0,J=f.c!==void 0?f.c:0,B=w0(m).ranges.c,W=(B[1]-B[0])/Math.pow(2,13),Q=z.c;while(J-X>W)if(z.c=X+(J-X)*0.5,$(z))Q=z.c,X=z.c;else J=z.c;return Z($(z)?z:{...z,c:Q})}var v9=(f,m=O1(),M=($)=>$)=>{let $=f.map((H,Z)=>({color:M(H),i:Z}));return(H,Z=1,z=1/0)=>{if(isFinite(Z))Z=Math.max(1,Math.min(Z,$.length-1));return $.forEach((X)=>{X.d=m(H,X.color)}),$.sort((X,J)=>X.d-J.d).slice(0,Z).filter((X)=>X.d<z).map((X)=>f[X.i])}},t1=v9;var z5=(f={mode:"oklch",l:0.48,c:0.23,h:263})=>D0((m)=>(m?e("oklch")(m):f)??f);var _0=(f,m)=>{let M=2*Math.log((1-f.l)/f.l),$=f.l!==0.5?(Math.exp(M*m)-1)/(Math.exp(M)-1):m,H=f.c>0?f.c*(8*Math.sin(Math.PI*(4*m+1)/6)**3-1)/7:0;return{mode:"oklch",l:$,c:H,h:f.h??0}};var o9=0.71,AQ=R("card-colorscale",({expose:f,first:m,host:M,watch:$})=>{let H=m(".label strong","Add a <strong> element inside .label."),Z=m(".label small","Add a <small> element inside .label.");return f({name:H.textContent?.trim()??"",color:z5()}),[$("name",w(H)),$("color",(z)=>{Z.textContent=Xf(z);let X=new Map,J=z.l>o9,B=J?0.1:0.9;X.set("base",p0(z)),X.set("text",J?"black":"white"),X.set("text-soft",p0(_0(z,B)));for(let W=4;W>0;W--)X.set(`lighten${W*20}`,p0(_0(z,(5+W)/10)));for(let W=1;W<5;W++)X.set(`darken${W*20}`,p0(_0(z,(5-W)/10)));for(let[W,Q]of X)M.style.setProperty(`--color-${W}`,Q)})]});var lf="media-motion",cf="media-theme",hf="media-viewport",tf="media-orientation",wQ=R("context-media",({expose:f,host:m,provideContexts:M})=>{let $=(H,Z)=>{let X=m.getAttribute(H)?.trim();if(!X)return Z;let J=X.match(/em$/)?"em":"px",B=parseFloat(X);return Number.isFinite(B)?B+J:Z};return f({[lf]:Q5((H)=>{let Z=matchMedia("(prefers-reduced-motion: reduce)"),z=(X)=>H(X.matches?"reduce":"no-preference");return Z.addEventListener("change",z),()=>Z.removeEventListener("change",z)},{value:matchMedia("(prefers-reduced-motion: reduce)").matches?"reduce":"no-preference"}),[cf]:Q5((H)=>{let Z=matchMedia("(prefers-color-scheme: dark)"),z=(X)=>H(X.matches?"dark":"light");return Z.addEventListener("change",z),()=>Z.removeEventListener("change",z)},{value:matchMedia("(prefers-color-scheme: dark)").matches?"dark":"light"}),[hf]:(()=>{let H=matchMedia(`(min-width: ${$("sm","32em")})`),Z=matchMedia(`(min-width: ${$("md","48em")})`),z=matchMedia(`(min-width: ${$("lg","72em")})`),X=matchMedia(`(min-width: ${$("xl","104em")})`),J=()=>{if(X.matches)return"xl";if(z.matches)return"lg";if(Z.matches)return"md";if(H.matches)return"sm";return"xs"};return Q5((B)=>{let W=()=>B(J());return H.addEventListener("change",W),Z.addEventListener("change",W),z.addEventListener("change",W),X.addEventListener("change",W),()=>{H.removeEventListener("change",W),Z.removeEventListener("change",W),z.removeEventListener("change",W),X.removeEventListener("change",W)}},{value:J()})})(),[tf]:Q5((H)=>{let Z=matchMedia("(orientation: landscape)"),z=(X)=>H(X.matches?"landscape":"portrait");return Z.addEventListener("change",z),()=>Z.removeEventListener("change",z)},{value:matchMedia("(orientation: landscape)").matches?"landscape":"portrait"})}),[M([lf,cf,hf,tf])]});var EQ=R("card-mediaqueries",({expose:f,first:m,requestContext:M,watch:$})=>{let H=m(".motion"),Z=m(".theme"),z=m(".viewport"),X=m(".orientation");return f({motion:M(lf,"unknown"),theme:M(cf,"unknown"),viewport:M(hf,"unknown"),orientation:M(tf,"unknown")}),[H&&$("motion",w(H)),Z&&$("theme",w(Z)),z&&$("viewport",w(z)),X&&$("orientation",w(X))]});var dQ=R("form-checkbox",({expose:f,first:m,host:M,on:$,watch:H})=>{let Z=m('input[type="checkbox"]',"Add a native checkbox."),z=m(".label")??m("label");return f({checked:Z.checked,label:z?.textContent??""}),[$(Z,"change",()=>({checked:Z.checked})),H("checked",(X)=>{Z.checked=X,M.toggleAttribute("checked",X)}),z&&H("label",w(z))]});var _5=E5("p3"),r9=E5("rgb"),n9=new Intl.NumberFormat("en-US",{maximumFractionDigits:2}).format,l9=new Intl.NumberFormat("en-US",{maximumFractionDigits:4}).format,D2=20,a1=0.71,K0={l:1,c:0.4,h:360},s1={l:0.0025,c:0.001,h:1},c9={l:0.05,c:0.02,h:15},V2=(f,m)=>m?c9[f]:s1[f],lQ=R("form-colorgraph",({all:f,expose:m,first:M,host:$,on:H,watch:Z})=>{let z={l:M('input[name="lightness"]','Add an <input[name="lightness"]> element to control the lightness of the color.'),c:M('input[name="chroma"]','Add an <input[name="chroma"]> element to control the chroma of the color.'),h:M('input[name="hue"]','Add an <input[name="hue"]> element to control the hue of the color.')},X=M(".graph","Add a <.graph> element as a container for the color graph."),J=M(".graph canvas","Add a <canvas> element inside the graph to display the lightness/chroma graph."),B=M(".slider","Add a <.slider> element as a container for track and thumb."),W=M(".slider canvas","Add a <canvas> element inside the slider to display the hue slider track."),Q=M(".knob","Add a <.knob> element as a drag knob to control lightness and chroma."),j=M(".thumb","Add a <.thumb> element as a drag knob to control the hue."),b=f("input"),U=f(".error"),p=f("button.decrement"),N=f("button.increment");for(let[P,q]of Object.entries(z))q.min="0",q.max=P==="l"?"100":P==="c"?"0.4":"360",q.step="any";B.setAttribute("aria-valuemin","0"),B.setAttribute("aria-valuemax","360");let K=k(X.getBoundingClientRect().width),u=v(()=>K.get()-2*D2),o={l:k(""),c:k(""),h:k("")},D=(P,q)=>{let V=P==="l"?q*100:q;return P==="c"?l9(V):n9(V)},Y=(P)=>{if(P.closest(".lightness"))return"l";if(P.closest(".chroma"))return"c";if(P.closest(".hue"))return"h";return null},C=(P,q,V,_=1)=>p0({mode:"oklch",l:1-q,c:P*K0.c,h:V,alpha:_}),G=(P,q)=>{let V=K.get(),_=Math.round(q.c*V/K0.c),L=Math.round((1-q.l)*V);P.style.setProperty("background-color",p0(q)),P.style.setProperty("border-color",q.l>a1?"black":"white"),P.style.setProperty("left",`${_}px`),P.style.setProperty("top",`${L}px`)},I=(P)=>{let q={...$.color,h:P*K0.h};if(r9(q))return q;if(_5(q))q.alpha=0.5;else q.alpha=0;return q},r=(P)=>{i(()=>{$.color=P;for(let q of["l","c","h"])o[q].set("")})},d=(P)=>P==="l"?$.lightness:P==="c"?$.chroma:$.hue,l=(P,q)=>{let V=Math.round(q/s1[P])*s1[P];if(V<0||V>K0[P])return;let _={...$.color,[P]:V};if(_5(_))r(_);else z[P].setCustomValidity("Color out of gamut"),o[P].set(z[P].validationMessage)},J0=a0((P,q,V,_,L)=>{let F={...$.color,c:Math.min(Math.max((P-_)/L,0),1)*K0.c,l:1-Math.min(Math.max((q-V)/L,0),1)};if(_5(F))r(F)}),j0=a0((P,q,V)=>{let _={...$.color,h:Math.min(Math.max((P-q)/V,0),1)*K0.h};if(_5(_))r(_)});m({color:z5(),lightness:()=>$.color.l,chroma:()=>$.color.c,hue:()=>$.color.h??0,stepDown:s0((P,q=!1)=>{l(P,d(P)-V2(P,q))}),stepUp:s0((P,q=!1)=>{l(P,d(P)+V2(P,q))})});let m0=[Z(()=>X,()=>{let P=a0((V)=>{K.set(V)}),q=new ResizeObserver(()=>{P(X.clientWidth)});return q.observe(X),()=>{q.disconnect(),P.cancel()}}),Z(()=>p0($.color),e0($,"--color-base")),t(b,(P)=>{let q=Y(P);return[q&&Z(o[q],(V)=>{if(P.ariaInvalid=String(!!V),V&&P.id)P.setAttribute("aria-errormessage",`${P.id}-error`);else P.removeAttribute("aria-errormessage")}),Z("color",(V)=>{if(q)P.value=D(q,V[q]??0)}),H(P,"change",()=>{if(!q)return;let V=P.valueAsNumber,_={...$.color,[q]:q==="l"?V/100:V};if(_5(_))r(_);else P.setCustomValidity("Color out of gamut"),o[q].set(P.validationMessage)})]}),t(U,(P)=>{let q=Y(P);return[q?Z(o[q],w(P)):!1]}),H(X,"pointerdown",(P)=>{let{top:q,left:V}=J.getBoundingClientRect(),_=K.get();Q.ariaPressed="true",X.setPointerCapture(P.pointerId);let L=(y)=>{let O=(y.getCoalescedEvents?.()||[]).pop()||y;J0(O.clientX,O.clientY,q,V,_)},F=()=>{X.removeEventListener("pointermove",L),X.removeEventListener("pointerup",F),J0.cancel(),Q.ariaPressed="false"};X.addEventListener("pointermove",L,{passive:!0}),X.addEventListener("pointerup",F)}),Z(()=>`${K.get()}px`,e0(X,"--canvas-size")),Z(()=>({hue:$.hue,n:Math.round(K.get())}),({hue:P,n:q})=>{J.width=q,J.height=q;let V=J.getContext("2d",{colorSpace:"display-p3"});if(!V)return;let _=(y,O="rgb")=>h1({mode:"oklch",l:y,c:K0.c,h:P},"oklch",O).c/K0.c,L=(y,O,E,M0=1)=>[C(y,E,P,M0),C(O,E,P,M0)],F=(y,O,E="rgb")=>{let M0=_(1-O/q,E)*q,N5=V.createLinearGradient(y,0,M0,0),Wf=L(y/q,M0/q,O/q,E==="p3"?0.5:1);return N5.addColorStop(0,Wf[0]),N5.addColorStop(1,Wf[1]),V.fillStyle=N5,V.fillRect(y,O,M0-y,1),[M0,Wf[1]]};V.clearRect(0,0,q,q);for(let y=0;y<q;y++){let[O,E]=F(0,y);if(_5(E))F(O,y,"p3")}}),Z(()=>({l:$.lightness,c:$.chroma,size:K.get()}),({l:P,c:q,size:V})=>{Q.style.setProperty("top",`${Math.round((1-P)*V)}px`),Q.style.setProperty("left",`${Math.round(q*V/K0.c)}px`),Q.style.setProperty("--color-border",P>a1?"black":"white")}),H(B,"pointerdown",(P)=>{let q=W.getBoundingClientRect().left,V=u.get();j.ariaPressed="true",B.setPointerCapture(P.pointerId);let _=(F)=>{let y=(F.getCoalescedEvents?.()||[]).pop()||F;j0(y.clientX,q,V)},L=()=>{B.removeEventListener("pointermove",_),B.removeEventListener("pointerup",L),j0.cancel(),j.ariaPressed="false"};B.addEventListener("pointermove",_,{passive:!0}),B.addEventListener("pointerup",L)}),Z(()=>`${u.get()}px`,e0(B,"--track-width")),Z("hue",(P)=>{B.setAttribute("aria-valuenow",String(P)),B.setAttribute("aria-valuetext",`${D("h",P)}°`)}),Z(()=>({color:$.color,n:Math.round(u.get())}),({n:P})=>{W.width=P;let q=W.getContext("2d",{colorSpace:"display-p3"});if(!q)return;q.clearRect(0,0,P,1);for(let V=0;V<P;V++)q.fillStyle=p0(I(V/P)),q.fillRect(V,0,1,1)}),Z(()=>({hue:$.hue,l:$.lightness,tw:u.get()}),({hue:P,l:q,tw:V})=>{j.style.setProperty("left",`${Math.round(P*V/K0.h)+D2}px`),j.style.setProperty("--color-border",q>a1?"black":"white")}),t(p,(P)=>{let q=Y(P);return[H(P,"click",(V)=>{if(q)$.stepDown(q,V.shiftKey)}),Z("color",(V)=>{if(!q){P.disabled=!0;return}P.disabled=(V[q]??0)<=0})]}),t(N,(P)=>{let q=Y(P);return[H(P,"click",(V)=>{if(q)$.stepUp(q,V.shiftKey)}),Z("color",(V)=>{if(!q){P.disabled=!0;return}P.disabled=(V[q]??0)>=K0[q]})]}),H($,"keydown",(P)=>{let{key:q,shiftKey:V}=P,_=P.target;if(!_||_.localName==="input"&&(q==="ArrowLeft"||q==="ArrowRight"))return;if(q.substring(0,5)==="Arrow"||["+","-"].includes(q)){P.preventDefault(),P.stopPropagation();let L=Y(_);if(L){if(q==="ArrowLeft"||q==="ArrowDown"||q==="-")$.stepDown(L,V);else if(q==="ArrowRight"||q==="ArrowUp"||q==="+")$.stepUp(L,V)}else if(_.role==="slider"){if(q==="ArrowLeft"||q==="ArrowDown"||q==="-")$.stepDown("h",V);else if(q==="ArrowRight"||q==="ArrowUp"||q==="+")$.stepUp("h",V)}else switch(q){case"ArrowDown":$.stepDown("l",V);break;case"ArrowUp":$.stepUp("l",V);break;case"ArrowLeft":$.stepDown("c",V);break;case"ArrowRight":$.stepUp("c",V);break;case"-":$.stepDown("h");break;case"+":$.stepUp("h");break}}})];for(let P=1;P<5;P++){let q=M(`li.lighten${(5-P)*20}`);if(q)m0.push(Z(()=>({color:$.color,size:K.get()}),({color:V})=>{G(q,_0(V,1-P/10))}))}for(let P=1;P<5;P++){let q=M(`li.darken${P*20}`);if(q)m0.push(Z(()=>({color:$.color,size:K.get()}),({color:V})=>{G(q,_0(V,1-(P+5)/10))}))}return m0});var hQ=R("form-combobox",({expose:f,first:m,host:M,on:$,pass:H,watch:Z})=>{let z=m("input","Needed to enter value."),X=m("form-listbox","Needed to display options."),J=m("button.clear"),B=m("form-combobox > .error"),W=m(".description"),Q=B?.id,j=W?.id;if(j)z.setAttribute("aria-describedby",j);let b=k(!1),U=v(()=>b.get()&&X.options.length>0),p=k(z.value.length);return f({value:z.value,length:p.get,error:"",description:W?.textContent?.trim()??"",clear:s0(()=>{M.value="",z.value="",z.setCustomValidity(""),z.checkValidity(),z.dispatchEvent(new Event("input",{bubbles:!0})),z.dispatchEvent(new Event("change",{bubbles:!0})),z.focus()})}),[H(X,{filter:()=>M.value}),$(M,"keyup",({key:N})=>{if(N==="Escape")b.set(!1),z.focus();if(N==="Delete")M.clear()}),$(z,"input",()=>{p.set(z.value.length),z.checkValidity(),i(()=>{M.value=z.value,M.error=z.validationMessage??"",b.set(!0)})}),$(z,"keydown",({key:N,altKey:K})=>{if(N==="ArrowDown"){if(K)b.set(!0);if(U.get())X.options[0]?.focus()}}),$(X,"change",({target:N})=>{if(N instanceof HTMLInputElement)z.value=N.value,z.checkValidity(),i(()=>{M.value=N.value,M.error=z.validationMessage??"",b.set(!1),z.focus()})}),$(J,"click",()=>{M.clear()}),Z("value",q5(M,"value")),Z("error",(N)=>{if(z.ariaInvalid=String(!!N),N&&Q)z.setAttribute("aria-errormessage",Q);else z.removeAttribute("aria-errormessage")}),B&&Z("error",w(B)),W&&Z("description",w(W)),Z(U,(N)=>{X.hidden=!N,z.ariaExpanded=String(N)}),J&&Z(p,z0(J))]});var i9=0,sQ=R("form-inplace-edit",({expose:f,first:m,host:M,on:$,watch:H})=>{let Z=m(".text",'Add an element with "text" class for label display.'),z=m("button","Add a <button> element for edit mode toggle."),X=`form-inplace-edit-input${++i9}`,J=null;return f({editing:s5(),value:Z.textContent?.trim()??""}),[$(z,"click",(B)=>{if(B.stopPropagation(),M.editing&&J)return{editing:!1,value:J.value};else M.editing=!M.editing}),$(Z,"dblclick",()=>({editing:!0})),$(M,"keydown",(B)=>{if(!M.editing)return;if(B.key!=="Escape"&&B.key!=="Enter")return;if(B.preventDefault(),J&&B.key==="Enter")return{editing:!1,value:J.value};else M.editing=!1}),$(z,"mousedown",(B)=>{B.preventDefault()}),$(M,"focusout",(B)=>{if(!M.editing)return;let W=B.relatedTarget;if(W&&M.contains(W))return;M.editing=!1}),H("value",w(Z)),H("editing",(B)=>{if(M.toggleAttribute("editing",B),B){let W=document.createElement("form-textbox"),Q=document.createElement("label");Q.className="visually-hidden",Q.setAttribute("for",X),Q.textContent="Edit";let j=document.createElement("div");j.className="input",J=document.createElement("input"),J.type="text",J.id=X,J.value=M.value,j.append(J),W.append(Q,j),M.insertBefore(W,Z),z.textContent="✓",z.setAttribute("aria-label","Accept"),J.focus(),J.select()}else m("form-textbox")?.remove(),z.textContent="✎",z.setAttribute("aria-label","Edit")})]});var Y2=new Map,h9=(f)=>{let m=f.toLowerCase().split(",").map(($)=>$.trim()),M={noCache:!1,noStore:!1,maxAge:void 0};for(let $ of m)if($==="no-cache")M.noCache=!0;else if($==="no-store")M.noStore=!0;else if($.startsWith("max-age=")){let H=parseInt($.substring(8),10);if(!isNaN(H))M.maxAge=H}return M},t9=(f)=>{if(f.maxAge!==void 0)return(Date.now()-f.timestamp)/1000<f.maxAge;return!0},af=(f,m,M="src")=>!!f&&!!(m.parentElement||m.getRootNode().host)?.closest(`${m.localName}[${M}="${f}"]`),sf=(f)=>{if(!f)return!1;try{if(new URL(f,location.href).origin===location.origin)return!0}catch{return!1}return!1},ef=async(f,m,M=($)=>$.text())=>{let $=Y2.get(f),H={};if($?.etag)H["If-None-Match"]=$.etag;if($?.lastModified)H["If-Modified-Since"]=$.lastModified;let Z=await fetch(f,{signal:m??null,headers:H});if(Z.status===304&&$)return{content:$.content,fromCache:!0};if(!Z.ok)throw Error(`HTTP error: ${Z.statusText}`);let z=await M(Z),X=Z.headers.get("cache-control"),J=Z.headers.get("etag"),B=Z.headers.get("last-modified"),W=X?h9(X):{noCache:!1,noStore:!1,maxAge:void 0};if(!W.noStore){let Q={content:z,timestamp:Date.now(),etag:J||void 0,lastModified:B||void 0,maxAge:W.maxAge};if(!W.noCache||t9(Q))Y2.set(f,Q)}return{content:z,fromCache:!1}};function G2(f,m){if(!m)return V0(f);let M=typeof m==="string"?new RegExp(m.replace(/[.*+?^${}()|[\]\\]/g,"\\$&"),"giu"):m,$=[],H;M.lastIndex=0;while((H=M.exec(f))!==null)if($.push({start:H.index,end:H.index+H[0].length,text:H[0]}),H.index===M.lastIndex)M.lastIndex++;if($.length===0)return V0(f);let Z="",z=0;for(let{start:X,end:J,text:B}of $)Z+=V0(f.slice(z,X)),Z+=`<mark>${V0(B)}</mark>`,z=J;return Z+=V0(f.slice(z)),Z}var e1=(f,...m)=>String.raw({raw:f},...m);var a9="Enter",s9=["ArrowUp"],F2=["ArrowDown"],C2="Home",K2="End",e9=[...s9,...F2,C2,K2],XU=R("form-listbox",({all:f,expose:m,first:M,host:$,on:H,watch:Z})=>{let z=M('input[type="hidden"]',"Needed to store the selected value."),X=M("input.filter"),J=M("button.clear"),B=M("card-callout"),W=M(".loading"),Q=M(".error"),j=M('[role="listbox"]',"Needed to display list of options."),b=f('button[role="option"]'),U=(D)=>D.map((Y)=>e1`<button
+`||m[A]==="\t"||m[A]===" "))A++;continue}if($===",")return;if($===")"){M.push({type:x.ParenClose});continue}if($==="+"){if(A--,Of(m)){M.push(Mf(m));continue}return}if($==="-"){if(A--,Of(m)){M.push(Mf(m));continue}if(p1(m)){M.push({type:x.Ident,value:If(m)});continue}return}if($==="."){if(A--,Of(m)){M.push(Mf(m));continue}return}if($==="/"){while(A<m.length&&(m[A]===`
+`||m[A]==="\t"||m[A]===" "))A++;let H;if(Of(m)){if(H=Mf(m),H.type!==x.Hue){M.push({type:x.Alpha,value:H});continue}}if(p1(m)){if(If(m)==="none"){M.push({type:x.Alpha,value:{type:x.None,value:void 0}});continue}}return}if(/\d/.test($)){A--,M.push(Mf(m));continue}if(q1.test($)){A--,M.push(F8(m));continue}return}return M}function K8(f){f._i=0;let m=f[f._i++];if(!m||m.type!==x.Function||m.value!=="color")return;if(m=f[f._i++],m.type!==x.Ident)return;let M=P1[m.value];if(!M)return;let $={mode:M},H=E6(f,!1);if(!H)return;let Z=w0(M).channels;for(let z=0,X,J;z<Z.length;z++)if(X=H[z],J=Z[z],X.type!==x.None){if($[J]=X.type===x.Number?X.value:X.value/100,J==="alpha")$[J]=Math.max(0,Math.min(1,$[J]))}return $}function E6(f,m){let M=[],$;while(f._i<f.length){if($=f[f._i++],$.type===x.None||$.type===x.Number||$.type===x.Alpha||$.type===x.Percentage||m&&$.type===x.Hue){M.push($);continue}if($.type===x.ParenClose){if(f._i<f.length)return;continue}return}if(M.length<3||M.length>4)return;if(M.length===4){if(M[3].type!==x.Alpha)return;M[3]=M[3].value}if(M.length===3)M.push({type:x.None,value:void 0});return M.every((H)=>H.type!==x.Alpha)?M:void 0}function O8(f,m){f._i=0;let M=f[f._i++];if(!M||M.type!==x.Function)return;let $=E6(f,m);if(!$)return;return $.unshift(M.value),$}var A8=(f)=>{if(typeof f!=="string")return;let m=C8(f),M=m?O8(m,!0):void 0,$=void 0,H=0,Z=mf.length;while(H<Z)if(($=mf[H++](f,M))!==void 0)return $;return m?K8(m):void 0},T6=A8;function I8(f,m){if(!m||m[0]!=="rgb"&&m[0]!=="rgba")return;let M={mode:"rgb"},[,$,H,Z,z]=m;if($.type===x.Hue||H.type===x.Hue||Z.type===x.Hue)return;if($.type!==x.None)M.r=$.type===x.Number?$.value/255:$.value/100;if(H.type!==x.None)M.g=H.type===x.Number?H.value/255:H.value/100;if(Z.type!==x.None)M.b=Z.type===x.Number?Z.value/255:Z.value/100;if(z.type!==x.None)M.alpha=Math.min(1,Math.max(0,z.type===x.Number?z.value:z.value/100));return M}var _6=I8;var y8=(f)=>f==="transparent"?{mode:"rgb",r:0,g:0,b:0,alpha:0}:void 0,S6=y8;var d6=(f,m,M)=>f+M*(m-f);var w8=(f)=>{let m=[];for(let M=0;M<f.length-1;M++){let $=f[M],H=f[M+1];if($===void 0&&H===void 0)m.push(void 0);else if($!==void 0&&H!==void 0)m.push([$,H]);else m.push($!==void 0?[$,$]:[H,H])}return m},k6=(f)=>(m)=>{let M=w8(m);return($)=>{let H=$*M.length,Z=$>=1?M.length-1:Math.max(Math.floor(H),0),z=M[Z];return z===void 0?void 0:f(z[0],z[1],H-Z)}};var T=k6(d6);var B0=(f)=>{let m=!1,M=f.map(($)=>{if($!==void 0)return m=!0,$;return 1});return m?M:f};var u8={mode:"rgb",channels:["r","g","b","alpha"],parse:[_6,y6,u6,I6,S6,"srgb"],serialize:"srgb",interpolate:{r:T,g:T,b:T,alpha:{use:T,fixup:B0}},gamut:!0,white:{r:1,g:1,b:1},black:{r:0,g:0,b:0}},Y0=u8;var x1=(f=0)=>Math.pow(Math.abs(f),2.19921875)*Math.sign(f),T8=(f)=>{let m=x1(f.r),M=x1(f.g),$=x1(f.b),H={mode:"xyz65",x:0.5766690429101305*m+0.1855582379065463*M+0.1882286462349947*$,y:0.297344975250536*m+0.6273635662554661*M+0.0752914584939979*$,z:0.0270313613864123*m+0.0706888525358272*M+0.9913375368376386*$};if(f.alpha!==void 0)H.alpha=f.alpha;return H},N1=T8;var L1=(f)=>Math.pow(Math.abs(f),0.4547069271758437)*Math.sign(f),g8=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H={mode:"a98",r:L1(f*2.0415879038107465-m*0.5650069742788597-0.3447313507783297*M),g:L1(f*-0.9692436362808798+m*1.8759675015077206+0.0415550574071756*M),b:L1(f*0.0134442806320312-m*0.1183623922310184+1.0151749943912058*M)};if($!==void 0)H.alpha=$;return H},R1=g8;var D1=(f=0)=>{let m=Math.abs(f);if(m<=0.04045)return f/12.92;return(Math.sign(f)||1)*Math.pow((m+0.055)/1.055,2.4)},E8=({r:f,g:m,b:M,alpha:$})=>{let H={mode:"lrgb",r:D1(f),g:D1(m),b:D1(M)};if($!==void 0)H.alpha=$;return H},u0=E8;var _8=(f)=>{let{r:m,g:M,b:$,alpha:H}=u0(f),Z={mode:"xyz65",x:0.4123907992659593*m+0.357584339383878*M+0.1804807884018343*$,y:0.2126390058715102*m+0.715168678767756*M+0.0721923153607337*$,z:0.0193308187155918*m+0.119194779794626*M+0.9505321522496607*$};if(H!==void 0)Z.alpha=H;return Z},T0=_8;var V1=(f=0)=>{let m=Math.abs(f);if(m>0.0031308)return(Math.sign(f)||1)*(1.055*Math.pow(m,0.4166666666666667)-0.055);return f*12.92},S8=({r:f,g:m,b:M,alpha:$},H="rgb")=>{let Z={mode:H,r:V1(f),g:V1(m),b:V1(M)};if($!==void 0)Z.alpha=$;return Z},g0=S8;var d8=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=g0({r:f*3.2409699419045226-m*1.537383177570094-0.4986107602930034*M,g:f*-0.9692436362808796+m*1.8759675015077204+0.0415550574071756*M,b:f*0.0556300796969936-m*0.2039769588889765+1.0569715142428784*M});if($!==void 0)H.alpha=$;return H},E0=d8;var k8={...Y0,mode:"a98",parse:["a98-rgb"],serialize:"a98-rgb",fromMode:{rgb:(f)=>R1(T0(f)),xyz65:R1},toMode:{rgb:(f)=>E0(N1(f)),xyz65:N1}},v6=k8;var v8=(f)=>(f=f%360)<0?f+360:f,X0=v8;function Y1({h:f,s:m,l:M,alpha:$}){if(f=X0(f!==void 0?f:0),m===void 0)m=0;if(M===void 0)M=0;let H=M+m*(M<0.5?M:1-M),Z=H-(H-M)*2*Math.abs(f/60%2-1),z;switch(Math.floor(f/60)){case 0:z={r:H,g:Z,b:2*M-H};break;case 1:z={r:Z,g:H,b:2*M-H};break;case 2:z={r:2*M-H,g:H,b:Z};break;case 3:z={r:2*M-H,g:Z,b:H};break;case 4:z={r:Z,g:2*M-H,b:H};break;case 5:z={r:H,g:2*M-H,b:Z};break;default:z={r:2*M-H,g:2*M-H,b:2*M-H}}if(z.mode="rgb",$!==void 0)z.alpha=$;return z}function G1({r:f,g:m,b:M,alpha:$}){if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=Math.max(f,m,M),Z=Math.min(f,m,M),z={mode:"hsl",s:H===Z?0:(H-Z)/(1-Math.abs(H+Z-1)),l:0.5*(H+Z)};if(H-Z!==0)z.h=(H===f?(m-M)/(H-Z)+(m<M)*6:H===m?(M-f)/(H-Z)+2:(f-m)/(H-Z)+4)*60;if($!==void 0)z.alpha=$;return z}var o8=(f,m)=>{switch(m){case"deg":return+f;case"rad":return f/Math.PI*180;case"grad":return f/10*9;case"turn":return f*360}},o6=o8;var r8=new RegExp(`^hsla?\\(\\s*${w6}${f5}${x5}${f5}${x5}\\s*(?:,\\s*${ff}\\s*)?\\)$`),n8=(f)=>{let m=f.match(r8);if(!m)return;let M={mode:"hsl"};if(m[3]!==void 0)M.h=+m[3];else if(m[1]!==void 0&&m[2]!==void 0)M.h=o6(m[1],m[2]);if(m[4]!==void 0)M.s=Math.min(Math.max(0,m[4]/100),1);if(m[5]!==void 0)M.l=Math.min(Math.max(0,m[5]/100),1);if(m[6]!==void 0)M.alpha=Math.max(0,Math.min(1,m[6]/100));else if(m[7]!==void 0)M.alpha=Math.max(0,Math.min(1,+m[7]));return M},r6=n8;function l8(f,m){if(!m||m[0]!=="hsl"&&m[0]!=="hsla")return;let M={mode:"hsl"},[,$,H,Z,z]=m;if($.type!==x.None){if($.type===x.Percentage)return;M.h=$.value}if(H.type!==x.None){if(H.type===x.Hue)return;M.s=H.value/100}if(Z.type!==x.None){if(Z.type===x.Hue)return;M.l=Z.value/100}if(z.type!==x.None)M.alpha=Math.min(1,Math.max(0,z.type===x.Number?z.value:z.value/100));return M}var n6=l8;var c8=(f,m)=>{return f.map((M,$,H)=>{if(M===void 0)return M;let Z=X0(M);if($===0||f[$-1]===void 0)return Z;return m(Z-X0(H[$-1]))}).reduce((M,$)=>{if(!M.length||$===void 0||M[M.length-1]===void 0)return M.push($),M;return M.push($+M[M.length-1]),M},[])},M5=(f)=>c8(f,(m)=>Math.abs(m)<=180?m:m-360*Math.sign(m));var $f=(f,m)=>{if(f.h===void 0||m.h===void 0||!f.s||!m.s)return 0;let M=X0(f.h),$=X0(m.h),H=Math.sin(($-M+360)/2*Math.PI/180);return 2*Math.sqrt(f.s*m.s)*H},F1=(f,m)=>{if(f.h===void 0||m.h===void 0)return 0;let M=X0(f.h),$=X0(m.h);if(Math.abs($-M)>180)return M-($-360*Math.sign($-M));return $-M},C1=(f,m)=>{if(f.h===void 0||m.h===void 0||!f.c||!m.c)return 0;let M=X0(f.h),$=X0(m.h),H=Math.sin(($-M+360)/2*Math.PI/180);return 2*Math.sqrt(f.c*m.c)*H},K1=(f="rgb",m=[1,1,1,0])=>{let M=w0(f),$=M.channels,H=M.difference,Z=e(f);return(z,X)=>{let J=Z(z),B=Z(X);return Math.sqrt($.reduce((W,Q,j)=>{let b=H[Q]?H[Q](J,B):J[Q]-B[Q];return W+(m[j]||0)*Math.pow(isNaN(b)?0:b,2)},0))}};var O1=(f=1,m=1,M=1)=>{let $=e("lab65");return(H,Z)=>{let z=$(H),X=$(Z),J=z.l,B=z.a,W=z.b,Q=Math.sqrt(B*B+W*W),j=X.l,b=X.a,U=X.b,p=Math.sqrt(b*b+U*U),N=(Q+p)/2,K=0.5*(1-Math.sqrt(Math.pow(N,7)/(Math.pow(N,7)+Math.pow(25,7)))),u=B*(1+K),o=b*(1+K),D=Math.sqrt(u*u+W*W),Y=Math.sqrt(o*o+U*U),C=Math.abs(u)+Math.abs(W)===0?0:Math.atan2(W,u);C+=(C<0)*2*Math.PI;let G=Math.abs(o)+Math.abs(U)===0?0:Math.atan2(U,o);G+=(G<0)*2*Math.PI;let I=j-J,r=Y-D,d=D*Y===0?0:G-C;d-=(d>Math.PI)*2*Math.PI,d+=(d<-Math.PI)*2*Math.PI;let l=2*Math.sqrt(D*Y)*Math.sin(d/2),J0=(J+j)/2,j0=(D+Y)/2,m0;if(D*Y===0)m0=C+G;else m0=(C+G)/2,m0-=(Math.abs(C-G)>Math.PI)*Math.PI,m0+=(m0<0)*2*Math.PI;let P=Math.pow(J0-50,2),q=1-0.17*Math.cos(m0-Math.PI/6)+0.24*Math.cos(2*m0)+0.32*Math.cos(3*m0+Math.PI/30)-0.2*Math.cos(4*m0-63*Math.PI/180),V=1+0.015*P/Math.sqrt(20+P),_=1+0.045*j0,L=1+0.015*j0*q,F=30*Math.PI/180*Math.exp(-1*Math.pow((180/Math.PI*m0-275)/25,2)),y=2*Math.sqrt(Math.pow(j0,7)/(Math.pow(j0,7)+Math.pow(25,7))),O=-1*Math.sin(2*F)*y;return Math.sqrt(Math.pow(I/(f*V),2)+Math.pow(r/(m*_),2)+Math.pow(l/(M*L),2)+O*r/(m*_)*l/(M*L))}};var $5=(f)=>{let m=f.reduce(($,H)=>{if(H!==void 0){let Z=H*Math.PI/180;$.sin+=Math.sin(Z),$.cos+=Math.cos(Z)}return $},{sin:0,cos:0}),M=Math.atan2(m.sin,m.cos)*180/Math.PI;return M<0?360+M:M};var i8={mode:"hsl",toMode:{rgb:Y1},fromMode:{rgb:G1},channels:["h","s","l","alpha"],ranges:{h:[0,360]},gamut:"rgb",parse:[n6,r6],serialize:(f)=>`hsl(${f.h!==void 0?f.h:"none"} ${f.s!==void 0?f.s*100+"%":"none"} ${f.l!==void 0?f.l*100+"%":"none"}${f.alpha<1?` / ${f.alpha}`:""})`,interpolate:{h:{use:T,fixup:M5},s:T,l:T,alpha:{use:T,fixup:B0}},difference:{h:$f},average:{h:$5}},l6=i8;function Hf({h:f,s:m,v:M,alpha:$}){if(f=X0(f!==void 0?f:0),m===void 0)m=0;if(M===void 0)M=0;let H=Math.abs(f/60%2-1),Z;switch(Math.floor(f/60)){case 0:Z={r:M,g:M*(1-m*H),b:M*(1-m)};break;case 1:Z={r:M*(1-m*H),g:M,b:M*(1-m)};break;case 2:Z={r:M*(1-m),g:M,b:M*(1-m*H)};break;case 3:Z={r:M*(1-m),g:M*(1-m*H),b:M};break;case 4:Z={r:M*(1-m*H),g:M*(1-m),b:M};break;case 5:Z={r:M,g:M*(1-m),b:M*(1-m*H)};break;default:Z={r:M*(1-m),g:M*(1-m),b:M*(1-m)}}if(Z.mode="rgb",$!==void 0)Z.alpha=$;return Z}function Zf({r:f,g:m,b:M,alpha:$}){if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=Math.max(f,m,M),Z=Math.min(f,m,M),z={mode:"hsv",s:H===0?0:1-Z/H,v:H};if(H-Z!==0)z.h=(H===f?(m-M)/(H-Z)+(m<M)*6:H===m?(M-f)/(H-Z)+2:(f-m)/(H-Z)+4)*60;if($!==void 0)z.alpha=$;return z}var h8={mode:"hsv",toMode:{rgb:Hf},parse:["--hsv"],serialize:"--hsv",fromMode:{rgb:Zf},channels:["h","s","v","alpha"],ranges:{h:[0,360]},gamut:"rgb",interpolate:{h:{use:T,fixup:M5},s:T,v:T,alpha:{use:T,fixup:B0}},difference:{h:$f},average:{h:$5}},c6=h8;function A1({h:f,w:m,b:M,alpha:$}){if(m===void 0)m=0;if(M===void 0)M=0;if(m+M>1){let H=m+M;m/=H,M/=H}return Hf({h:f,s:M===1?1:1-m/(1-M),v:1-M,alpha:$})}function I1(f){let m=Zf(f);if(m===void 0)return;let M=m.s!==void 0?m.s:0,$=m.v!==void 0?m.v:0,H={mode:"hwb",w:(1-M)*$,b:1-$};if(m.h!==void 0)H.h=m.h;if(m.alpha!==void 0)H.alpha=m.alpha;return H}function t8(f,m){if(!m||m[0]!=="hwb")return;let M={mode:"hwb"},[,$,H,Z,z]=m;if($.type!==x.None){if($.type===x.Percentage)return;M.h=$.value}if(H.type!==x.None){if(H.type===x.Hue)return;M.w=H.value/100}if(Z.type!==x.None){if(Z.type===x.Hue)return;M.b=Z.value/100}if(z.type!==x.None)M.alpha=Math.min(1,Math.max(0,z.type===x.Number?z.value:z.value/100));return M}var i6=t8;var a8={mode:"hwb",toMode:{rgb:A1},fromMode:{rgb:I1},channels:["h","w","b","alpha"],ranges:{h:[0,360]},gamut:"rgb",parse:[i6],serialize:(f)=>`hwb(${f.h!==void 0?f.h:"none"} ${f.w!==void 0?f.w*100+"%":"none"} ${f.b!==void 0?f.b*100+"%":"none"}${f.alpha<1?` / ${f.alpha}`:""})`,interpolate:{h:{use:T,fixup:M5},w:T,b:T,alpha:{use:T,fixup:B0}},difference:{h:F1},average:{h:$5}},h6=a8;var yf=Math.pow(29,3)/Math.pow(3,3),wf=Math.pow(6,3)/Math.pow(29,3);var H5={X:0.9642956764295677,Y:1,Z:0.8251046025104602},Z5={X:0.9504559270516716,Y:1,Z:1.0890577507598784},X$=Math.pow(29,3)/Math.pow(3,3),J$=Math.pow(6,3)/Math.pow(29,3);var y1=(f)=>Math.pow(f,3)>wf?Math.pow(f,3):(116*f-16)/yf,s8=({l:f,a:m,b:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=(f+16)/116,Z=m/500+H,z=H-M/200,X={mode:"xyz50",x:y1(Z)*H5.X,y:y1(H)*H5.Y,z:y1(z)*H5.Z};if($!==void 0)X.alpha=$;return X},O5=s8;var e8=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=g0({r:f*3.1341359569958707-m*1.6173863321612538-0.4906619460083532*M,g:f*-0.978795502912089+m*1.916254567259524+0.03344273116131949*M,b:f*0.07195537988411677-m*0.2289768264158322+1.405386058324125*M});if($!==void 0)H.alpha=$;return H},A5=e8;var f9=(f)=>A5(O5(f)),uf=f9;var m9=(f)=>{let{r:m,g:M,b:$,alpha:H}=u0(f),Z={mode:"xyz50",x:0.436065742824811*m+0.3851514688337912*M+0.14307845442264197*$,y:0.22249319175623702*m+0.7168870538238823*M+0.06061979053616537*$,z:0.013923904500943465*m+0.09708128566574634*M+0.7140993584005155*$};if(H!==void 0)Z.alpha=H;return Z},I5=m9;var w1=(f)=>f>wf?Math.cbrt(f):(yf*f+16)/116,M9=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=w1(f/H5.X),Z=w1(m/H5.Y),z=w1(M/H5.Z),X={mode:"lab",l:116*Z-16,a:500*(H-Z),b:200*(Z-z)};if($!==void 0)X.alpha=$;return X},y5=M9;var $9=(f)=>{let m=y5(I5(f));if(f.r===f.b&&f.b===f.g)m.a=m.b=0;return m},Tf=$9;function H9(f,m){if(!m||m[0]!=="lab")return;let M={mode:"lab"},[,$,H,Z,z]=m;if($.type===x.Hue||H.type===x.Hue||Z.type===x.Hue)return;if($.type!==x.None)M.l=Math.min(Math.max(0,$.value),100);if(H.type!==x.None)M.a=H.type===x.Number?H.value:H.value*125/100;if(Z.type!==x.None)M.b=Z.type===x.Number?Z.value:Z.value*125/100;if(z.type!==x.None)M.alpha=Math.min(1,Math.max(0,z.type===x.Number?z.value:z.value/100));return M}var t6=H9;var Z9={mode:"lab",toMode:{xyz50:O5,rgb:uf},fromMode:{xyz50:y5,rgb:Tf},channels:["l","a","b","alpha"],ranges:{l:[0,100],a:[-125,125],b:[-125,125]},parse:[t6],serialize:(f)=>`lab(${f.l!==void 0?f.l:"none"} ${f.a!==void 0?f.a:"none"} ${f.b!==void 0?f.b:"none"}${f.alpha<1?` / ${f.alpha}`:""})`,interpolate:{l:T,a:T,b:T,alpha:{use:T,fixup:B0}}},w5=Z9;var gf=Math.pow(29,3)/Math.pow(3,3),Ef=Math.pow(6,3)/Math.pow(29,3);var u1=(f)=>Math.pow(f,3)>Ef?Math.pow(f,3):(116*f-16)/gf,z9=({l:f,a:m,b:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=(f+16)/116,Z=m/500+H,z=H-M/200,X={mode:"xyz65",x:u1(Z)*Z5.X,y:u1(H)*Z5.Y,z:u1(z)*Z5.Z};if($!==void 0)X.alpha=$;return X},_f=z9;var B9=(f)=>E0(_f(f)),Sf=B9;var T1=(f)=>f>Ef?Math.cbrt(f):(gf*f+16)/116,X9=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=T1(f/Z5.X),Z=T1(m/Z5.Y),z=T1(M/Z5.Z),X={mode:"lab65",l:116*Z-16,a:500*(H-Z),b:200*(Z-z)};if($!==void 0)X.alpha=$;return X},df=X9;var J9=(f)=>{let m=df(T0(f));if(f.r===f.b&&f.b===f.g)m.a=m.b=0;return m},kf=J9;var W9={...w5,mode:"lab65",parse:["--lab-d65"],serialize:"--lab-d65",toMode:{xyz65:_f,rgb:Sf},fromMode:{xyz65:df,rgb:kf},ranges:{l:[0,100],a:[-125,125],b:[-125,125]}},a6=W9;var b9=({l:f,a:m,b:M,alpha:$},H="lch")=>{if(m===void 0)m=0;if(M===void 0)M=0;let Z=Math.sqrt(m*m+M*M),z={mode:H,l:f,c:Z};if(Z)z.h=X0(Math.atan2(M,m)*180/Math.PI);if($!==void 0)z.alpha=$;return z},l0=b9;var Q9=({l:f,c:m,h:M,alpha:$},H="lab")=>{if(M===void 0)M=0;let Z={mode:H,l:f,a:m?m*Math.cos(M/180*Math.PI):0,b:m?m*Math.sin(M/180*Math.PI):0};if($!==void 0)Z.alpha=$;return Z},c0=Q9;function U9(f,m){if(!m||m[0]!=="lch")return;let M={mode:"lch"},[,$,H,Z,z]=m;if($.type!==x.None){if($.type===x.Hue)return;M.l=Math.min(Math.max(0,$.value),100)}if(H.type!==x.None)M.c=Math.max(0,H.type===x.Number?H.value:H.value*150/100);if(Z.type!==x.None){if(Z.type===x.Percentage)return;M.h=Z.value}if(z.type!==x.None)M.alpha=Math.min(1,Math.max(0,z.type===x.Number?z.value:z.value/100));return M}var s6=U9;var j9={mode:"lch",toMode:{lab:c0,rgb:(f)=>uf(c0(f))},fromMode:{rgb:(f)=>l0(Tf(f)),lab:l0},channels:["l","c","h","alpha"],ranges:{l:[0,100],c:[0,150],h:[0,360]},parse:[s6],serialize:(f)=>`lch(${f.l!==void 0?f.l:"none"} ${f.c!==void 0?f.c:"none"} ${f.h!==void 0?f.h:"none"}${f.alpha<1?` / ${f.alpha}`:""})`,interpolate:{h:{use:T,fixup:M5},c:T,l:T,alpha:{use:T,fixup:B0}},difference:{h:C1},average:{h:$5}},u5=j9;var P9={...u5,mode:"lch65",parse:["--lch-d65"],serialize:"--lch-d65",toMode:{lab65:(f)=>c0(f,"lab65"),rgb:(f)=>Sf(c0(f,"lab65"))},fromMode:{rgb:(f)=>l0(kf(f),"lch65"),lab65:(f)=>l0(f,"lch65")},ranges:{l:[0,100],c:[0,150],h:[0,360]}},e6=P9;var q9={...Y0,mode:"lrgb",toMode:{rgb:g0},fromMode:{rgb:u0},parse:["srgb-linear"],serialize:"srgb-linear"},f2=q9;var p9=({l:f,a:m,b:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=Math.pow(f+0.3963377773761749*m+0.2158037573099136*M,3),Z=Math.pow(f-0.1055613458156586*m-0.0638541728258133*M,3),z=Math.pow(f-0.0894841775298119*m-1.2914855480194092*M,3),X={mode:"lrgb",r:4.076741636075957*H-3.3077115392580616*Z+0.2309699031821044*z,g:-1.2684379732850317*H+2.6097573492876887*Z-0.3413193760026573*z,b:-0.0041960761386756*H-0.7034186179359362*Z+1.7076146940746117*z};if($!==void 0)X.alpha=$;return X},vf=p9;var x9=({r:f,g:m,b:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=Math.cbrt(0.412221469470763*f+0.5363325372617348*m+0.0514459932675022*M),Z=Math.cbrt(0.2119034958178252*f+0.6806995506452344*m+0.1073969535369406*M),z=Math.cbrt(0.0883024591900564*f+0.2817188391361215*m+0.6299787016738222*M),X={mode:"oklab",l:0.210454268309314*H+0.7936177747023054*Z-0.0040720430116193*z,a:1.9779985324311684*H-2.42859224204858*Z+0.450593709617411*z,b:0.0259040424655478*H+0.7827717124575296*Z-0.8086757549230774*z};if($!==void 0)X.alpha=$;return X},of=x9;var N9=(f)=>{let m=of(u0(f));if(f.r===f.b&&f.b===f.g)m.a=m.b=0;return m},rf=N9;var L9=(f)=>g0(vf(f)),nf=L9;function R9(f,m){if(!m||m[0]!=="oklab")return;let M={mode:"oklab"},[,$,H,Z,z]=m;if($.type===x.Hue||H.type===x.Hue||Z.type===x.Hue)return;if($.type!==x.None)M.l=Math.min(Math.max(0,$.type===x.Number?$.value:$.value/100),1);if(H.type!==x.None)M.a=H.type===x.Number?H.value:H.value*0.4/100;if(Z.type!==x.None)M.b=Z.type===x.Number?Z.value:Z.value*0.4/100;if(z.type!==x.None)M.alpha=Math.min(1,Math.max(0,z.type===x.Number?z.value:z.value/100));return M}var m2=R9;var D9={...w5,mode:"oklab",toMode:{lrgb:vf,rgb:nf},fromMode:{lrgb:of,rgb:rf},ranges:{l:[0,1],a:[-0.4,0.4],b:[-0.4,0.4]},parse:[m2],serialize:(f)=>`oklab(${f.l!==void 0?f.l:"none"} ${f.a!==void 0?f.a:"none"} ${f.b!==void 0?f.b:"none"}${f.alpha<1?` / ${f.alpha}`:""})`},M2=D9;function V9(f,m){if(!m||m[0]!=="oklch")return;let M={mode:"oklch"},[,$,H,Z,z]=m;if($.type!==x.None){if($.type===x.Hue)return;M.l=Math.min(Math.max(0,$.type===x.Number?$.value:$.value/100),1)}if(H.type!==x.None)M.c=Math.max(0,H.type===x.Number?H.value:H.value*0.4/100);if(Z.type!==x.None){if(Z.type===x.Percentage)return;M.h=Z.value}if(z.type!==x.None)M.alpha=Math.min(1,Math.max(0,z.type===x.Number?z.value:z.value/100));return M}var $2=V9;var Y9={...u5,mode:"oklch",toMode:{oklab:(f)=>c0(f,"oklab"),rgb:(f)=>nf(c0(f,"oklab"))},fromMode:{rgb:(f)=>l0(rf(f),"oklch"),oklab:(f)=>l0(f,"oklch")},parse:[$2],serialize:(f)=>`oklch(${f.l!==void 0?f.l:"none"} ${f.c!==void 0?f.c:"none"} ${f.h!==void 0?f.h:"none"}${f.alpha<1?` / ${f.alpha}`:""})`,ranges:{l:[0,1],c:[0,0.4],h:[0,360]}},H2=Y9;var G9=(f)=>{let{r:m,g:M,b:$,alpha:H}=u0(f),Z={mode:"xyz65",x:0.486570948648216*m+0.265667693169093*M+0.1982172852343625*$,y:0.2289745640697487*m+0.6917385218365062*M+0.079286914093745*$,z:0*m+0.0451133818589026*M+1.043944368900976*$};if(H!==void 0)Z.alpha=H;return Z},g1=G9;var F9=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H=g0({r:f*2.4934969119414263-m*0.9313836179191242-0.402710784450717*M,g:f*-0.8294889695615749+m*1.7626640603183465+0.0236246858419436*M,b:f*0.0358458302437845-m*0.0761723892680418+0.9568845240076871*M},"p3");if($!==void 0)H.alpha=$;return H},E1=F9;var C9={...Y0,mode:"p3",parse:["display-p3"],serialize:"display-p3",fromMode:{rgb:(f)=>E1(T0(f)),xyz65:E1},toMode:{rgb:(f)=>E0(g1(f)),xyz65:g1}},Z2=C9;var _1=(f)=>{let m=Math.abs(f);if(m>=0.001953125)return Math.sign(f)*Math.pow(m,0.5555555555555556);return 16*f},K9=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H={mode:"prophoto",r:_1(f*1.3457868816471585-m*0.2555720873797946-0.0511018649755453*M),g:_1(f*-0.5446307051249019+m*1.5082477428451466+0.0205274474364214*M),b:_1(f*0+m*0+1.2119675456389452*M)};if($!==void 0)H.alpha=$;return H},S1=K9;var d1=(f=0)=>{let m=Math.abs(f);if(m>=0.03125)return Math.sign(f)*Math.pow(m,1.8);return f/16},O9=(f)=>{let m=d1(f.r),M=d1(f.g),$=d1(f.b),H={mode:"xyz50",x:0.7977666449006423*m+0.1351812974005331*M+0.0313477341283922*$,y:0.2880748288194013*m+0.7118352342418731*M+0.0000899369387256*$,z:0*m+0*M+0.8251046025104602*$};if(f.alpha!==void 0)H.alpha=f.alpha;return H},k1=O9;var A9={...Y0,mode:"prophoto",parse:["prophoto-rgb"],serialize:"prophoto-rgb",fromMode:{xyz50:S1,rgb:(f)=>S1(I5(f))},toMode:{xyz50:k1,rgb:(f)=>A5(k1(f))}},z2=A9;var v1=(f)=>{let m=Math.abs(f);if(m>0.018053968510807)return(Math.sign(f)||1)*(1.09929682680944*Math.pow(m,0.45)-0.09929682680944008);return 4.5*f},I9=({x:f,y:m,z:M,alpha:$})=>{if(f===void 0)f=0;if(m===void 0)m=0;if(M===void 0)M=0;let H={mode:"rec2020",r:v1(f*1.7166511879712683-m*0.3556707837763925-0.2533662813736599*M),g:v1(f*-0.6666843518324893+m*1.6164812366349395+0.0157685458139111*M),b:v1(f*0.0176398574453108-m*0.0427706132578085+0.9421031212354739*M)};if($!==void 0)H.alpha=$;return H},o1=I9;var r1=(f=0)=>{let m=Math.abs(f);if(m<0.08124285829863151)return f/4.5;return(Math.sign(f)||1)*Math.pow((m+1.09929682680944-1)/1.09929682680944,2.2222222222222223)},y9=(f)=>{let m=r1(f.r),M=r1(f.g),$=r1(f.b),H={mode:"xyz65",x:0.6369580483012911*m+0.1446169035862083*M+0.1688809751641721*$,y:0.262700212011267*m+0.6779980715188708*M+0.059301716469862*$,z:0*m+0.0280726930490874*M+1.0609850577107909*$};if(f.alpha!==void 0)H.alpha=f.alpha;return H},n1=y9;var w9={...Y0,mode:"rec2020",fromMode:{xyz65:o1,rgb:(f)=>o1(T0(f))},toMode:{xyz65:n1,rgb:(f)=>E0(n1(f))},parse:["rec2020"],serialize:"rec2020"},B2=w9;var u9={mode:"xyz50",parse:["xyz-d50"],serialize:"xyz-d50",toMode:{rgb:A5,lab:y5},fromMode:{rgb:I5,lab:O5},channels:["x","y","z","alpha"],ranges:{x:[0,0.964],y:[0,0.999],z:[0,0.825]},interpolate:{x:T,y:T,z:T,alpha:{use:T,fixup:B0}}},X2=u9;var T9=(f)=>{let{x:m,y:M,z:$,alpha:H}=f;if(m===void 0)m=0;if(M===void 0)M=0;if($===void 0)$=0;let Z={mode:"xyz50",x:1.0479298208405488*m+0.0229467933410191*M-0.0501922295431356*$,y:0.0296278156881593*m+0.990434484573249*M-0.0170738250293851*$,z:-0.0092430581525912*m+0.0150551448965779*M+0.7518742899580008*$};if(H!==void 0)Z.alpha=H;return Z},J2=T9;var g9=(f)=>{let{x:m,y:M,z:$,alpha:H}=f;if(m===void 0)m=0;if(M===void 0)M=0;if($===void 0)$=0;let Z={mode:"xyz65",x:0.9554734527042182*m-0.0230985368742614*M+0.0632593086610217*$,y:-0.0283697069632081*m+1.0099954580058226*M+0.021041398966943*$,z:0.0123140016883199*m-0.0205076964334779*M+1.3303659366080753*$};if(H!==void 0)Z.alpha=H;return Z},W2=g9;var E9={mode:"xyz65",toMode:{rgb:E0,xyz50:J2},fromMode:{rgb:T0,xyz50:W2},ranges:{x:[0,0.95],y:[0,1],z:[0,1.088]},channels:["x","y","z","alpha"],parse:["xyz","xyz-d65"],serialize:"xyz-d65",interpolate:{x:T,y:T,z:T,alpha:{use:T,fixup:B0}}},b2=E9;var zz=f0(v6),Bz=f0(l6),Xz=f0(c6),Jz=f0(h6),Wz=f0(w5),bz=f0(a6),Qz=f0(u5),Uz=f0(e6),jz=f0(f2),Pz=f0(M2),qz=f0(H2),pz=f0(Z2),xz=f0(z2),Nz=f0(B2),Lz=f0(Y0),Rz=f0(X2),Dz=f0(b2);var _9=(f,m)=>Math.round(f*(m=Math.pow(10,m)))/m,S9=(f=4)=>(m)=>typeof m==="number"?_9(m,f):m,Q2=S9;var zf=Q2(2),Bf=(f)=>Math.max(0,Math.min(1,f||0)),T5=(f)=>Math.round(Bf(f)*255),U2=e("rgb"),d9=e("hsl"),j2=(f)=>{if(f===void 0)return;let m=T5(f.r),M=T5(f.g),$=T5(f.b);return"#"+(16777216|m<<16|M<<8|$).toString(16).slice(1)};var P2=(f)=>{if(f===void 0)return;let m=T5(f.r),M=T5(f.g),$=T5(f.b);if(f.alpha===void 0||f.alpha===1)return`rgb(${m}, ${M}, ${$})`;else return`rgba(${m}, ${M}, ${$}, ${zf(Bf(f.alpha))})`},q2=(f)=>{if(f===void 0)return;let m=zf(f.h||0),M=zf(Bf(f.s)*100)+"%",$=zf(Bf(f.l)*100)+"%";if(f.alpha===void 0||f.alpha===1)return`hsl(${m}, ${M}, ${$})`;else return`hsla(${m}, ${M}, ${$}, ${zf(Bf(f.alpha))})`},p0=(f)=>{let m=m5(f);if(!m)return;let M=w0(m.mode);if(!M.serialize||typeof M.serialize==="string"){let $=`color(${M.serialize||`--${m.mode}`} `;if(M.channels.forEach((H,Z)=>{if(H!=="alpha")$+=(Z?" ":"")+(m[H]!==void 0?m[H]:"none")}),m.alpha!==void 0&&m.alpha<1)$+=` / ${m.alpha}`;return $+")"}if(typeof M.serialize==="function")return M.serialize(m);return},Xf=(f)=>j2(U2(f));var l1=(f)=>P2(U2(f)),c1=(f)=>q2(d9(f));var p2=e("rgb"),x2=(f)=>{let m={mode:f.mode,r:Math.max(0,Math.min(f.r!==void 0?f.r:0,1)),g:Math.max(0,Math.min(f.g!==void 0?f.g:0,1)),b:Math.max(0,Math.min(f.b!==void 0?f.b:0,1))};if(f.alpha!==void 0)m.alpha=f.alpha;return m},k9=(f)=>x2(p2(f)),N2=(f)=>{return f!==void 0&&(f.r===void 0||f.r>=0&&f.r<=1)&&(f.g===void 0||f.g>=0&&f.g<=1)&&(f.b===void 0||f.b>=0&&f.b<=1)};function L2(f){return N2(p2(f))}function g5(f="rgb"){let{gamut:m}=w0(f);if(!m)return($)=>!0;let M=e(typeof m==="string"?m:f);return($)=>N2(M($))}function R2(f="rgb"){let{gamut:m}=w0(f);if(!m)return(Z)=>m5(Z);let M=typeof m==="string"?m:f,$=e(M),H=g5(M);return(Z)=>{let z=m5(Z);if(!z)return;let X=$(z);if(H(X))return z;let J=x2(X);if(z.mode===J.mode)return J;return e(z.mode)(J)}}function i1(f,m="lch",M="rgb"){f=m5(f);let $=M==="rgb"?L2:g5(M),H=M==="rgb"?k9:R2(M);if(f===void 0||$(f))return f;let Z=e(f.mode);f=e(m)(f);let z={...f,c:0};if(!$(z))return Z(H(z));let X=0,J=f.c!==void 0?f.c:0,B=w0(m).ranges.c,W=(B[1]-B[0])/Math.pow(2,13),Q=z.c;while(J-X>W)if(z.c=X+(J-X)*0.5,$(z))Q=z.c,X=z.c;else J=z.c;return Z($(z)?z:{...z,c:Q})}var v9=(f,m=K1(),M=($)=>$)=>{let $=f.map((H,Z)=>({color:M(H),i:Z}));return(H,Z=1,z=1/0)=>{if(isFinite(Z))Z=Math.max(1,Math.min(Z,$.length-1));return $.forEach((X)=>{X.d=m(H,X.color)}),$.sort((X,J)=>X.d-J.d).slice(0,Z).filter((X)=>X.d<z).map((X)=>f[X.i])}},h1=v9;var z5=(f={mode:"oklch",l:0.48,c:0.23,h:263})=>D0((m)=>(m?e("oklch")(m):f)??f);var _0=(f,m)=>{let M=2*Math.log((1-f.l)/f.l),$=f.l!==0.5?(Math.exp(M*m)-1)/(Math.exp(M)-1):m,H=f.c>0?f.c*(8*Math.sin(Math.PI*(4*m+1)/6)**3-1)/7:0;return{mode:"oklch",l:$,c:H,h:f.h??0}};var o9=0.71,CQ=R("card-colorscale",({expose:f,first:m,host:M,watch:$})=>{let H=m(".label strong","Add a <strong> element inside .label."),Z=m(".label small","Add a <small> element inside .label.");return f({name:H.textContent?.trim()??"",color:z5()}),[$("name",w(H)),$("color",(z)=>{Z.textContent=Xf(z);let X=new Map,J=z.l>o9,B=J?0.1:0.9;X.set("base",p0(z)),X.set("text",J?"black":"white"),X.set("text-soft",p0(_0(z,B)));for(let W=4;W>0;W--)X.set(`lighten${W*20}`,p0(_0(z,(5+W)/10)));for(let W=1;W<5;W++)X.set(`darken${W*20}`,p0(_0(z,(5-W)/10)));for(let[W,Q]of X)M.style.setProperty(`--color-${W}`,Q)})]});var lf="media-motion",cf="media-theme",hf="media-viewport",tf="media-orientation",AQ=R("context-media",({expose:f,host:m,provideContexts:M})=>{let $=(H,Z)=>{let X=m.getAttribute(H)?.trim();if(!X)return Z;let J=X.match(/em$/)?"em":"px",B=parseFloat(X);return Number.isFinite(B)?B+J:Z};return f({[lf]:Q5((H)=>{let Z=matchMedia("(prefers-reduced-motion: reduce)"),z=(X)=>H(X.matches?"reduce":"no-preference");return Z.addEventListener("change",z),()=>Z.removeEventListener("change",z)},{value:matchMedia("(prefers-reduced-motion: reduce)").matches?"reduce":"no-preference"}),[cf]:Q5((H)=>{let Z=matchMedia("(prefers-color-scheme: dark)"),z=(X)=>H(X.matches?"dark":"light");return Z.addEventListener("change",z),()=>Z.removeEventListener("change",z)},{value:matchMedia("(prefers-color-scheme: dark)").matches?"dark":"light"}),[hf]:(()=>{let H=matchMedia(`(min-width: ${$("sm","32em")})`),Z=matchMedia(`(min-width: ${$("md","48em")})`),z=matchMedia(`(min-width: ${$("lg","72em")})`),X=matchMedia(`(min-width: ${$("xl","104em")})`),J=()=>{if(X.matches)return"xl";if(z.matches)return"lg";if(Z.matches)return"md";if(H.matches)return"sm";return"xs"};return Q5((B)=>{let W=()=>B(J());return H.addEventListener("change",W),Z.addEventListener("change",W),z.addEventListener("change",W),X.addEventListener("change",W),()=>{H.removeEventListener("change",W),Z.removeEventListener("change",W),z.removeEventListener("change",W),X.removeEventListener("change",W)}},{value:J()})})(),[tf]:Q5((H)=>{let Z=matchMedia("(orientation: landscape)"),z=(X)=>H(X.matches?"landscape":"portrait");return Z.addEventListener("change",z),()=>Z.removeEventListener("change",z)},{value:matchMedia("(orientation: landscape)").matches?"landscape":"portrait"})}),[M([lf,cf,hf,tf])]});var uQ=R("card-mediaqueries",({expose:f,first:m,requestContext:M,watch:$})=>{let H=m(".motion"),Z=m(".theme"),z=m(".viewport"),X=m(".orientation");return f({motion:M(lf,"unknown"),theme:M(cf,"unknown"),viewport:M(hf,"unknown"),orientation:M(tf,"unknown")}),[H&&$("motion",w(H)),Z&&$("theme",w(Z)),z&&$("viewport",w(z)),X&&$("orientation",w(X))]});var EQ=R("form-checkbox",({expose:f,first:m,host:M,on:$,watch:H})=>{let Z=m('input[type="checkbox"]',"Add a native checkbox."),z=m(".label")??m("label");return f({checked:Z.checked,label:z?.textContent??""}),[$(Z,"change",()=>({checked:Z.checked})),H("checked",(X)=>{Z.checked=X,M.toggleAttribute("checked",X)}),z&&H("label",w(z))]});var E5=g5("p3"),r9=g5("rgb"),n9=new Intl.NumberFormat("en-US",{maximumFractionDigits:2}).format,l9=new Intl.NumberFormat("en-US",{maximumFractionDigits:4}).format,D2=20,t1=0.71,K0={l:1,c:0.4,h:360},a1={l:0.0025,c:0.001,h:1},c9={l:0.05,c:0.02,h:15},V2=(f,m)=>m?c9[f]:a1[f],oQ=R("form-colorgraph",({all:f,expose:m,first:M,host:$,on:H,watch:Z})=>{let z={l:M('input[name="lightness"]','Add an <input[name="lightness"]> element to control the lightness of the color.'),c:M('input[name="chroma"]','Add an <input[name="chroma"]> element to control the chroma of the color.'),h:M('input[name="hue"]','Add an <input[name="hue"]> element to control the hue of the color.')},X=M(".graph","Add a <.graph> element as a container for the color graph."),J=M(".graph canvas","Add a <canvas> element inside the graph to display the lightness/chroma graph."),B=M(".slider","Add a <.slider> element as a container for track and thumb."),W=M(".slider canvas","Add a <canvas> element inside the slider to display the hue slider track."),Q=M(".knob","Add a <.knob> element as a drag knob to control lightness and chroma."),j=M(".thumb","Add a <.thumb> element as a drag knob to control the hue."),b=f("input"),U=f(".error"),p=f("button.decrement"),N=f("button.increment");for(let[P,q]of Object.entries(z))q.min="0",q.max=P==="l"?"100":P==="c"?"0.4":"360",q.step="any";B.setAttribute("aria-valuemin","0"),B.setAttribute("aria-valuemax","360");let K=k(X.getBoundingClientRect().width),u=v(()=>K.get()-2*D2),o={l:k(""),c:k(""),h:k("")},D=(P,q)=>{let V=P==="l"?q*100:q;return P==="c"?l9(V):n9(V)},Y=(P)=>{if(P.closest(".lightness"))return"l";if(P.closest(".chroma"))return"c";if(P.closest(".hue"))return"h";return null},C=(P,q,V,_=1)=>p0({mode:"oklch",l:1-q,c:P*K0.c,h:V,alpha:_}),G=(P,q)=>{let V=K.get(),_=Math.round(q.c*V/K0.c),L=Math.round((1-q.l)*V);P.style.setProperty("background-color",p0(q)),P.style.setProperty("border-color",q.l>t1?"black":"white"),P.style.setProperty("left",`${_}px`),P.style.setProperty("top",`${L}px`)},I=(P)=>{let q={...$.color,h:P*K0.h};if(r9(q))return q;if(E5(q))q.alpha=0.5;else q.alpha=0;return q},r=(P)=>{i(()=>{$.color=P;for(let q of["l","c","h"])o[q].set("")})},d=(P)=>P==="l"?$.lightness:P==="c"?$.chroma:$.hue,l=(P,q)=>{let V=Math.round(q/a1[P])*a1[P];if(V<0||V>K0[P])return;let _={...$.color,[P]:V};if(E5(_))r(_);else z[P].setCustomValidity("Color out of gamut"),o[P].set(z[P].validationMessage)},J0=a0((P,q,V,_,L)=>{let F={...$.color,c:Math.min(Math.max((P-_)/L,0),1)*K0.c,l:1-Math.min(Math.max((q-V)/L,0),1)};if(E5(F))r(F)}),j0=a0((P,q,V)=>{let _={...$.color,h:Math.min(Math.max((P-q)/V,0),1)*K0.h};if(E5(_))r(_)});m({color:z5(),lightness:()=>$.color.l,chroma:()=>$.color.c,hue:()=>$.color.h??0,stepDown:s0((P,q=!1)=>{l(P,d(P)-V2(P,q))}),stepUp:s0((P,q=!1)=>{l(P,d(P)+V2(P,q))})});let m0=[Z(()=>X,()=>{let P=a0((V)=>{K.set(V)}),q=new ResizeObserver(()=>{P(X.clientWidth)});return q.observe(X),()=>{q.disconnect(),P.cancel()}}),Z(()=>p0($.color),e0($,"--color-base")),t(b,(P)=>{let q=Y(P);return[q&&Z(o[q],(V)=>{if(P.ariaInvalid=String(!!V),V&&P.id)P.setAttribute("aria-errormessage",`${P.id}-error`);else P.removeAttribute("aria-errormessage")}),Z("color",(V)=>{if(q)P.value=D(q,V[q]??0)}),H(P,"change",()=>{if(!q)return;let V=P.valueAsNumber,_={...$.color,[q]:q==="l"?V/100:V};if(E5(_))r(_);else P.setCustomValidity("Color out of gamut"),o[q].set(P.validationMessage)})]}),t(U,(P)=>{let q=Y(P);return[q?Z(o[q],w(P)):!1]}),H(X,"pointerdown",(P)=>{let{top:q,left:V}=J.getBoundingClientRect(),_=K.get();Q.ariaPressed="true",X.setPointerCapture(P.pointerId);let L=(y)=>{let O=(y.getCoalescedEvents?.()||[]).pop()||y;J0(O.clientX,O.clientY,q,V,_)},F=()=>{X.removeEventListener("pointermove",L),X.removeEventListener("pointerup",F),J0.cancel(),Q.ariaPressed="false"};X.addEventListener("pointermove",L,{passive:!0}),X.addEventListener("pointerup",F)}),Z(()=>`${K.get()}px`,e0(X,"--canvas-size")),Z(()=>({hue:$.hue,n:Math.round(K.get())}),({hue:P,n:q})=>{J.width=q,J.height=q;let V=J.getContext("2d",{colorSpace:"display-p3"});if(!V)return;let _=(y,O="rgb")=>i1({mode:"oklch",l:y,c:K0.c,h:P},"oklch",O).c/K0.c,L=(y,O,E,M0=1)=>[C(y,E,P,M0),C(O,E,P,M0)],F=(y,O,E="rgb")=>{let M0=_(1-O/q,E)*q,N5=V.createLinearGradient(y,0,M0,0),Wf=L(y/q,M0/q,O/q,E==="p3"?0.5:1);return N5.addColorStop(0,Wf[0]),N5.addColorStop(1,Wf[1]),V.fillStyle=N5,V.fillRect(y,O,M0-y,1),[M0,Wf[1]]};V.clearRect(0,0,q,q);for(let y=0;y<q;y++){let[O,E]=F(0,y);if(E5(E))F(O,y,"p3")}}),Z(()=>({l:$.lightness,c:$.chroma,size:K.get()}),({l:P,c:q,size:V})=>{Q.style.setProperty("top",`${Math.round((1-P)*V)}px`),Q.style.setProperty("left",`${Math.round(q*V/K0.c)}px`),Q.style.setProperty("--color-border",P>t1?"black":"white")}),H(B,"pointerdown",(P)=>{let q=W.getBoundingClientRect().left,V=u.get();j.ariaPressed="true",B.setPointerCapture(P.pointerId);let _=(F)=>{let y=(F.getCoalescedEvents?.()||[]).pop()||F;j0(y.clientX,q,V)},L=()=>{B.removeEventListener("pointermove",_),B.removeEventListener("pointerup",L),j0.cancel(),j.ariaPressed="false"};B.addEventListener("pointermove",_,{passive:!0}),B.addEventListener("pointerup",L)}),Z(()=>`${u.get()}px`,e0(B,"--track-width")),Z("hue",(P)=>{B.setAttribute("aria-valuenow",String(P)),B.setAttribute("aria-valuetext",`${D("h",P)}°`)}),Z(()=>({color:$.color,n:Math.round(u.get())}),({n:P})=>{W.width=P;let q=W.getContext("2d",{colorSpace:"display-p3"});if(!q)return;q.clearRect(0,0,P,1);for(let V=0;V<P;V++)q.fillStyle=p0(I(V/P)),q.fillRect(V,0,1,1)}),Z(()=>({hue:$.hue,l:$.lightness,tw:u.get()}),({hue:P,l:q,tw:V})=>{j.style.setProperty("left",`${Math.round(P*V/K0.h)+D2}px`),j.style.setProperty("--color-border",q>t1?"black":"white")}),t(p,(P)=>{let q=Y(P);return[H(P,"click",(V)=>{if(q)$.stepDown(q,V.shiftKey)}),Z("color",(V)=>{if(!q){P.disabled=!0;return}P.disabled=(V[q]??0)<=0})]}),t(N,(P)=>{let q=Y(P);return[H(P,"click",(V)=>{if(q)$.stepUp(q,V.shiftKey)}),Z("color",(V)=>{if(!q){P.disabled=!0;return}P.disabled=(V[q]??0)>=K0[q]})]}),H($,"keydown",(P)=>{let{key:q,shiftKey:V}=P,_=P.target;if(!_||_.localName==="input"&&(q==="ArrowLeft"||q==="ArrowRight"))return;if(q.substring(0,5)==="Arrow"||["+","-"].includes(q)){P.preventDefault(),P.stopPropagation();let L=Y(_);if(L){if(q==="ArrowLeft"||q==="ArrowDown"||q==="-")$.stepDown(L,V);else if(q==="ArrowRight"||q==="ArrowUp"||q==="+")$.stepUp(L,V)}else if(_.role==="slider"){if(q==="ArrowLeft"||q==="ArrowDown"||q==="-")$.stepDown("h",V);else if(q==="ArrowRight"||q==="ArrowUp"||q==="+")$.stepUp("h",V)}else switch(q){case"ArrowDown":$.stepDown("l",V);break;case"ArrowUp":$.stepUp("l",V);break;case"ArrowLeft":$.stepDown("c",V);break;case"ArrowRight":$.stepUp("c",V);break;case"-":$.stepDown("h");break;case"+":$.stepUp("h");break}}})];for(let P=1;P<5;P++){let q=M(`li.lighten${(5-P)*20}`);if(q)m0.push(Z(()=>({color:$.color,size:K.get()}),({color:V})=>{G(q,_0(V,1-P/10))}))}for(let P=1;P<5;P++){let q=M(`li.darken${P*20}`);if(q)m0.push(Z(()=>({color:$.color,size:K.get()}),({color:V})=>{G(q,_0(V,1-(P+5)/10))}))}return m0});var lQ=R("form-combobox",({expose:f,first:m,host:M,on:$,pass:H,watch:Z})=>{let z=m("input","Needed to enter value."),X=m("form-listbox","Needed to display options."),J=m("button.clear"),B=m("form-combobox > .error"),W=m(".description"),Q=B?.id,j=W?.id;if(j)z.setAttribute("aria-describedby",j);let b=k(!1),U=v(()=>b.get()&&X.options.length>0),p=k(z.value.length);return f({value:z.value,length:p.get,error:"",description:W?.textContent?.trim()??"",clear:s0(()=>{M.value="",z.value="",z.setCustomValidity(""),z.checkValidity(),z.dispatchEvent(new Event("input",{bubbles:!0})),z.dispatchEvent(new Event("change",{bubbles:!0})),z.focus()})}),[H(X,{filter:()=>M.value}),$(M,"keyup",({key:N})=>{if(N==="Escape")b.set(!1),z.focus();if(N==="Delete")M.clear()}),$(z,"input",()=>{p.set(z.value.length),z.checkValidity(),i(()=>{M.value=z.value,M.error=z.validationMessage??"",b.set(!0)})}),$(z,"keydown",({key:N,altKey:K})=>{if(N==="ArrowDown"){if(K)b.set(!0);if(U.get())X.options[0]?.focus()}}),$(X,"change",({target:N})=>{if(N instanceof HTMLInputElement)z.value=N.value,z.checkValidity(),i(()=>{M.value=N.value,M.error=z.validationMessage??"",b.set(!1),z.focus()})}),$(J,"click",()=>{M.clear()}),Z("value",q5(M,"value")),Z("error",(N)=>{if(z.ariaInvalid=String(!!N),N&&Q)z.setAttribute("aria-errormessage",Q);else z.removeAttribute("aria-errormessage")}),B&&Z("error",w(B)),W&&Z("description",w(W)),Z(U,(N)=>{X.hidden=!N,z.ariaExpanded=String(N)}),J&&Z(p,z0(J))]});var i9=0,hQ=R("form-inplace-edit",({expose:f,first:m,host:M,on:$,watch:H})=>{let Z=m(".text",'Add an element with "text" class for label display.'),z=m("button","Add a <button> element for edit mode toggle."),X=`form-inplace-edit-input${++i9}`,J=null;return f({editing:s5(),value:Z.textContent?.trim()??""}),[$(z,"click",(B)=>{if(B.stopPropagation(),M.editing&&J)return{editing:!1,value:J.value};else M.editing=!M.editing}),$(Z,"dblclick",()=>({editing:!0})),$(M,"keydown",(B)=>{if(!M.editing)return;if(B.key!=="Escape"&&B.key!=="Enter")return;if(B.preventDefault(),J&&B.key==="Enter")return{editing:!1,value:J.value};else M.editing=!1}),$(z,"mousedown",(B)=>{B.preventDefault()}),$(M,"focusout",(B)=>{if(!M.editing)return;let W=B.relatedTarget;if(W&&M.contains(W))return;M.editing=!1}),H("value",w(Z)),H("editing",(B)=>{if(M.toggleAttribute("editing",B),B){let W=document.createElement("form-textbox"),Q=document.createElement("label");Q.className="visually-hidden",Q.setAttribute("for",X),Q.textContent="Edit";let j=document.createElement("div");j.className="input",J=document.createElement("input"),J.type="text",J.id=X,J.value=M.value,j.append(J),W.append(Q,j),M.insertBefore(W,Z),z.textContent="✓",z.setAttribute("aria-label","Accept"),J.focus(),J.select()}else m("form-textbox")?.remove(),z.textContent="✎",z.setAttribute("aria-label","Edit")})]});var Y2=new Map,h9=(f)=>{let m=f.toLowerCase().split(",").map(($)=>$.trim()),M={noCache:!1,noStore:!1,maxAge:void 0};for(let $ of m)if($==="no-cache")M.noCache=!0;else if($==="no-store")M.noStore=!0;else if($.startsWith("max-age=")){let H=parseInt($.substring(8),10);if(!isNaN(H))M.maxAge=H}return M},t9=(f)=>{if(f.maxAge!==void 0)return(Date.now()-f.timestamp)/1000<f.maxAge;return!0},af=(f,m,M="src")=>!!f&&!!(m.parentElement||m.getRootNode().host)?.closest(`${m.localName}[${M}="${f}"]`),sf=(f)=>{if(!f)return!1;try{if(new URL(f,location.href).origin===location.origin)return!0}catch{return!1}return!1},ef=async(f,m,M=($)=>$.text())=>{let $=Y2.get(f),H={};if($?.etag)H["If-None-Match"]=$.etag;if($?.lastModified)H["If-Modified-Since"]=$.lastModified;let Z=await fetch(f,{signal:m??null,headers:H});if(Z.status===304&&$)return{content:$.content,fromCache:!0};if(!Z.ok)throw Error(`HTTP error: ${Z.statusText}`);let z=await M(Z),X=Z.headers.get("cache-control"),J=Z.headers.get("etag"),B=Z.headers.get("last-modified"),W=X?h9(X):{noCache:!1,noStore:!1,maxAge:void 0};if(!W.noStore){let Q={content:z,timestamp:Date.now(),etag:J||void 0,lastModified:B||void 0,maxAge:W.maxAge};if(!W.noCache||t9(Q))Y2.set(f,Q)}return{content:z,fromCache:!1}};function G2(f,m){if(!m)return V0(f);let M=typeof m==="string"?new RegExp(m.replace(/[.*+?^${}()|[\]\\]/g,"\\$&"),"giu"):m,$=[],H;M.lastIndex=0;while((H=M.exec(f))!==null)if($.push({start:H.index,end:H.index+H[0].length,text:H[0]}),H.index===M.lastIndex)M.lastIndex++;if($.length===0)return V0(f);let Z="",z=0;for(let{start:X,end:J,text:B}of $)Z+=V0(f.slice(z,X)),Z+=`<mark>${V0(B)}</mark>`,z=J;return Z+=V0(f.slice(z)),Z}var s1=(f,...m)=>String.raw({raw:f},...m);var a9="Enter",s9=["ArrowUp"],F2=["ArrowDown"],C2="Home",K2="End",e9=[...s9,...F2,C2,K2],ZU=R("form-listbox",({all:f,expose:m,first:M,host:$,on:H,watch:Z})=>{let z=M('input[type="hidden"]',"Needed to store the selected value."),X=M("input.filter"),J=M("button.clear"),B=M("card-callout"),W=M(".loading"),Q=M(".error"),j=M('[role="listbox"]',"Needed to display list of options."),b=f('button[role="option"]'),U=(D)=>D.map((Y)=>s1`<button
 							type="button"
 							role="option"
 							tabindex="-1"
 							value="${V0(Y.value)}"
 						>
 							${V0(Y.label)}
-						</button>`).join(""),p=(D)=>{let Y=$.id,C="";for(let[G,I]of Object.entries(D)){let r=`${Y}-${V0(G)}`;C+=e1`<div role="group" aria-labelledby="${r}">
+						</button>`).join(""),p=(D)=>{let Y=$.id,C="";for(let[G,I]of Object.entries(D)){let r=`${Y}-${V0(G)}`;C+=s1`<div role="group" aria-labelledby="${r}">
 					<div role="presentation" id="${r}">
 						${V0(I.label)}
 					</div>
 					${U(I.items)}
-				</div>`}return C},N=I0(async(D,Y)=>{let C=$.src;if(!C)throw Error("No URL provided");if(!sf(C))throw Error("Invalid URL");if(af(C,$))throw Error("Recursive URL detected");try{let{content:G}=await ef(C,Y,(I)=>I.json());return Array.isArray(G)?U(G):p(G)}catch(G){throw Error(`Failed to fetch content for "${C}": ${String(G)}`)}}),K=v(()=>$.filter.toLowerCase()),u=()=>Array.from(j.querySelectorAll('button[role="option"]:not([hidden])')),o=u().findIndex((D)=>D.ariaSelected==="true");return m({value:M('button[role="option"][aria-selected="true"]')?.value??"",options:Ff(j,'button[role="option"]:not([hidden])'),filter:"",src:y0()}),[H(X,"input",(D,Y)=>({filter:Y.value??""})),H(J,"click",()=>({filter:""})),H(j,"click",({target:D})=>{let Y=D.closest('[role="option"]');if(Y&&Y.value!==$.value)$.value=Y.value,z.dispatchEvent(new Event("change",{bubbles:!0}))}),H(j,"keydown",(D)=>{let{key:Y}=D;if(!e9.includes(Y))return;let C=u();if(D.preventDefault(),D.stopPropagation(),Y===C2)o=0;else if(Y===K2)o=C.length-1;else o=(o+(F2.includes(Y)?1:-1)+C.length)%C.length;C[o]?.focus()}),H(j,"keyup",({key:D})=>{if(D!==a9)return;u()[o]?.click()}),Z("value",(D)=>{$.setAttribute("value",D),z.value=D}),$.src&&Z(N,{nil:()=>{if(B)B.hidden=!1;if(W)return W.hidden=!1,()=>{W.hidden=!1}},ok:(D)=>{if(B)B.hidden=!0;if(W)W.hidden=!0;if(Q)Q.hidden=!0;return j.hidden=!1,t5(j,()=>{j.innerHTML=D}),()=>{j.hidden=!0}},err:(D)=>{if(B)B.hidden=!1,B.classList.add("danger");if(Q)Q.hidden=!1,Q.textContent=D.message;return()=>{if(B)B.classList.remove("danger");if(Q)Q.hidden=!0,Q.textContent=""}}}),t(b,(D)=>{let Y=D.textContent,C=Y?.trim().toLowerCase();return[Z(K,(G)=>{D.hidden=!C?.includes(G),D.innerHTML=G2(Y,G)}),Z("value",()=>{let G=$.value===D.value;D.tabIndex=G?0:-1,D.ariaSelected=String(G)})]}),J&&Z(K,z0(J))]});var f3="Enter",m3=["ArrowLeft","ArrowUp"],O2=["ArrowRight","ArrowDown"],A2="Home",I2="End",M3=[...m3,...O2,A2,I2],$3=(f)=>f.findIndex((m)=>m.checked),bU=R("form-radiogroup",({all:f,expose:m,host:M,on:$,watch:H})=>{let Z=f('input[type="radio"]',"Add at least two native radio buttons."),z=$3(Z.get());return m({value:Z.get()[z]?.value??""}),[$(Z,"change",(X,J)=>({value:J.value})),$(M,"click",({target:X})=>{if(!(X instanceof HTMLElement))return;if(X.hasAttribute("value"))z=Z.get().indexOf(X)}),$(M,"keydown",(X)=>{let{key:J}=X;if(!M3.includes(J))return;let B=Z.get();if(X.preventDefault(),X.stopPropagation(),J===A2)z=0;else if(J===I2)z=B.length-1;else z=(z+(O2.includes(J)?1:-1)+B.length)%B.length;B[z]?.focus()}),$(M,"keyup",({key:X})=>{if(X!==f3)return;Z.get()[z]?.click()}),t(Z,(X)=>H(()=>X.value===M.value,(J)=>{X.checked=J,X.tabIndex=J?0:-1,X.closest("label")?.classList.toggle("selected",J)}))]});var jU=R("form-spinbutton",({all:f,expose:m,first:M,host:$,on:H,watch:Z})=>{let z=f("button, input:not([disabled])"),X=M("button.increment","Add a native button to increment the value"),J=M("button.decrement","Add a native button to decrement the value"),B=M("input.value","Add a native input to display the value"),W=M(".zero"),Q=M(".other"),j=v(()=>$.value!==0),b=X.ariaLabel||"Increment";return m({value:Number.parseInt(B.value)||0,max:Number.parseInt(B.max)||10}),[H(z,"change",(U,p)=>{if(!(p instanceof HTMLInputElement))return;let N=Number(p.value);if(!Number.isInteger(N)){p.value=String($.value),p.checkValidity();return}let K=Math.min($.max,Math.max(0,N));if(N!==K)p.value=String(K),p.checkValidity();$.value=K}),H(z,"click",(U,p)=>{if(p.classList.contains("decrement"))$.value=Math.max(0,$.value-1);else if(p.classList.contains("increment"))$.value=Math.min($.max,$.value+1)}),H(z,"keydown",(U)=>{let{key:p}=U;if(["ArrowUp","ArrowDown","-","+"].includes(p)){U.stopPropagation(),U.preventDefault();let N=p==="ArrowDown"||p==="-"?-1:1;$.value=Math.min($.max,Math.max(0,$.value+N))}}),Z(j,(U)=>{B.hidden=!U,J.hidden=!U}),W&&Z(j,(U)=>{W.hidden=U,X.ariaLabel=U?b:W.textContent}),Q&&Z(j,z0(Q)),Z(()=>String($.value),H0(B,"value")),Z(()=>String($.max),H0(B,"max")),Z(()=>$.value>=$.max,H0(X,"disabled"))]});var pU=R("form-textbox",({expose:f,first:m,host:M,on:$,watch:H})=>{let Z=m("input, textarea","Add a native input or textarea as descendant element."),z=m("button.clear"),X=m(".error"),J=m(".description"),B=X?.id,W=J?.id;if(W)Z.setAttribute("aria-describedby",W);let Q=J&&Z.maxLength>0&&J.dataset.remaining?v(()=>J.dataset.remaining.replace("${n}",String(Z.maxLength-M.length))):null,j=k(Z.value.length);return f({value:Z.value,length:j.get,error:"",description:Q??J?.textContent?.trim()??"",clear:s0(()=>{M.value="",Z.value="",Z.setCustomValidity(""),Z.checkValidity(),Z.dispatchEvent(new Event("input",{bubbles:!0})),Z.dispatchEvent(new Event("change",{bubbles:!0})),Z.focus()})}),[$(Z,"change",()=>{return Z.checkValidity(),{value:Z.value,error:Z.validationMessage}}),$(Z,"input",()=>{j.set(Z.value.length)}),$(z,"click",()=>{M.clear()}),H("value",H0(Z,"value")),H("error",(b)=>{if(Z.ariaInvalid=String(!!b),b&&B)Z.setAttribute("aria-errormessage",B);else Z.removeAttribute("aria-errormessage")}),X&&H("error",w(X)),J&&H("description",w(J)),z&&H(j,z0(z))]});var f6=(f,m)=>Math.max(0,Math.min(f,m-1)),LU=R("module-carousel",({all:f,expose:m,first:M,host:$,on:H,watch:Z})=>{let z=f('button[role="tab"]'),X=f('[role="tabpanel"]'),J=f("nav button"),B=M("button.prev","Add a previous button"),W=M("button.next","Add a next button"),Q=!1,j=-1;return m({index:Math.max(X.get().findIndex((b)=>b.ariaCurrent==="true"),0)}),[()=>{let b=new IntersectionObserver((U)=>{for(let p of U)if(p.intersectionRatio>0.5){let N=X.get().indexOf(p.target);if(Q){if(N===$.index)Q=!1}else if(N!==$.index&&N>=0)j=N,$.index=N;break}},{root:$,threshold:0.5});for(let U of X.get())b.observe(U);return()=>b.disconnect()},Z("index",(b)=>{if(j<0){j=b;return}if(j!==b)j=b,Q=!0,X.get()[b].scrollIntoView({behavior:"smooth",block:"nearest"})}),H(B,"click",()=>{let b=f6($.index-1,X.get().length);if($.index=b,b===0)W.focus()}),H(W,"click",()=>{let b=f6($.index+1,X.get().length);if($.index=b,b===X.get().length-1)B.focus()}),H(z,"click",(b,U)=>{$.index=parseInt(U.dataset.index||"0")}),H(J,"keyup",(b)=>{let{key:U}=b;if(!["ArrowLeft","ArrowRight","Home","End"].includes(U))return;b.preventDefault(),b.stopPropagation();let p=X.get().length,N=U==="Home"?0:U==="End"?p-1:f6($.index+(U==="ArrowLeft"?-1:1),p);if($.index=N,N===0&&document.activeElement===B)W.focus();else if(N===p-1&&document.activeElement===W)B.focus();else if(document.activeElement){let K=z.get();if(K.includes(document.activeElement))K[N].focus()}}),t(X,(b)=>Z(()=>String(b.id===X.get()[$.index]?.id),H0(b,"ariaCurrent"))),t(z,(b)=>Z(()=>b.dataset.index===String($.index),(U)=>{b.ariaSelected=String(U),b.tabIndex=U?0:-1})),Z(()=>$.index>0,z0(B)),Z(()=>$.index<X.get().length-1,z0(W))]});var VU=R("module-catalog",({all:f,first:m,pass:M})=>{let $=m("basic-button","Add a button to go to the Shopping Cart"),H=f("form-spinbutton","Add spinbutton components to calculate sum from."),Z=v(()=>H.get().reduce((z,X)=>z+X.value,0));return[M($,{disabled:()=>!Z.get(),badge:()=>Z.get()>0?String(Z.get()):""})]});var y2=(f,m,M)=>()=>{let $=async()=>{let H=m.label,Z="success";try{await navigator.clipboard.writeText(f.textContent?.trim()??"")}catch(X){console.error("Error while trying to use navigator.clipboard.writeText()",X),Z="error"}m.disabled=!0,m.label=M[Z]??H;let z=setTimeout(()=>{if(m.isConnected)m.disabled=!1,m.label=H;else clearTimeout(z)},Z==="success"?1000:3000)};return m.addEventListener("click",$),()=>m.removeEventListener("click",$)};var KU=R("module-codeblock",({expose:f,first:m,host:M,on:$,watch:H})=>{let Z=m("code","Needed as source container to copy from."),z=m("button.overlay"),X=m("basic-button.copy");return f({collapsed:s5()}),[$(z,"click",()=>({collapsed:!1})),X&&y2(Z,X,{success:X.getAttribute("copy-success")||"Copied!",error:X.getAttribute("copy-error")||"Error trying to copy to clipboard!"}),H("collapsed",q5(M,"collapsed"))]});var H3=t1(Object.keys(e5),A1()),uU=R("module-coloreditor",({expose:f,first:m,host:M,on:$,pass:H})=>{let Z=m("form-textbox"),z=m("form-colorgraph"),X=m("card-colorscale"),J=m("module-colorinfo.base");f({color:z5(),name:y0("Blue"),nearest:()=>H3(M.color)[0]??"",lightness:()=>M.color.l,chroma:()=>M.color.c,hue:()=>M.color.h??0});let B=[$(M,"change",(W)=>{let{target:Q}=W;if(Q instanceof HTMLInputElement&&Q.name==="name")return{name:Q.value}}),Z&&H(Z,{value:"name",description:()=>`Nearest named CSS color: ${M.nearest}`}),z&&H(z,{color:"color"}),X&&H(X,{color:"color",name:"name"}),J&&H(J,{color:"color",name:()=>`${M.name} 500`})];for(let W=1;W<5;W++){let Q=m(`module-colorinfo.lighten${(5-W)*20}`);B.push(H(Q,{color:()=>_0(M.color,1-W/10),name:()=>`${M.name} ${W*100}`}))}for(let W=1;W<5;W++){let Q=m(`module-colorinfo.darken${W*20}`);B.push(H(Q,{color:()=>_0(M.color,1-(W+5)/10),name:()=>`${M.name} ${(W+5)*100}`}))}return B});var dU=R("module-colorinfo",({all:f,expose:m,first:M,host:$,pass:H,watch:Z})=>{let z=M(".label strong","Add a <strong> element inside .label."),X=M(".hex"),J=M(".rgb"),B=M(".hsl"),W=f("basic-number.lightness"),Q=f("basic-number.chroma"),j=f("basic-number.hue");return m({name:z.textContent?.trim()??"",color:z5(),css:()=>p0($.color),hex:()=>Xf($.color),rgb:()=>c1($.color)??"",hsl:()=>i1($.color)??"",lightness:()=>$.color.l,chroma:()=>$.color.c,hue:()=>$.color.h??0}),[H(W,{value:"lightness"}),H(Q,{value:"chroma"}),H(j,{value:"hue"}),Z("css",e0($,"--color-swatch")),Z("hex",e0($,"--color-fallback")),Z("name",w(z)),X&&Z("hex",w(X)),J&&Z("rgb",w(J)),B&&Z("hsl",w(B))]});var m6="scroll-lock",oU=R("module-dialog",({expose:f,first:m,on:M,watch:$})=>{let H=m('button[aria-haspopup="dialog"]',"Add a button to open the dialog."),Z=m("dialog","Add a native dialog element."),z=m("dialog button.close","Add a close button in the dialog."),X=0,J=null;return f({open:!1}),[M(H,"click",()=>({open:!0})),M(z,"click",()=>({open:!1})),M(Z,"click",({target:B})=>B===Z&&{open:!1}),M(Z,"keydown",(B)=>{if(B.key!=="Escape")return;return B.preventDefault(),{open:!1}}),$("open",(B)=>{if(B)X=document.documentElement.scrollTop,J=document.activeElement,Z.showModal(),document.body.classList.add(m6),document.body.style.setProperty("top",`-${X}px`),z.focus();else if(document.body.classList.remove(m6),window.scrollTo({top:X,left:0,behavior:"instant"}),document.body.style.removeProperty("top"),Z.close(),J)J.focus();return()=>{document.body.classList.remove(m6),document.body.style.removeProperty("top"),Z.close()}})]});var cU=R("module-lazyload",({expose:f,first:m,host:M,watch:$})=>{let H=m("card-callout","Needed to display loading state and error messages."),Z=m(".loading","Needed to display loading state."),z=m(".error","Needed to display error messages."),X=m(".content","Needed to display content."),J=I0(async(W,Q)=>{let j=M.src;if(!j)throw Error("No URL provided");if(!sf(j))throw Error("Invalid URL");if(af(j,M))throw Error("Recursive URL detected");try{let{content:b}=await ef(j,Q);return b}catch(b){throw Error(`Failed to fetch content for "${j}": ${String(b)}`)}}),{ok:B}=Q1(X,{allowScripts:M.hasAttribute("allow-scripts")});return f({src:y0()}),[$(J,{ok:(W)=>{H.hidden=!0,Z.hidden=!0,X.hidden=!1,B(W)},nil:()=>{H.hidden=!1,Z.hidden=!1,X.hidden=!0},stale:()=>{return X.style.setProperty("opacity","var(--opacity-dimmed)"),()=>{X.style.removeProperty("opacity")}},err:(W)=>{return H.hidden=!1,H.classList.add("danger"),Z.hidden=!0,z.hidden=!1,z.textContent=W.message,X.hidden=!0,()=>{H.classList.remove("danger"),z.hidden=!0,z.textContent=""}}})]});var u2=(f)=>{let m=f.querySelector('button[role="option"]');if(!m?.value)return null;let M=m.value;if(!M.startsWith("./"))return null;let $=M.indexOf("/",2);if($===-1)return null;return{base:M.slice(0,$+1),ext:M.slice(M.lastIndexOf("."))}},w2=(f,m)=>{if(!f)return null;let M=f.slice(1);if(!M)return null;let $=u2(m);if(!$)return null;return`${$.base}${M}${$.ext}`},Z3=(f,m)=>{if(!f)return"";let M=u2(m);if(!M)return"";let $=f;if($.startsWith(M.base))$=$.slice(M.base.length);let H=$.lastIndexOf(".");if(H>0)$=$.slice(0,H);return $},tU=R("module-listnav",({first:f,pass:m})=>{let M=f("form-listbox","Required to select a partial to load"),$=f("module-lazyload","Required to load a partial into"),H=(X)=>!!M.querySelector(`button[role="option"][value="${CSS.escape(X)}"]`);if(location.hash){let X=w2(location.hash,M);if(X&&H(X))M.value=X}let Z=!1,z=()=>{if(Z)return;let X=w2(location.hash,M);if(X&&X!==M.value&&H(X))i(()=>{M.filter="",M.value=X})};return[m($,{src:()=>M.value}),()=>{let X=R0(()=>{let J=M.value;if(!J)return;let B=Z3(J,M);if(B&&location.hash!==`#${B}`)Z=!0,history.replaceState(null,"",`#${B}`),Z=!1});return window.addEventListener("hashchange",z),()=>{X(),window.removeEventListener("hashchange",z)}}]});var eU=R("module-pagination",({expose:f,first:m,host:M,on:$,watch:H})=>{let Z=m("input",'Add an <input[type="number"]> to enter the page number to go to.'),z=m("button.prev","Add a <button.prev> to go to the previous page."),X=m("button.next","Add a <button.next> to go to the next page."),J=m(".value"),B=m(".max");return f({max:K5(Number(Z.max)??1),value:K5(Z.valueAsNumber??1,M.max)}),[$(M,"keyup",(W)=>{let{key:Q}=W;if(W.target instanceof HTMLInputElement)return;let j=M.value;if((Q==="ArrowLeft"||Q==="-")&&M.value>1)j--;else if((Q==="ArrowRight"||Q==="+")&&M.value<M.max)j++;if(document.activeElement===z&&j<=1)X.focus();else if(document.activeElement===X&&j>=M.max)z.focus();M.value=j}),$(Z,"change",()=>{let W=Z.valueAsNumber,Q=Number.isNaN(W)?1:Math.max(1,Math.min(W,M.max));Z.valueAsNumber=Q,M.value=Q}),$(z,"click",()=>{if(M.value--,M.value<=1)X.focus()}),$(X,"click",()=>{if(M.value++,M.value>=M.max)z.focus()}),H("value",(W)=>{M.setAttribute("value",String(W)),Z.value=String(W),z.disabled=W<=1}),H("max",(W)=>{M.hidden=W<=1,M.setAttribute("max",String(W)),Z.max=String(W)}),H(()=>M.value>=M.max,H0(X,"disabled")),J&&H("value",w(J)),B&&H("max",w(B))]});var T2=0,g2=0.99,z3=(f,m,M)=>($)=>{let H=new IntersectionObserver(([Z])=>{if(!Z)return;if(Z.intersectionRatio>T2&&Z.intersectionRatio<g2)m();else i(M)},{root:$,threshold:[T2,g2]});return H.observe(f),()=>{H.disconnect()}},Mj=R("module-scrollarea",({host:f,on:m,watch:M})=>{let $=f.firstElementChild;if(!$)return[];let H=k(!1),Z=k(!1),z=()=>H.get()||Z.get(),X=f.getAttribute("orientation")==="horizontal"?()=>{H.set(f.scrollLeft>0),Z.set(f.scrollLeft<f.scrollWidth-f.offsetWidth)}:()=>{H.set(f.scrollTop>0),Z.set(f.scrollTop<f.scrollHeight-f.offsetHeight)};return[m(f,"scroll",()=>{if(z())i(X)}),M(()=>H.get()||Z.get(),a5(f,"overflow")),M(H,a5(f,"overflow-start")),M(Z,a5(f,"overflow-end")),()=>z3($,()=>{Z.set(!0)},()=>{H.set(!1),Z.set(!1)})(f)]});var M6=0.1,$6=0.9,E2=0.05,Zj=R("module-splitview",({expose:f,first:m,host:M,on:$,watch:H})=>{let Z=m("button.divider","Add a button.divider resize handle."),z=M.getAttribute("orientation")==="vertical",X=!1;return f({split:p5(0.5)}),[$(Z,"pointerdown",(J)=>{X=!0,J.target.setPointerCapture(J.pointerId)}),$(Z,"pointermove",(J)=>{if(!X)return;let B=M.getBoundingClientRect(),W=z?(J.clientY-B.top)/B.height:(J.clientX-B.left)/B.width;return{split:Math.max(M6,Math.min($6,W))}}),$(Z,"pointerup",()=>{X=!1}),$(Z,"lostpointercapture",()=>{X=!1}),$(Z,"keydown",(J)=>{let{key:B}=J,W=z?B==="ArrowUp":B==="ArrowLeft",Q=z?B==="ArrowDown":B==="ArrowRight";if(W||Q||B==="Home"||B==="End")J.preventDefault();if(W)return{split:Math.max(M6,M.split-E2)};if(Q)return{split:Math.min($6,M.split+E2)};if(B==="Home")return{split:M6};if(B==="End")return{split:$6}}),H("split",(J)=>{M.style.setProperty("--split",`${(J*100).toFixed(2)}%`),Z.setAttribute("aria-valuenow",String(Math.round(J*100)))})]});var Jf=(f)=>f.getAttribute("aria-controls")??"",_2=(f,m,M=0)=>{let H=(f.findIndex(m)+M+f.length)%f.length;return Jf(f[H])},Xj=R("module-tabgroup",({all:f,expose:m,host:M,on:$,watch:H})=>{let Z=f('button[role="tab"]','At least 2 tabs as children of a <[role="tablist"]> element are needed. Each tab must reference a unique id of a <[role="tabpanel"]> element.'),z=f('[role="tabpanel"]',"At least 2 tabpanels are needed. Each tabpanel must have a unique id."),X=(B)=>M.selected===B.getAttribute("aria-controls"),J=k(_2(Z.get(),(B)=>B.ariaSelected==="true"));return m({selected:J.get}),[$(Z,"click",(B,W)=>{J.set(Jf(W))}),$(Z,"keyup",(B,W)=>{let Q=B.key;if(["ArrowLeft","ArrowRight","ArrowUp","ArrowDown","Home","End"].includes(Q)){B.preventDefault(),B.stopPropagation();let j=Z.get(),b=Q==="Home"?Jf(j[0]):Q==="End"?Jf(j[j.length-1]):_2(j,(U)=>U===W,Q==="ArrowLeft"||Q==="ArrowUp"?-1:1);j.filter((U)=>Jf(U)===b)[0].focus(),J.set(b)}}),H("selected",()=>{for(let B of Z.get())B.ariaSelected=String(X(B)),B.tabIndex=X(B)?0:-1;for(let B of z.get())B.hidden=M.selected!==B.id})]});var H6="ABCDEFGHIJKLMNOPQRSTUVWXYZ",B3=0,z6=new Set;function X3(){while(!0){let f=B3++,m=(H6[Math.floor(f/676)%26]??"A")+(H6[Math.floor(f/26)%26]??"A")+(H6[f%26]??"A");if(!z6.has(m))return z6.add(m),m}}var Z6=new Intl.NumberFormat("en-US",{minimumFractionDigits:2,maximumFractionDigits:2}),J3=new Intl.NumberFormat("en-US",{style:"percent",minimumFractionDigits:2,maximumFractionDigits:2,signDisplay:"always"}),W3=new Intl.NumberFormat("en-US",{notation:"compact"}),b3=100,bj=R("module-ticker",({all:f,expose:m,first:M,host:$,on:H,pass:Z,watch:z})=>{let X=M("basic-button.toggle"),J=M("basic-button.add-rows"),B=M("template"),W=$.querySelector("table"),Q=f("tr[data-symbol]"),j=Array.from($.querySelectorAll("tr[data-symbol]")).map((D)=>{let Y=D.dataset.symbol??"";z6.add(Y);let C=parseFloat((D.querySelector(".price")?.textContent??"0").replace(/,/g,""));return{symbol:Y,open:C,price:C,volume:0}}),b=n0(j,{keyConfig:(D)=>D.symbol}),U=new Map,p=$.querySelector("tbody");if(p)U.set(p,j.map((D)=>D.symbol));function N(D){let Y=U.get(D);if(!Y||!B)return;D.innerHTML="";let C=document.createDocumentFragment();for(let G of Y){let I=b.byKey(G)?.get().price??0,d=B.content.cloneNode(!0).firstElementChild;d.dataset.symbol=G;let l=d.querySelector("th"),J0=d.querySelector(".price");if(l)l.textContent=G;if(J0)J0.textContent=Z6.format(I);C.append(d)}D.append(C)}function K(D){let Y=D.offsetHeight;D.innerHTML=`<tr><td colspan="4" style="height:${Y}px;padding:0;border:none"></td></tr>`}let u=new IntersectionObserver((D)=>{for(let Y of D){let C=Y.target,G=!C.querySelector("tr[data-symbol]");if(Y.isIntersecting&&G)N(C);else if(!Y.isIntersecting&&!G)K(C)}},{rootMargin:"400px"});if(p)u.observe(p);let o=()=>{for(let D of b.keys()){if(Math.random()>=$.fraction)continue;b.byKey(D)?.update((Y)=>({...Y,price:Math.max(0.01,Y.price+(Math.random()-0.5)*Y.price*0.004),volume:Math.max(0,Y.volume+Math.round(Math.random()*50000))}))}};return m({running:!0,fraction:p5(0.1)}),[()=>()=>u.disconnect(),H(X,"click",()=>({running:!$.running})),Z(X,{label:()=>$.running?"⏸️ Pause":"▶️ Resume"}),z("running",(D)=>{if(!D)return;let Y=setInterval(o,10);return()=>clearInterval(Y)}),t(Q,(D)=>{let Y=D.dataset.symbol??"",C=b.byKey(Y);if(!C||!D.isConnected)return;let G=D.querySelector(".price"),I=D.querySelector(".change"),r=D.querySelector(".volume");if(!G||!I||!r)return;let d=v(()=>{let{open:l,price:J0}=C.get();return(J0-l)/l});return[z(()=>Z6.format(C.get().price),w(G)),z(()=>J3.format(d.get()),w(I)),z(()=>{let l=d.get();return l>0?"up":l<0?"down":"flat"},H0(D.dataset,"direction")),z(()=>W3.format(C.get().volume),w(r))]}),H(J,"click",()=>{if(!B||!W)return;let D=Array.from({length:b3},()=>{let G=X3(),I=Math.round((10+Math.random()*1000)*100)/100;return{symbol:G,open:I,price:I,volume:0}});b.splice(b.length,0,...D);let Y=document.createElement("tbody");U.set(Y,D.map((G)=>G.symbol));let C=document.createDocumentFragment();for(let{symbol:G,price:I}of D){let d=B.content.cloneNode(!0).firstElementChild;d.dataset.symbol=G;let l=d.querySelector("th"),J0=d.querySelector(".price");if(l)l.textContent=G;if(J0)J0.textContent=Z6.format(I);C.append(d)}Y.append(C),W.append(Y),u.observe(Y)})]});var S2=5,d2="reorder",f1=`button.${d2}`,B6="dragging",Q3=0,jj=R("module-todo",({all:f,first:m,host:M,on:$,pass:H,watch:Z})=>{let z=m("form","Add a form element to enter a new todo item."),X=m("form-textbox","Add <form-textbox> component to enter a new todo item."),J=m("basic-button.submit","Add <basic-button.submit> component to submit the form."),B=m("[data-container]","Add a container element for items."),W=m("template","Add a template element for items."),Q=m('[role="status"]',"Add a live region for status messages."),j=m("basic-pluralize","Add <basic-pluralize> component to display the number of todo items."),b=m("form-radiogroup","Add <form-radiogroup> component to filter todo items."),U=m("basic-button.clear-completed","Add <basic-button.clear-completed> component to clear completed todo items."),p=f(f1),N=f("form-checkbox"),K=f("form-inplace-edit"),u=n0([],{keyConfig:(L)=>L.id,createItem:Y5}),o=v(()=>u.get().filter((L)=>L.completed).length),D=v(()=>u.length-o.get()),Y=k(Q.textContent),C=null,G=null,I=null,r=0,d=null,l=0,J0=0,j0=!1;function m0(L){return L.querySelector("label.text, span")?.textContent?.trim()??"item"}function P(L){if(C=L,L){let F=Array.from(B.children);Y.set(`${m0(L)} selected, position ${F.indexOf(L)+1} of ${u.length}. Press Up or Down arrow to move.`)}}function q(L,F){let y=Array.from(B.children),O=y.indexOf(L)+F;if(O<0||O>=y.length)return;if(F===1)y[O]?.after(L);else y[O]?.before(L);let E=Array.from(B.children).indexOf(L)+1;Y.set(`${m0(L)} moved to position ${E} of ${u.length}.`),L.querySelector(f1)?.focus(),_()}function V(L){if(!I||!G)return;let F=Array.from(B.children).filter((O)=>O!==I&&O!==G),y=null;for(let O of F){let E=O.getBoundingClientRect();if(L<E.top+E.height/2){y=O;break}}if(y)B.insertBefore(I,y);else B.appendChild(I)}function _(){let L=Array.from(B.children).filter((F)=>F instanceof HTMLElement&&F.dataset.key).map((F)=>F.dataset.key);u.update((F)=>{let y=new Map(F.map((O,E)=>[u.keyAt(E),O]));return L.map((O)=>y.get(O)).filter(Boolean)})}return[H(J,{disabled:()=>!X.length}),H(j,{count:()=>D.get()}),H(U,{disabled:()=>!o.get(),badge:()=>o.get()?String(o.get()):""}),t(p,(L)=>{return Z(()=>u.length===1,H0(L,"disabled"))}),t(N,(L)=>{let F=L.closest("[data-key]")?.dataset.key;if(!F||!L.isConnected)return;return H(L,{checked:{get:()=>u.byKey(F)?.completed.get()??!1,set:(y)=>u.byKey(F)?.completed.set(y)}})}),t(K,(L)=>{let F=L.closest("[data-key]")?.dataset.key;if(!F||!L.isConnected)return;return H(L,{value:{get:()=>u.byKey(F)?.label.get()??"",set:(y)=>u.byKey(F)?.label.set(y)}})}),Z(()=>Array.from(u.keys()),(L)=>{let F=new Map;for(let O of B.children){let E=O;if(E.dataset.key)F.set(E.dataset.key,E)}let y=new Set(L);for(let[O,E]of F)if(!y.has(O))E.remove();for(let O=0;O<L.length;O++){let E=L[O],M0=E&&F.get(E);if(E&&!M0){M0=W.content.cloneNode(!0).firstElementChild,M0.dataset.key=E;let X6=`${E}-checkbox`,J6=M0.querySelector("input");if(J6)J6.id=X6;let W6=M0.querySelector("label");if(W6)W6.htmlFor=X6;let b6=M0.querySelector("slot");if(b6)b6.replaceWith(document.createTextNode(u.byKey(E)?.label.get()??""))}let N5=B.children[O];if(M0&&N5!==M0)B.insertBefore(M0,N5??null)}}),$(z,"submit",(L)=>{L.preventDefault();let F=X.value.trim();if(!F)return;u.add({id:`todo${++Q3}`,label:F,createdAt:new Date,completed:!1}),X.clear()}),$(M,"click",(L)=>{if(j0){j0=!1;return}let F=L.target,y=F.closest("[data-key]");if(!(y instanceof HTMLElement))return;if(F.closest("basic-button.remove")){if(L.stopPropagation(),y===C)P(null);let O=y.dataset.key;if(O)u.remove(O)}else if(F.closest(f1))P(y)}),$(M,"keydown",(L)=>{if(!C)return;if(!L.target.classList.contains(d2))return;if(L.key==="Escape"){P(null);return}if(L.key!=="ArrowUp"&&L.key!=="ArrowDown")return;if(L.preventDefault(),L.key==="ArrowUp")q(C,-1);else q(C,1)}),$(M,"pointerdown",(L)=>{let F=L.target.closest(f1);if(!(F instanceof HTMLElement))return;if(!(F.closest("[data-key]")instanceof HTMLElement))return;L.preventDefault(),d=F,l=L.clientY,J0=L.clientX,j0=!1,F.setPointerCapture(L.pointerId),F.focus()}),$(M,"pointermove",(L)=>{if(!d)return;let F=Math.abs(L.clientY-l),y=Math.abs(L.clientX-J0);if(!G&&(F>S2||y>S2)){let O=d.closest("[data-key]");if(!(O instanceof HTMLElement))return;G=O;let E=O.getBoundingClientRect();r=l-E.top,I=document.createElement("li"),I.className="drop-marker",I.style.height=`${E.height-4}px`,B.insertBefore(I,O),O.style.top=`${E.top}px`,O.style.left=`${E.left}px`,O.style.width=`${E.width}px`,O.classList.add(B6)}if(G)G.style.top=`${L.clientY-r}px`,V(L.clientY)}),$(M,"pointerup",()=>{if(G&&I)I.replaceWith(G),G.style.cssText="",G.classList.remove(B6),G=null,I=null,j0=!0,_();d=null}),$(M,"pointercancel",()=>{if(G&&I)I.remove(),G.style.cssText="",G.classList.remove(B6),G=null,I=null;d=null,j0=!1}),$(U,"click",()=>{for(let L=u.length-1;L>=0;L--){let F=u.keyAt(L);if(F&&u.byKey(F)?.completed.get())u.remove(F)}}),Z(()=>b.value||"all",q5(M,"filter")),Z(Y,w(Q))]});var U3="count",pj=R("test-context-provider",({expose:f,provideContexts:m})=>{return f({count:0}),[m(["count"])]}),xj=R("test-context-consumer",({expose:f,first:m,watch:M,requestContext:$})=>{let H=m("#output",'Add element with id="output".');return f({count:$(U3,-1)}),[M("count",(Z)=>{H.textContent=String(Z)})]});R("security-onevil",({expose:f,first:m,host:M,watch:$})=>{let H=m("a","Add an anchor element.");return f({src:y0()}),[$("src",(Z)=>{try{C5(H,"onclick",Z)}catch(z){console.error(`Failed to set attribute 'onclick' on <${H.localName}> in ${M.localName}`,z)}})]});R("security-urlhref",({expose:f,first:m,host:M,watch:$})=>{let H=m("a","Add an anchor element.");return f({href:y0()}),[$("href",(Z)=>{try{C5(H,"href",Z)}catch(z){console.error(`Failed to set attribute 'href' on <${H.localName}> in ${M.localName}`,z)}})]});var Dj=R("test-each",({expose:f,first:m,all:M,on:$,watch:H})=>{m("ul","Add a <ul> element.");let Z=M("li");return f({selected:-1}),[t(Z,(z)=>{let X=Number(z.dataset.index??-1);return[H("selected",(J)=>{z.classList.toggle("active",J===X)}),$(z,"click",()=>({selected:X}))]})]});var Gj=R("test-expose",({expose:f})=>{return f({greeting:"Hello",count:0}),[]});var Kj=R("test-on",({expose:f,first:m,all:M,on:$,watch:H})=>{let Z=m("button","Add a <button> element."),z=M("input"),X=m("#output",'Add element with id="output".'),J=m("#focus-log",'Add element with id="focus-log".');return f({count:0,focused:!1}),[$(Z,"click",()=>({count:1})),H("count",(B)=>{X.textContent=String(B)}),$(z,"click",(B,W)=>({count:Number(W.dataset.value??0)})),$(z,"keydown",({key:B})=>B==="Enter"&&{count:100}),$(z,"focus",()=>({focused:!0})),H("focused",(B)=>{J.textContent=B?"focused":"blurred"})]});var Ij=R("test-pass",({expose:f,first:m,all:M,watch:$,pass:H})=>{let Z=m("basic-number#single",'Add basic-number with id="single".'),z=M("basic-number.group"),X=m("#output",'Add element with id="output".');return f({count:0}),[$("count",(J)=>{X.textContent=String(J)}),H(Z,{value:"count"}),H(z,{value:"count"})]});var uj=R("test-watch",({expose:f,first:m,watch:M})=>{let $=m("#output",'Add element with id="output".'),H=m("#combined",'Add element with id="combined".'),Z=m("#direct",'Add element with id="direct".'),z=m("#handlers",'Add element with id="handlers".'),X=k(0);return f({count:0,label:"hello"}),[M("count",(J)=>{$.textContent=String(J)}),M(["count","label"],([J,B])=>{H.textContent=`${J}:${B}`}),M(X,(J)=>{Z.textContent=String(J),Z.dataset.signal=String(J)}),M("count",{ok:(J)=>{z.textContent=`ok:${J}`}}),[!1,void 0,null,"",0,0n].map((J)=>J&&M("label",()=>{}))]});customElements.define("card-blogpost",class extends HTMLElement{});customElements.define("card-callout",class extends HTMLElement{});customElements.define("module-demo",class extends HTMLElement{});customElements.define("module-toc",class extends HTMLElement{});customElements.define("section-hero",class extends HTMLElement{});customElements.define("section-menu",class extends HTMLElement{});
-
-//# debugId=2ECBC33789E7BB2E64756E2164756E21
+				</div>`}return C},N=I0(async(D,Y)=>{let C=$.src;if(!C)throw Error("No URL provided");if(!sf(C))throw Error("Invalid URL");if(af(C,$))throw Error("Recursive URL detected");try{let{content:G}=await ef(C,Y,(I)=>I.json());return Array.isArray(G)?U(G):p(G)}catch(G){throw Error(`Failed to fetch content for "${C}": ${String(G)}`)}}),K=v(()=>$.filter.toLowerCase()),u=()=>Array.from(j.querySelectorAll('button[role="option"]:not([hidden])')),o=u().findIndex((D)=>D.ariaSelected==="true");return m({value:M('button[role="option"][aria-selected="true"]')?.value??"",options:Ff(j,'button[role="option"]:not([hidden])'),filter:"",src:y0()}),[H(X,"input",(D,Y)=>({filter:Y.value??""})),H(J,"click",()=>({filter:""})),H(j,"click",({target:D})=>{let Y=D.closest('[role="option"]');if(Y&&Y.value!==$.value)$.value=Y.value,z.dispatchEvent(new Event("change",{bubbles:!0}))}),H(j,"keydown",(D)=>{let{key:Y}=D;if(!e9.includes(Y))return;let C=u();if(D.preventDefault(),D.stopPropagation(),Y===C2)o=0;else if(Y===K2)o=C.length-1;else o=(o+(F2.includes(Y)?1:-1)+C.length)%C.length;C[o]?.focus()}),H(j,"keyup",({key:D})=>{if(D!==a9)return;u()[o]?.click()}),Z("value",(D)=>{$.setAttribute("value",D),z.value=D}),$.src&&Z(N,{nil:()=>{if(B)B.hidden=!1;if(W)return W.hidden=!1,()=>{W.hidden=!1}},ok:(D)=>{if(B)B.hidden=!0;if(W)W.hidden=!0;if(Q)Q.hidden=!0;return j.hidden=!1,h5(j,()=>{j.innerHTML=D}),()=>{j.hidden=!0}},err:(D)=>{if(B)B.hidden=!1,B.classList.add("danger");if(Q)Q.hidden=!1,Q.textContent=D.message;return()=>{if(B)B.classList.remove("danger");if(Q)Q.hidden=!0,Q.textContent=""}}}),t(b,(D)=>{let Y=D.textContent,C=Y?.trim().toLowerCase();return[Z(K,(G)=>{D.hidden=!C?.includes(G),D.innerHTML=G2(Y,G)}),Z("value",()=>{let G=$.value===D.value;D.tabIndex=G?0:-1,D.ariaSelected=String(G)})]}),J&&Z(K,z0(J))]});var f3="Enter",m3=["ArrowLeft","ArrowUp"],O2=["ArrowRight","ArrowDown"],A2="Home",I2="End",M3=[...m3,...O2,A2,I2],$3=(f)=>f.findIndex((m)=>m.checked),XU=R("form-radiogroup",({all:f,expose:m,host:M,on:$,watch:H})=>{let Z=f('input[type="radio"]',"Add at least two native radio buttons."),z=$3(Z.get());return m({value:Z.get()[z]?.value??""}),[$(Z,"change",(X,J)=>({value:J.value})),$(M,"click",({target:X})=>{if(!(X instanceof HTMLElement))return;if(X.hasAttribute("value"))z=Z.get().indexOf(X)}),$(M,"keydown",(X)=>{let{key:J}=X;if(!M3.includes(J))return;let B=Z.get();if(X.preventDefault(),X.stopPropagation(),J===A2)z=0;else if(J===I2)z=B.length-1;else z=(z+(O2.includes(J)?1:-1)+B.length)%B.length;B[z]?.focus()}),$(M,"keyup",({key:X})=>{if(X!==f3)return;Z.get()[z]?.click()}),t(Z,(X)=>H(()=>X.value===M.value,(J)=>{X.checked=J,X.tabIndex=J?0:-1,X.closest("label")?.classList.toggle("selected",J)}))]});var bU=R("form-spinbutton",({all:f,expose:m,first:M,host:$,on:H,watch:Z})=>{let z=f("button, input:not([disabled])"),X=M("button.increment","Add a native button to increment the value"),J=M("button.decrement","Add a native button to decrement the value"),B=M("input.value","Add a native input to display the value"),W=M(".zero"),Q=M(".other"),j=v(()=>$.value!==0),b=X.ariaLabel||"Increment";return m({value:Number.parseInt(B.value)||0,max:Number.parseInt(B.max)||10}),[H(z,"change",(U,p)=>{if(!(p instanceof HTMLInputElement))return;let N=Number(p.value);if(!Number.isInteger(N)){p.value=String($.value),p.checkValidity();return}let K=Math.min($.max,Math.max(0,N));if(N!==K)p.value=String(K),p.checkValidity();$.value=K}),H(z,"click",(U,p)=>{if(p.classList.contains("decrement"))$.value=Math.max(0,$.value-1);else if(p.classList.contains("increment"))$.value=Math.min($.max,$.value+1)}),H(z,"keydown",(U)=>{let{key:p}=U;if(["ArrowUp","ArrowDown","-","+"].includes(p)){U.stopPropagation(),U.preventDefault();let N=p==="ArrowDown"||p==="-"?-1:1;$.value=Math.min($.max,Math.max(0,$.value+N))}}),Z(j,(U)=>{B.hidden=!U,J.hidden=!U}),W&&Z(j,(U)=>{W.hidden=U,X.ariaLabel=U?b:W.textContent}),Q&&Z(j,z0(Q)),Z(()=>String($.value),H0(B,"value")),Z(()=>String($.max),H0(B,"max")),Z(()=>$.value>=$.max,H0(X,"disabled"))]});var jU=R("form-textbox",({expose:f,first:m,host:M,on:$,watch:H})=>{let Z=m("input, textarea","Add a native input or textarea as descendant element."),z=m("button.clear"),X=m(".error"),J=m(".description"),B=X?.id,W=J?.id;if(W)Z.setAttribute("aria-describedby",W);let Q=J&&Z.maxLength>0&&J.dataset.remaining?v(()=>J.dataset.remaining.replace("${n}",String(Z.maxLength-M.length))):null,j=k(Z.value.length);return f({value:Z.value,length:j.get,error:"",description:Q??J?.textContent?.trim()??"",clear:s0(()=>{M.value="",Z.value="",Z.setCustomValidity(""),Z.checkValidity(),Z.dispatchEvent(new Event("input",{bubbles:!0})),Z.dispatchEvent(new Event("change",{bubbles:!0})),Z.focus()})}),[$(Z,"change",()=>{return Z.checkValidity(),{value:Z.value,error:Z.validationMessage}}),$(Z,"input",()=>{j.set(Z.value.length)}),$(z,"click",()=>{M.clear()}),H("value",H0(Z,"value")),H("error",(b)=>{if(Z.ariaInvalid=String(!!b),b&&B)Z.setAttribute("aria-errormessage",B);else Z.removeAttribute("aria-errormessage")}),X&&H("error",w(X)),J&&H("description",w(J)),z&&H(j,z0(z))]});var e1=(f,m)=>Math.max(0,Math.min(f,m-1)),pU=R("module-carousel",({all:f,expose:m,first:M,host:$,on:H,watch:Z})=>{let z=f('button[role="tab"]'),X=f('[role="tabpanel"]'),J=f("nav button"),B=M("button.prev","Add a previous button"),W=M("button.next","Add a next button"),Q=!1,j=-1;return m({index:Math.max(X.get().findIndex((b)=>b.ariaCurrent==="true"),0)}),[()=>{let b=new IntersectionObserver((U)=>{for(let p of U)if(p.intersectionRatio>0.5){let N=X.get().indexOf(p.target);if(Q){if(N===$.index)Q=!1}else if(N!==$.index&&N>=0)j=N,$.index=N;break}},{root:$,threshold:0.5});for(let U of X.get())b.observe(U);return()=>b.disconnect()},Z("index",(b)=>{if(j<0){j=b;return}if(j!==b)j=b,Q=!0,X.get()[b].scrollIntoView({behavior:"smooth",block:"nearest"})}),H(B,"click",()=>{let b=e1($.index-1,X.get().length);if($.index=b,b===0)W.focus()}),H(W,"click",()=>{let b=e1($.index+1,X.get().length);if($.index=b,b===X.get().length-1)B.focus()}),H(z,"click",(b,U)=>{$.index=parseInt(U.dataset.index||"0")}),H(J,"keyup",(b)=>{let{key:U}=b;if(!["ArrowLeft","ArrowRight","Home","End"].includes(U))return;b.preventDefault(),b.stopPropagation();let p=X.get().length,N=U==="Home"?0:U==="End"?p-1:e1($.index+(U==="ArrowLeft"?-1:1),p);if($.index=N,N===0&&document.activeElement===B)W.focus();else if(N===p-1&&document.activeElement===W)B.focus();else if(document.activeElement){let K=z.get();if(K.includes(document.activeElement))K[N].focus()}}),t(X,(b)=>Z(()=>String(b.id===X.get()[$.index]?.id),H0(b,"ariaCurrent"))),t(z,(b)=>Z(()=>b.dataset.index===String($.index),(U)=>{b.ariaSelected=String(U),b.tabIndex=U?0:-1})),Z(()=>$.index>0,z0(B)),Z(()=>$.index<X.get().length-1,z0(W))]});var LU=R("module-catalog",({all:f,first:m,pass:M})=>{let $=m("basic-button","Add a button to go to the Shopping Cart"),H=f("form-spinbutton","Add spinbutton components to calculate sum from."),Z=v(()=>H.get().reduce((z,X)=>z+X.value,0));return[M($,{disabled:()=>!Z.get(),badge:()=>Z.get()>0?String(Z.get()):""})]});var y2=(f,m,M)=>()=>{let $=async()=>{let H=m.label,Z="success";try{await navigator.clipboard.writeText(f.textContent?.trim()??"")}catch(X){console.error("Error while trying to use navigator.clipboard.writeText()",X),Z="error"}m.disabled=!0,m.label=M[Z]??H;let z=setTimeout(()=>{if(m.isConnected)m.disabled=!1,m.label=H;else clearTimeout(z)},Z==="success"?1000:3000)};return m.addEventListener("click",$),()=>m.removeEventListener("click",$)};var GU=R("module-codeblock",({expose:f,first:m,host:M,on:$,watch:H})=>{let Z=m("code","Needed as source container to copy from."),z=m("button.overlay"),X=m("basic-button.copy");return f({collapsed:s5()}),[$(z,"click",()=>({collapsed:!1})),X&&y2(Z,X,{success:X.getAttribute("copy-success")||"Copied!",error:X.getAttribute("copy-error")||"Error trying to copy to clipboard!"}),H("collapsed",q5(M,"collapsed"))]});var H3=h1(Object.keys(e5),O1()),IU=R("module-coloreditor",({expose:f,first:m,host:M,on:$,pass:H})=>{let Z=m("form-textbox"),z=m("form-colorgraph"),X=m("card-colorscale"),J=m("module-colorinfo.base");f({color:z5(),name:y0("Blue"),nearest:()=>H3(M.color)[0]??"",lightness:()=>M.color.l,chroma:()=>M.color.c,hue:()=>M.color.h??0});let B=[$(M,"change",(W)=>{let{target:Q}=W;if(Q instanceof HTMLInputElement&&Q.name==="name")return{name:Q.value}}),Z&&H(Z,{value:"name",description:()=>`Nearest named CSS color: ${M.nearest}`}),z&&H(z,{color:"color"}),X&&H(X,{color:"color",name:"name"}),J&&H(J,{color:"color",name:()=>`${M.name} 500`})];for(let W=1;W<5;W++){let Q=m(`module-colorinfo.lighten${(5-W)*20}`);B.push(H(Q,{color:()=>_0(M.color,1-W/10),name:()=>`${M.name} ${W*100}`}))}for(let W=1;W<5;W++){let Q=m(`module-colorinfo.darken${W*20}`);B.push(H(Q,{color:()=>_0(M.color,1-(W+5)/10),name:()=>`${M.name} ${(W+5)*100}`}))}return B});var EU=R("module-colorinfo",({all:f,expose:m,first:M,host:$,pass:H,watch:Z})=>{let z=M(".label strong","Add a <strong> element inside .label."),X=M(".hex"),J=M(".rgb"),B=M(".hsl"),W=f("basic-number.lightness"),Q=f("basic-number.chroma"),j=f("basic-number.hue");return m({name:z.textContent?.trim()??"",color:z5(),css:()=>p0($.color),hex:()=>Xf($.color),rgb:()=>l1($.color)??"",hsl:()=>c1($.color)??"",lightness:()=>$.color.l,chroma:()=>$.color.c,hue:()=>$.color.h??0}),[H(W,{value:"lightness"}),H(Q,{value:"chroma"}),H(j,{value:"hue"}),Z("css",e0($,"--color-swatch")),Z("hex",e0($,"--color-fallback")),Z("name",w(z)),X&&Z("hex",w(X)),J&&Z("rgb",w(J)),B&&Z("hsl",w(B))]});var f6="scroll-lock",dU=R("module-dialog",({expose:f,first:m,on:M,watch:$})=>{let H=m('button[aria-haspopup="dialog"]',"Add a button to open the dialog."),Z=m("dialog","Add a native dialog element."),z=m("dialog button.close","Add a close button in the dialog."),X=0,J=null;return f({open:!1}),[M(H,"click",()=>({open:!0})),M(z,"click",()=>({open:!1})),M(Z,"click",({target:B})=>B===Z&&{open:!1}),M(Z,"keydown",(B)=>{if(B.key!=="Escape")return;return B.preventDefault(),{open:!1}}),$("open",(B)=>{if(B)X=document.documentElement.scrollTop,J=document.activeElement,Z.showModal(),document.body.classList.add(f6),document.body.style.setProperty("top",`-${X}px`),z.focus();else if(document.body.classList.remove(f6),window.scrollTo({top:X,left:0,behavior:"instant"}),document.body.style.removeProperty("top"),Z.close(),J)J.focus();return()=>{document.body.classList.remove(f6),document.body.style.removeProperty("top"),Z.close()}})]});var rU=R("module-lazyload",({expose:f,first:m,host:M,watch:$})=>{let H=m("card-callout","Needed to display loading state and error messages."),Z=m(".loading","Needed to display loading state."),z=m(".error","Needed to display error messages."),X=m(".content","Needed to display content."),J=I0(async(W,Q)=>{let j=M.src;if(!j)throw Error("No URL provided");if(!sf(j))throw Error("Invalid URL");if(af(j,M))throw Error("Recursive URL detected");try{let{content:b}=await ef(j,Q);return b}catch(b){throw Error(`Failed to fetch content for "${j}": ${String(b)}`)}}),{ok:B}=b1(X,{allowScripts:M.hasAttribute("allow-scripts")});return f({src:y0()}),[$(J,{ok:(W)=>{H.hidden=!0,Z.hidden=!0,X.hidden=!1,B(W)},nil:()=>{H.hidden=!1,Z.hidden=!1,X.hidden=!0},stale:()=>{return X.style.setProperty("opacity","var(--opacity-dimmed)"),()=>{X.style.removeProperty("opacity")}},err:(W)=>{return H.hidden=!1,H.classList.add("danger"),Z.hidden=!0,z.hidden=!1,z.textContent=W.message,X.hidden=!0,()=>{H.classList.remove("danger"),z.hidden=!0,z.textContent=""}}})]});var u2=(f)=>{let m=f.querySelector('button[role="option"]');if(!m?.value)return null;let M=m.value;if(!M.startsWith("./"))return null;let $=M.indexOf("/",2);if($===-1)return null;return{base:M.slice(0,$+1),ext:M.slice(M.lastIndexOf("."))}},w2=(f,m)=>{if(!f)return null;let M=f.slice(1);if(!M)return null;let $=u2(m);if(!$)return null;return`${$.base}${M}${$.ext}`},Z3=(f,m)=>{if(!f)return"";let M=u2(m);if(!M)return"";let $=f;if($.startsWith(M.base))$=$.slice(M.base.length);let H=$.lastIndexOf(".");if(H>0)$=$.slice(0,H);return $},cU=R("module-listnav",({first:f,pass:m})=>{let M=f("form-listbox","Required to select a partial to load"),$=f("module-lazyload","Required to load a partial into"),H=(X)=>!!M.querySelector(`button[role="option"][value="${CSS.escape(X)}"]`);if(location.hash){let X=w2(location.hash,M);if(X&&H(X))M.value=X}let Z=!1,z=()=>{if(Z)return;let X=w2(location.hash,M);if(X&&X!==M.value&&H(X))i(()=>{M.filter="",M.value=X})};return[m($,{src:()=>M.value}),()=>{let X=R0(()=>{let J=M.value;if(!J)return;let B=Z3(J,M);if(B&&location.hash!==`#${B}`)Z=!0,history.replaceState(null,"",`#${B}`),Z=!1});return window.addEventListener("hashchange",z),()=>{X(),window.removeEventListener("hashchange",z)}}]});var tU=R("module-pagination",({expose:f,first:m,host:M,on:$,watch:H})=>{let Z=m("input",'Add an <input[type="number"]> to enter the page number to go to.'),z=m("button.prev","Add a <button.prev> to go to the previous page."),X=m("button.next","Add a <button.next> to go to the next page."),J=m(".value"),B=m(".max");return f({max:C5(Number(Z.max)??1),value:C5(Z.valueAsNumber??1,M.max)}),[$(M,"keyup",(W)=>{let{key:Q}=W;if(W.target instanceof HTMLInputElement)return;let j=M.value;if((Q==="ArrowLeft"||Q==="-")&&M.value>1)j--;else if((Q==="ArrowRight"||Q==="+")&&M.value<M.max)j++;if(document.activeElement===z&&j<=1)X.focus();else if(document.activeElement===X&&j>=M.max)z.focus();M.value=j}),$(Z,"change",()=>{let W=Z.valueAsNumber,Q=Number.isNaN(W)?1:Math.max(1,Math.min(W,M.max));Z.valueAsNumber=Q,M.value=Q}),$(z,"click",()=>{if(M.value--,M.value<=1)X.focus()}),$(X,"click",()=>{if(M.value++,M.value>=M.max)z.focus()}),H("value",(W)=>{M.setAttribute("value",String(W)),Z.value=String(W),z.disabled=W<=1}),H("max",(W)=>{M.hidden=W<=1,M.setAttribute("max",String(W)),Z.max=String(W)}),H(()=>M.value>=M.max,H0(X,"disabled")),J&&H("value",w(J)),B&&H("max",w(B))]});var T2=0,g2=0.99,z3=(f,m,M)=>($)=>{let H=new IntersectionObserver(([Z])=>{if(!Z)return;if(Z.intersectionRatio>T2&&Z.intersectionRatio<g2)m();else i(M)},{root:$,threshold:[T2,g2]});return H.observe(f),()=>{H.disconnect()}},eU=R("module-scrollarea",({host:f,on:m,watch:M})=>{let $=f.firstElementChild;if(!$)return[];let H=k(!1),Z=k(!1),z=()=>H.get()||Z.get(),X=f.getAttribute("orientation")==="horizontal"?()=>{H.set(f.scrollLeft>0),Z.set(f.scrollLeft<f.scrollWidth-f.offsetWidth)}:()=>{H.set(f.scrollTop>0),Z.set(f.scrollTop<f.scrollHeight-f.offsetHeight)};return[m(f,"scroll",()=>{if(z())i(X)}),M(()=>H.get()||Z.get(),a5(f,"overflow")),M(H,a5(f,"overflow-start")),M(Z,a5(f,"overflow-end")),()=>z3($,()=>{Z.set(!0)},()=>{H.set(!1),Z.set(!1)})(f)]});var m6=0.1,M6=0.9,E2=0.05,Mj=R("module-splitview",({expose:f,first:m,host:M,on:$,watch:H})=>{let Z=m("button.divider","Add a button.divider resize handle."),z=M.getAttribute("orientation")==="vertical",X=!1;return f({split:p5(0.5)}),[$(Z,"pointerdown",(J)=>{X=!0,J.target.setPointerCapture(J.pointerId)}),$(Z,"pointermove",(J)=>{if(!X)return;let B=M.getBoundingClientRect(),W=z?(J.clientY-B.top)/B.height:(J.clientX-B.left)/B.width;return{split:Math.max(m6,Math.min(M6,W))}}),$(Z,"pointerup",()=>{X=!1}),$(Z,"lostpointercapture",()=>{X=!1}),$(Z,"keydown",(J)=>{let{key:B}=J,W=z?B==="ArrowUp":B==="ArrowLeft",Q=z?B==="ArrowDown":B==="ArrowRight";if(W||Q||B==="Home"||B==="End")J.preventDefault();if(W)return{split:Math.max(m6,M.split-E2)};if(Q)return{split:Math.min(M6,M.split+E2)};if(B==="Home")return{split:m6};if(B==="End")return{split:M6}}),H("split",(J)=>{M.style.setProperty("--split",`${(J*100).toFixed(2)}%`),Z.setAttribute("aria-valuenow",String(Math.round(J*100)))})]});var Jf=(f)=>f.getAttribute("aria-controls")??"",_2=(f,m,M=0)=>{let H=(f.findIndex(m)+M+f.length)%f.length;return Jf(f[H])},Zj=R("module-tabgroup",({all:f,expose:m,host:M,on:$,watch:H})=>{let Z=f('button[role="tab"]','At least 2 tabs as children of a <[role="tablist"]> element are needed. Each tab must reference a unique id of a <[role="tabpanel"]> element.'),z=f('[role="tabpanel"]',"At least 2 tabpanels are needed. Each tabpanel must have a unique id."),X=(B)=>M.selected===B.getAttribute("aria-controls"),J=k(_2(Z.get(),(B)=>B.ariaSelected==="true"));return m({selected:J.get}),[$(Z,"click",(B,W)=>{J.set(Jf(W))}),$(Z,"keyup",(B,W)=>{let Q=B.key;if(["ArrowLeft","ArrowRight","ArrowUp","ArrowDown","Home","End"].includes(Q)){B.preventDefault(),B.stopPropagation();let j=Z.get(),b=Q==="Home"?Jf(j[0]):Q==="End"?Jf(j[j.length-1]):_2(j,(U)=>U===W,Q==="ArrowLeft"||Q==="ArrowUp"?-1:1);j.filter((U)=>Jf(U)===b)[0].focus(),J.set(b)}}),H("selected",()=>{for(let B of Z.get())B.ariaSelected=String(X(B)),B.tabIndex=X(B)?0:-1;for(let B of z.get())B.hidden=M.selected!==B.id})]});var $6="ABCDEFGHIJKLMNOPQRSTUVWXYZ",B3=0,Z6=new Set;function X3(){while(!0){let f=B3++,m=($6[Math.floor(f/676)%26]??"A")+($6[Math.floor(f/26)%26]??"A")+($6[f%26]??"A");if(!Z6.has(m))return Z6.add(m),m}}var H6=new Intl.NumberFormat("en-US",{minimumFractionDigits:2,maximumFractionDigits:2}),J3=new Intl.NumberFormat("en-US",{style:"percent",minimumFractionDigits:2,maximumFractionDigits:2,signDisplay:"always"}),W3=new Intl.NumberFormat("en-US",{notation:"compact"}),b3=100,Xj=R("module-ticker",({all:f,expose:m,first:M,host:$,on:H,pass:Z,watch:z})=>{let X=M("basic-button.toggle"),J=M("basic-button.add-rows"),B=M("template"),W=$.querySelector("table"),Q=f("tr[data-symbol]"),j=Array.from($.querySelectorAll("tr[data-symbol]")).map((D)=>{let Y=D.dataset.symbol??"";Z6.add(Y);let C=parseFloat((D.querySelector(".price")?.textContent??"0").replace(/,/g,""));return{symbol:Y,open:C,price:C,volume:0}}),b=n0(j,{keyConfig:(D)=>D.symbol}),U=new Map,p=$.querySelector("tbody");if(p)U.set(p,j.map((D)=>D.symbol));function N(D){let Y=U.get(D);if(!Y||!B)return;D.innerHTML="";let C=document.createDocumentFragment();for(let G of Y){let I=b.byKey(G)?.get().price??0,d=B.content.cloneNode(!0).firstElementChild;d.dataset.symbol=G;let l=d.querySelector("th"),J0=d.querySelector(".price");if(l)l.textContent=G;if(J0)J0.textContent=H6.format(I);C.append(d)}D.append(C)}function K(D){let Y=D.offsetHeight;D.innerHTML=`<tr><td colspan="4" style="height:${Y}px;padding:0;border:none"></td></tr>`}let u=new IntersectionObserver((D)=>{for(let Y of D){let C=Y.target,G=!C.querySelector("tr[data-symbol]");if(Y.isIntersecting&&G)N(C);else if(!Y.isIntersecting&&!G)K(C)}},{rootMargin:"400px"});if(p)u.observe(p);let o=()=>{for(let D of b.keys()){if(Math.random()>=$.fraction)continue;b.byKey(D)?.update((Y)=>({...Y,price:Math.max(0.01,Y.price+(Math.random()-0.5)*Y.price*0.004),volume:Math.max(0,Y.volume+Math.round(Math.random()*50000))}))}};return m({running:!0,fraction:p5(0.1)}),[()=>()=>u.disconnect(),H(X,"click",()=>({running:!$.running})),Z(X,{label:()=>$.running?"⏸️ Pause":"▶️ Resume"}),z("running",(D)=>{if(!D)return;let Y=setInterval(o,10);return()=>clearInterval(Y)}),t(Q,(D)=>{let Y=D.dataset.symbol??"",C=b.byKey(Y);if(!C||!D.isConnected)return;let G=D.querySelector(".price"),I=D.querySelector(".change"),r=D.querySelector(".volume");if(!G||!I||!r)return;let d=v(()=>{let{open:l,price:J0}=C.get();return(J0-l)/l});return[z(()=>H6.format(C.get().price),w(G)),z(()=>J3.format(d.get()),w(I)),z(()=>{let l=d.get();return l>0?"up":l<0?"down":"flat"},H0(D.dataset,"direction")),z(()=>W3.format(C.get().volume),w(r))]}),H(J,"click",()=>{if(!B||!W)return;let D=Array.from({length:b3},()=>{let G=X3(),I=Math.round((10+Math.random()*1000)*100)/100;return{symbol:G,open:I,price:I,volume:0}});b.splice(b.length,0,...D);let Y=document.createElement("tbody");U.set(Y,D.map((G)=>G.symbol));let C=document.createDocumentFragment();for(let{symbol:G,price:I}of D){let d=B.content.cloneNode(!0).firstElementChild;d.dataset.symbol=G;let l=d.querySelector("th"),J0=d.querySelector(".price");if(l)l.textContent=G;if(J0)J0.textContent=H6.format(I);C.append(d)}Y.append(C),W.append(Y),u.observe(Y)})]});var S2=5,d2="reorder",f1=`button.${d2}`,z6="dragging",Q3=0,bj=R("module-todo",({all:f,first:m,host:M,on:$,pass:H,watch:Z})=>{let z=m("form","Add a form element to enter a new todo item."),X=m("form-textbox","Add <form-textbox> component to enter a new todo item."),J=m("basic-button.submit","Add <basic-button.submit> component to submit the form."),B=m("[data-container]","Add a container element for items."),W=m("template","Add a template element for items."),Q=m('[role="status"]',"Add a live region for status messages."),j=m("basic-pluralize","Add <basic-pluralize> component to display the number of todo items."),b=m("form-radiogroup","Add <form-radiogroup> component to filter todo items."),U=m("basic-button.clear-completed","Add <basic-button.clear-completed> component to clear completed todo items."),p=f(f1),N=f("form-checkbox"),K=f("form-inplace-edit"),u=n0([],{keyConfig:(L)=>L.id,createItem:Y5}),o=v(()=>u.get().filter((L)=>L.completed).length),D=v(()=>u.length-o.get()),Y=k(Q.textContent),C=null,G=null,I=null,r=0,d=null,l=0,J0=0,j0=!1;function m0(L){return L.querySelector("label.text, span")?.textContent?.trim()??"item"}function P(L){if(C=L,L){let F=Array.from(B.children);Y.set(`${m0(L)} selected, position ${F.indexOf(L)+1} of ${u.length}. Press Up or Down arrow to move.`)}}function q(L,F){let y=Array.from(B.children),O=y.indexOf(L)+F;if(O<0||O>=y.length)return;if(F===1)y[O]?.after(L);else y[O]?.before(L);let E=Array.from(B.children).indexOf(L)+1;Y.set(`${m0(L)} moved to position ${E} of ${u.length}.`),L.querySelector(f1)?.focus(),_()}function V(L){if(!I||!G)return;let F=Array.from(B.children).filter((O)=>O!==I&&O!==G),y=null;for(let O of F){let E=O.getBoundingClientRect();if(L<E.top+E.height/2){y=O;break}}if(y)B.insertBefore(I,y);else B.appendChild(I)}function _(){let L=Array.from(B.children).filter((F)=>F instanceof HTMLElement&&F.dataset.key).map((F)=>F.dataset.key);u.update((F)=>{let y=new Map(F.map((O,E)=>[u.keyAt(E),O]));return L.map((O)=>y.get(O)).filter(Boolean)})}return[H(J,{disabled:()=>!X.length}),H(j,{count:()=>D.get()}),H(U,{disabled:()=>!o.get(),badge:()=>o.get()?String(o.get()):""}),t(p,(L)=>{return Z(()=>u.length===1,H0(L,"disabled"))}),t(N,(L)=>{let F=L.closest("[data-key]")?.dataset.key;if(!F||!L.isConnected)return;return H(L,{checked:{get:()=>u.byKey(F)?.completed.get()??!1,set:(y)=>u.byKey(F)?.completed.set(y)}})}),t(K,(L)=>{let F=L.closest("[data-key]")?.dataset.key;if(!F||!L.isConnected)return;return H(L,{value:{get:()=>u.byKey(F)?.label.get()??"",set:(y)=>u.byKey(F)?.label.set(y)}})}),Z(()=>Array.from(u.keys()),(L)=>{let F=new Map;for(let O of B.children){let E=O;if(E.dataset.key)F.set(E.dataset.key,E)}let y=new Set(L);for(let[O,E]of F)if(!y.has(O))E.remove();for(let O=0;O<L.length;O++){let E=L[O],M0=E&&F.get(E);if(E&&!M0){M0=W.content.cloneNode(!0).firstElementChild,M0.dataset.key=E;let B6=`${E}-checkbox`,X6=M0.querySelector("input");if(X6)X6.id=B6;let J6=M0.querySelector("label");if(J6)J6.htmlFor=B6;let W6=M0.querySelector("slot");if(W6)W6.replaceWith(document.createTextNode(u.byKey(E)?.label.get()??""))}let N5=B.children[O];if(M0&&N5!==M0)B.insertBefore(M0,N5??null)}}),$(z,"submit",(L)=>{L.preventDefault();let F=X.value.trim();if(!F)return;u.add({id:`todo${++Q3}`,label:F,createdAt:new Date,completed:!1}),X.clear()}),$(M,"click",(L)=>{if(j0){j0=!1;return}let F=L.target,y=F.closest("[data-key]");if(!(y instanceof HTMLElement))return;if(F.closest("basic-button.remove")){if(L.stopPropagation(),y===C)P(null);let O=y.dataset.key;if(O)u.remove(O)}else if(F.closest(f1))P(y)}),$(M,"keydown",(L)=>{if(!C)return;if(!L.target.classList.contains(d2))return;if(L.key==="Escape"){P(null);return}if(L.key!=="ArrowUp"&&L.key!=="ArrowDown")return;if(L.preventDefault(),L.key==="ArrowUp")q(C,-1);else q(C,1)}),$(M,"pointerdown",(L)=>{let F=L.target.closest(f1);if(!(F instanceof HTMLElement))return;if(!(F.closest("[data-key]")instanceof HTMLElement))return;L.preventDefault(),d=F,l=L.clientY,J0=L.clientX,j0=!1,F.setPointerCapture(L.pointerId),F.focus()}),$(M,"pointermove",(L)=>{if(!d)return;let F=Math.abs(L.clientY-l),y=Math.abs(L.clientX-J0);if(!G&&(F>S2||y>S2)){let O=d.closest("[data-key]");if(!(O instanceof HTMLElement))return;G=O;let E=O.getBoundingClientRect();r=l-E.top,I=document.createElement("li"),I.className="drop-marker",I.style.height=`${E.height-4}px`,B.insertBefore(I,O),O.style.top=`${E.top}px`,O.style.left=`${E.left}px`,O.style.width=`${E.width}px`,O.classList.add(z6)}if(G)G.style.top=`${L.clientY-r}px`,V(L.clientY)}),$(M,"pointerup",()=>{if(G&&I)I.replaceWith(G),G.style.cssText="",G.classList.remove(z6),G=null,I=null,j0=!0,_();d=null}),$(M,"pointercancel",()=>{if(G&&I)I.remove(),G.style.cssText="",G.classList.remove(z6),G=null,I=null;d=null,j0=!1}),$(U,"click",()=>{for(let L=u.length-1;L>=0;L--){let F=u.keyAt(L);if(F&&u.byKey(F)?.completed.get())u.remove(F)}}),Z(()=>b.value||"all",q5(M,"filter")),Z(Y,w(Q))]});var U3="count",jj=R("test-context-provider",({expose:f,provideContexts:m})=>{return f({count:0}),[m(["count"])]}),Pj=R("test-context-consumer",({expose:f,first:m,watch:M,requestContext:$})=>{let H=m("#output",'Add element with id="output".');return f({count:$(U3,-1)}),[M("count",(Z)=>{H.textContent=String(Z)})]});R("security-onevil",({expose:f,first:m,host:M,watch:$})=>{let H=m("a","Add an anchor element.");return f({src:y0()}),[$("src",(Z)=>{try{t5(H,"onclick",Z)}catch(z){console.error(`Failed to set attribute 'onclick' on <${H.localName}> in ${M.localName}`,z)}})]});R("security-urlhref",({expose:f,first:m,host:M,watch:$})=>{let H=m("a","Add an anchor element.");return f({href:y0()}),[$("href",(Z)=>{try{t5(H,"href",Z)}catch(z){console.error(`Failed to set attribute 'href' on <${H.localName}> in ${M.localName}`,z)}})]});var Nj=R("test-each",({expose:f,first:m,all:M,on:$,watch:H})=>{m("ul","Add a <ul> element.");let Z=M("li");return f({selected:-1}),[t(Z,(z)=>{let X=Number(z.dataset.index??-1);return[H("selected",(J)=>{z.classList.toggle("active",J===X)}),$(z,"click",()=>({selected:X}))]})]});var Dj=R("test-expose",({expose:f})=>{return f({greeting:"Hello",count:0}),[]});var Gj=R("test-on",({expose:f,first:m,all:M,on:$,watch:H})=>{let Z=m("button","Add a <button> element."),z=M("input"),X=m("#output",'Add element with id="output".'),J=m("#focus-log",'Add element with id="focus-log".');return f({count:0,focused:!1}),[$(Z,"click",()=>({count:1})),H("count",(B)=>{X.textContent=String(B)}),$(z,"click",(B,W)=>({count:Number(W.dataset.value??0)})),$(z,"keydown",({key:B})=>B==="Enter"&&{count:100}),$(z,"focus",()=>({focused:!0})),H("focused",(B)=>{J.textContent=B?"focused":"blurred"})]});var Kj=R("test-pass",({expose:f,first:m,all:M,watch:$,pass:H})=>{let Z=m("basic-number#single",'Add basic-number with id="single".'),z=M("basic-number.group"),X=m("#output",'Add element with id="output".');return f({count:0}),[$("count",(J)=>{X.textContent=String(J)}),H(Z,{value:"count"}),H(z,{value:"count"})]});var Ij=R("test-watch",({expose:f,first:m,watch:M})=>{let $=m("#output",'Add element with id="output".'),H=m("#combined",'Add element with id="combined".'),Z=m("#direct",'Add element with id="direct".'),z=m("#handlers",'Add element with id="handlers".'),X=k(0);return f({count:0,label:"hello"}),[M("count",(J)=>{$.textContent=String(J)}),M(["count","label"],([J,B])=>{H.textContent=`${J}:${B}`}),M(X,(J)=>{Z.textContent=String(J),Z.dataset.signal=String(J)}),M("count",{ok:(J)=>{z.textContent=`ok:${J}`}}),[!1,void 0,null,"",0,0n].map((J)=>J&&M("label",()=>{}))]});customElements.define("card-blogpost",class extends HTMLElement{});customElements.define("card-callout",class extends HTMLElement{});customElements.define("module-demo",class extends HTMLElement{});customElements.define("module-toc",class extends HTMLElement{});customElements.define("section-hero",class extends HTMLElement{});customElements.define("section-menu",class extends HTMLElement{});
