+++ index.js
@@ -1,3 +1,3 @@
-var P4=Object.getPrototypeOf(async()=>{});function x(Z){return typeof Z==="function"}function n(Z){return x(Z)&&Object.getPrototypeOf(Z)===P4}function YZ(Z){return x(Z)&&Object.getPrototypeOf(Z)!==P4}function w(Z,$){return Z!=null&&Z[Symbol.toStringTag]===$}function L(Z){return Z!==null&&typeof Z==="object"&&Object.getPrototypeOf(Z)===Object.prototype}function $4(Z,$=(J)=>J!=null){return Array.isArray(Z)&&Z.every($)}function HZ(Z){return typeof Z==="string"?`"${Z}"`:!!Z&&typeof Z==="object"?JSON.stringify(Z):String(Z)}class LZ extends Error{constructor(Z){super(`[${Z}] Circular dependency detected`);this.name="CircularDependencyError"}}class kZ extends TypeError{constructor(Z){super(`[${Z}] Signal value cannot be null or undefined`);this.name="NullishSignalValueError"}}class a extends Error{constructor(Z){super(`[${Z}] Signal value is unset`);this.name="UnsetSignalValueError"}}class qZ extends TypeError{constructor(Z,$){super(`[${Z}] Signal value ${HZ($)} is invalid`);this.name="InvalidSignalValueError"}}class hZ extends TypeError{constructor(Z,$){super(`[${Z}] Callback ${HZ($)} is invalid`);this.name="InvalidCallbackError"}}class _Z extends Error{constructor(Z){super(`[${Z}] Signal is read-only`);this.name="ReadonlySignalError"}}class bZ extends Error{constructor(Z){super(`[${Z}] Active owner is required`);this.name="RequiredOwnerError"}}class UZ extends Error{constructor(Z,$,J){super(`[${Z}] Could not add key "${$}"${J?` with value ${JSON.stringify(J)}`:""} because it already exists`);this.name="DuplicateKeyError"}}function C(Z,$,J){if($==null)throw new kZ(Z);if(J&&!J($))throw new qZ(Z,$)}function IZ(Z,$){if($==null)throw new a(Z)}function m(Z,$,J=x){if(!J($))throw new hZ(Z,$)}var e="State",ZZ="Memo",$Z="Task",JZ="Sensor",S="List",XZ="Collection",jZ="Store",WZ="Slot",h=0,MZ=1,P=2,fZ=4,b=8,N=null,T=null,pZ=[],O=0,J4=!1,p=(Z,$)=>Z===$,gZ=(Z,$)=>!1,X4=(Z,$)=>{if(Object.is(Z,$))return!0;if(typeof Z!==typeof $)return!1;if(Z==null||typeof Z!=="object"||$==null||typeof $!=="object")return!1;let J=Array.isArray(Z);if(J!==Array.isArray($))return!1;if(J){let X=Z,j=$;if(X.length!==j.length)return!1;for(let B=0;B<X.length;B++)if(!X4(X[B],j[B]))return!1;return!0}if(L(Z)&&L($)){let X=Object.keys(Z);if(X.length!==Object.keys($).length)return!1;for(let j of X){if(!(j in $))return!1;if(!X4(Z[j],$[j]))return!1}return!0}return!1},s=(Z,$)=>X4(Z,$);function c4(Z,$){let J=$.sourcesTail;if(J){let X=$.sources;while(X){if(X===Z)return!0;if(X===J)break;X=X.nextSource}}return!1}function d(Z,$){let J=$.sourcesTail;if(J?.source===Z)return;let X=null,j=$.flags&fZ;if(j){if(X=J?J.nextSource:$.sources,X?.source===Z){$.sourcesTail=X;return}}let B=Z.sinksTail;if(B?.sink===$&&(!j||c4(B,$)))return;let q={source:Z,sink:$,nextSource:X,prevSink:B,nextSink:null};if($.sourcesTail=Z.sinksTail=q,J)J.nextSource=q;else $.sources=q;if(B)B.nextSink=q;else Z.sinks=q}function u4(Z){let{source:$,nextSource:J,nextSink:X,prevSink:j}=Z;if(X)X.prevSink=j;else $.sinksTail=j;if(j)j.nextSink=X;else $.sinks=X;if(!$.sinks){if($.stop)$.stop(),$.stop=void 0;if("sources"in $&&$.sources){let B=$;B.sourcesTail=null,OZ(B),B.flags|=P}}return J}function OZ(Z){let $=Z.sourcesTail,J=$?$.nextSource:Z.sources;while(J)J=u4(J);if($)$.nextSource=null;else Z.sources=null}function Y(Z,$=P){let J=Z.flags;if("sinks"in Z){if((J&(P|MZ))>=$)return;if(Z.flags=J|$,"controller"in Z&&Z.controller)Z.controller.abort(),Z.controller=void 0;for(let X=Z.sinks;X;X=X.nextSink)Y(X.sink,MZ)}else{if((J&(P|MZ))>=$)return;let X=J&(P|MZ);if(Z.flags=$,!X)pZ.push(Z)}}function l(Z,$){if(Z.equals(Z.value,$))return;Z.value=$;for(let J=Z.sinks;J;J=J.nextSink)Y(J.sink);if(O===0)A()}function VZ(Z,$){if(!Z.cleanup)Z.cleanup=$;else if(Array.isArray(Z.cleanup))Z.cleanup.push($);else Z.cleanup=[Z.cleanup,$]}function vZ(Z){if(!Z.cleanup)return;if(Array.isArray(Z.cleanup))for(let $=0;$<Z.cleanup.length;$++)Z.cleanup[$]();else Z.cleanup();Z.cleanup=null}function i4(Z){let $=N;N=Z,Z.sourcesTail=null,Z.flags=fZ;let J=!1;try{let X=Z.fn(Z.value);if(Z.error||!Z.equals(X,Z.value))Z.value=X,Z.error=void 0,J=!0}catch(X){J=!0,Z.error=X instanceof Error?X:Error(String(X))}finally{N=$,OZ(Z)}if(J){for(let X=Z.sinks;X;X=X.nextSink)if(X.sink.flags&MZ)X.sink.flags|=P}Z.flags=h}function o4(Z){Z.controller?.abort();let $=new AbortController;Z.controller=$,Z.error=void 0;let J=N;N=Z,Z.sourcesTail=null,Z.flags=fZ;let X;try{X=Z.fn(Z.value,$.signal)}catch(j){Z.controller=void 0,Z.error=j instanceof Error?j:Error(String(j));return}finally{N=J,OZ(Z)}l(Z.pendingNode,!0),X.then((j)=>{if($.signal.aborted)return;Z.controller=void 0,k(()=>{if(Z.error||!Z.equals(j,Z.value)){Z.value=j,Z.error=void 0;for(let B=Z.sinks;B;B=B.nextSink)Y(B.sink)}l(Z.pendingNode,!1)})},(j)=>{if($.signal.aborted)return;Z.controller=void 0;let B=j instanceof Error?j:Error(String(j));k(()=>{if(!Z.error||B.name!==Z.error.name||B.message!==Z.error.message){Z.error=B;for(let q=Z.sinks;q;q=q.nextSink)Y(q.sink)}l(Z.pendingNode,!1)})}),Z.flags=h}function j4(Z){vZ(Z);let $=N,J=T;N=T=Z,Z.sourcesTail=null,Z.flags=fZ;try{let X=Z.fn();if(typeof X==="function")VZ(Z,X)}finally{N=$,T=J,OZ(Z)}Z.flags=h}function R(Z){if(Z.flags&MZ)for(let $=Z.sources;$;$=$.nextSource){if("fn"in $.source)R($.source);if(Z.flags&P)break}if(Z.flags&fZ)throw new LZ("controller"in Z?$Z:("value"in Z)?ZZ:"Effect");if(Z.flags&P)if("controller"in Z)o4(Z);else if("value"in Z)i4(Z);else j4(Z);else Z.flags=h}function A(){if(J4)return;J4=!0;try{for(let Z=0;Z<pZ.length;Z++){let $=pZ[Z];if($.flags&(P|MZ))R($)}pZ.length=0}finally{J4=!1}}function k(Z){O++;try{Z()}finally{if(O--,O===0)A()}}function f(Z){let $=N;N=null;try{return Z()}finally{N=$}}function E(Z,$){let J=T,X={cleanup:null};T=X;let j=()=>vZ(X);try{let B=Z();if(typeof B==="function")VZ(X,B);return j}finally{if(T=J,!$?.root&&J)VZ(J,j)}}function N4(Z){let $=T;T=null;try{return Z()}finally{T=$}}function i(Z,$){return $?()=>{if(N){if(!Z.sinks)Z.stop=$();d(Z,N)}}:()=>{if(N)d(Z,N)}}function g(Z,$){C(e,Z,$?.guard);let J={value:Z,sinks:null,sinksTail:null,equals:$?.equals??p,guard:$?.guard};return{[Symbol.toStringTag]:e,get(){if(N)d(J,N);return J.value},set(X){C(e,X,J.guard),l(J,X)},update(X){m(e,X);let j=X(J.value);C(e,j,J.guard),l(J,j)}}}function xZ(Z){return w(Z,e)}function dZ(Z,$){if(Z.length!==$.length)return!1;for(let J=0;J<Z.length;J++)if(Z[J]!==$[J])return!1;return!0}function W4(Z){let $=0,J=typeof Z==="function";return[typeof Z==="string"?()=>`${Z}${$++}`:J?(X)=>Z(X)||String($++):()=>String($++),J]}function l4(Z,$,J,X,j,B){let q={},G={},H={},W=[],V=!1,U=new Map;for(let Q=0;Q<Z.length;Q++){let M=J[Q],z=Z[Q];if(M&&z!==void 0)U.set(M,z)}let D=new Set;for(let Q=0;Q<$.length;Q++){let M=$[Q];if(M===void 0)continue;let z=j?X(M):J[Q]??X(M);if(D.has(z))throw new UZ(S,z,M);if(W.push(z),D.add(z),!U.has(z))q[z]=M,V=!0;else if(!B(U.get(z),M))G[z]=M,V=!0}for(let[Q]of U)if(!D.has(Q))H[Q]=null,V=!0;if(!V&&!dZ(J,W))V=!0;return{add:q,change:G,remove:H,newKeys:W,changed:V}}function GZ(Z,$){C(S,Z,Array.isArray);let J=new Map,X=[],[j,B]=W4($?.keyConfig),q=$?.itemEquals??s,G=$?.createItem??((Q)=>g(Q,{equals:q})),H=()=>X.map((Q)=>J.get(Q)?.get()).filter((Q)=>Q!==void 0),W={fn:H,value:Z,flags:P,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:s,error:void 0},V=(Q)=>{let M=!1;for(let z in Q.add){let K=Q.add[z];C(`${S} item for key "${z}"`,K),J.set(z,G(K)),M=!0}if(Object.keys(Q.change).length)k(()=>{for(let z in Q.change){let K=Q.change[z];C(`${S} item for key "${z}"`,K);let F=J.get(z);if(F)F.set(K)}});for(let z in Q.remove){J.delete(z);let K=X.indexOf(z);if(K!==-1)X.splice(K,1);M=!0}if(M)W.flags|=b;return Q.changed},U=i(W,$?.watched);for(let Q=0;Q<Z.length;Q++){let M=Z[Q];if(M===void 0)continue;let z=X[Q];if(!z)z=j(M),X[Q]=z;C(`${S} item for key "${z}"`,M),J.set(z,G(M))}W.value=Z,W.flags=0;let D={[Symbol.toStringTag]:S,[Symbol.isConcatSpreadable]:!0,*[Symbol.iterator](){for(let Q of X){let M=J.get(Q);if(M)yield M}},get length(){return U(),X.length},get(){if(U(),W.sources){if(W.flags){let Q=W.flags&b;if(W.value=f(H),Q){if(W.flags=P,R(W),W.error)throw W.error}else W.flags=h}}else if(R(W),W.error)throw W.error;return W.value},set(Q){let M=W.flags&P?H():W.value,z=l4(M,Q,X,j,B,q);if(z.changed){X=z.newKeys,V(z),W.flags|=P;for(let K=W.sinks;K;K=K.nextSink)Y(K.sink);if(O===0)A()}},update(Q){D.set(Q(D.get()))},at(Q){let M=X[Q];return M!==void 0?J.get(M):void 0},keys(){return U(),X.values()},byKey(Q){return J.get(Q)},keyAt(Q){return X[Q]},indexOfKey(Q){return X.indexOf(Q)},add(Q){let M=j(Q);if(J.has(M))throw new UZ(S,M,Q);if(!X.includes(M))X.push(M);C(`${S} item for key "${M}"`,Q),J.set(M,G(Q)),W.flags|=P|b;for(let z=W.sinks;z;z=z.nextSink)Y(z.sink);if(O===0)A();return M},remove(Q){let M=typeof Q==="number"?X[Q]:Q;if(M===void 0)return;if(J.delete(M)){let K=typeof Q==="number"?Q:X.indexOf(M);if(K>=0)X.splice(K,1);W.flags|=P|b;for(let F=W.sinks;F;F=F.nextSink)Y(F.sink);if(O===0)A()}},replace(Q,M){let z=J.get(Q);if(!z)return;if(C(`${S} item for key "${Q}"`,M),q(f(()=>z.get()),M))return;z.set(M),W.flags|=P;for(let K=W.sinks;K;K=K.nextSink)Y(K.sink);if(O===0)A()},sort(Q){let z=X.map((K)=>[K,J.get(K)?.get()]).sort(x(Q)?(K,F)=>Q(K[1],F[1]):(K,F)=>String(K[1]).localeCompare(String(F[1]))).map(([K])=>K);if(!dZ(X,z)){X=z,W.flags|=P;for(let K=W.sinks;K;K=K.nextSink)Y(K.sink);if(O===0)A()}},splice(Q,M,...z){let K=X.length,F=Q<0?Math.max(0,K+Q):Math.min(Q,K),CZ=Math.max(0,Math.min(M??Math.max(0,K-Math.max(0,F)),K-F)),zZ={},I={};for(let y=0;y<CZ;y++){let t=F+y,Z4=X[t];if(Z4){let D4=J.get(Z4);if(D4)I[Z4]=D4.get()}}let _=X.slice(0,F),u={};for(let y of z){let t=j(y);if(t in I)delete I[t],u[t]=y;else if(J.has(t))throw new UZ(S,t,y);else zZ[t]=y;_.push(t)}_.push(...X.slice(F+CZ));let K4=!!(Object.keys(zZ).length||Object.keys(I).length||Object.keys(u).length);if(K4){V({add:zZ,change:u,remove:I,changed:K4}),X=_,W.flags|=P;for(let y=W.sinks;y;y=y.nextSink)Y(y.sink);if(O===0)A()}return Object.values(I)},deriveCollection(Q){return uZ(D,Q)}};return D}function cZ(Z){return w(Z,S)}function v(Z,$){if(m(ZZ,Z,YZ),$?.value!==void 0)C(ZZ,$.value,$?.guard);let J={fn:Z,value:$?.value,flags:P,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:$?.equals??p,error:void 0,stop:void 0},X=$?.watched,j=i(J,X?()=>X(()=>{if(Y(J),O===0)A()}):void 0);return{[Symbol.toStringTag]:ZZ,get(){if(j(),R(J),J.error)throw J.error;return IZ(ZZ,J.value),J.value}}}function BZ(Z){return w(Z,ZZ)}function KZ(Z,$){if(m($Z,Z,n),$?.value!==void 0)C($Z,$.value,$?.guard);let J={value:!1,sinks:null,sinksTail:null,equals:p},X={fn:Z,value:$?.value,sources:null,sourcesTail:null,sinks:null,sinksTail:null,flags:P,equals:$?.equals??p,controller:void 0,error:void 0,stop:void 0,pendingNode:J},j=$?.watched,B=i(X,j?()=>j(()=>{if(Y(X),O===0)A()}):void 0),q=i(J);return{[Symbol.toStringTag]:$Z,get(){if(B(),R(X),X.error)throw X.error;return IZ($Z,X.value),X.value},isPending(){return q(),X.pendingNode.value},abort(){X.controller?.abort(),X.controller=void 0,l(X.pendingNode,!1)}}}function DZ(Z){return w(Z,$Z)}function uZ(Z,$){m(XZ,$);let J=n($),X=new Map,j=[],B=(Q)=>{let M=J?KZ(async(z,K)=>{let F=Z.byKey(Q)?.get();if(F==null)return z;return $(F,K)}):v(()=>{let z=Z.byKey(Q)?.get();if(z==null)return;return $(z)});X.set(Q,M)};function q(Q){if(!dZ(j,Q)){let M=new Set(j),z=new Set(Q);for(let K of j)if(!z.has(K))X.delete(K);for(let K of Q)if(!M.has(K))B(K);j=Q,W.flags|=b}}function G(){q(Array.from(Z.keys()));let Q=[];for(let M of j)try{let z=X.get(M)?.get();if(z!=null)Q.push(z)}catch(z){if(!(z instanceof a))throw z}return Q}let W={fn:G,value:[],flags:P,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:(Q,M)=>{if(Q.length!==M.length)return!1;for(let z=0;z<Q.length;z++)if(Q[z]!==M[z])return!1;return!0},error:void 0};function V(){if(W.sources){if(W.flags)if(W.value=f(G),W.flags&b){if(W.flags=P,R(W),W.error)throw W.error}else W.flags=h}else if(W.sinks){if(R(W),W.error)throw W.error}else W.value=f(G)}let U=Array.from(f(()=>Z.keys()));for(let Q of U)B(Q);j=U;let D={[Symbol.toStringTag]:XZ,[Symbol.isConcatSpreadable]:!0,*[Symbol.iterator](){for(let Q of j){let M=X.get(Q);if(M)yield M}},get length(){if(N)d(W,N);return V(),j.length},keys(){if(N)d(W,N);return V(),j.values()},get(){if(N)d(W,N);return V(),W.value},at(Q){let M=j[Q];return M!==void 0?X.get(M):void 0},byKey(Q){return X.get(Q)},keyAt(Q){return j[Q]},indexOfKey(Q){return j.indexOf(Q)},deriveCollection(Q){return uZ(D,Q)}};return D}function F4(Z,$){let J=$?.value??[];if(J.length)C(XZ,J,Array.isArray);m(XZ,Z,YZ);let X=new Map,j=[],B=new Map,[q,G]=W4($?.keyConfig),H=(z)=>B.get(z)??(G?q(z):void 0),W=$?.createItem??((z)=>g(z,{equals:$?.itemEquals??s}));function V(){let z=[];for(let K of j)try{let F=X.get(K)?.get();if(F!=null)z.push(F)}catch(F){if(!(F instanceof a))throw F}return z}let U={fn:V,value:J,flags:P,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:gZ,error:void 0};for(let z of J){let K=q(z);X.set(K,W(z)),B.set(z,K),j.push(K)}U.value=J,U.flags=P;let D=(z)=>{let{add:K,change:F,remove:CZ}=z;if(!K?.length&&!F?.length&&!CZ?.length)return;let zZ=!1;k(()=>{if(K)for(let I of K){let _=q(I);if(X.set(_,W(I)),B.set(I,_),!j.includes(_))j.push(_);zZ=!0}if(F)for(let I of F){let _=H(I);if(!_)continue;let u=X.get(_);if(u&&xZ(u))B.delete(u.get()),u.set(I),B.set(I,_)}if(CZ)for(let I of CZ){let _=H(I);if(!_)continue;B.delete(I),X.delete(_);let u=j.indexOf(_);if(u!==-1)j.splice(u,1);zZ=!0}U.flags=P|(zZ?b:0);for(let I=U.sinks;I;I=I.nextSink)Y(I.sink)})},Q=i(U,()=>Z(D)),M={[Symbol.toStringTag]:XZ,[Symbol.isConcatSpreadable]:!0,*[Symbol.iterator](){for(let z of j){let K=X.get(z);if(K)yield K}},get length(){return Q(),j.length},keys(){return Q(),j.values()},get(){if(Q(),U.sources){if(U.flags){let z=U.flags&b;if(U.value=f(V),z){if(U.flags=P,R(U),U.error)throw U.error}else U.flags=h}}else if(R(U),U.error)throw U.error;return U.value},at(z){let K=j[z];return K!==void 0?X.get(K):void 0},byKey(z){return X.get(z)},keyAt(z){return j[z]},indexOfKey(z){return j.indexOf(z)},deriveCollection(z){return uZ(M,z)}};return M}function C4(Z){return w(Z,XZ)}function r(Z){m("Effect",Z);let $={fn:Z,flags:P,sources:null,sourcesTail:null,cleanup:null},J=()=>{vZ($),$.fn=void 0,$.flags=h,$.sourcesTail=null,OZ($)};if(T)VZ(T,J);return j4($),J}function RZ(Z,$){if(!T)throw new bZ("match");let J=!Array.isArray(Z),X=J?[Z]:Z,{nil:j,stale:B}=$,q=J?(D)=>$.ok(D[0]):(D)=>$.ok(D),G=J&&$.err?(D)=>$.err(D[0]):$.err??console.error,H,W=!1,V=Array(X.length);for(let D=0;D<X.length;D++)try{V[D]=X[D].get()}catch(Q){if(Q instanceof a){W=!0;continue}if(!H)H=[];H.push(Q instanceof Error?Q:Error(String(Q)))}let U;try{if(W)U=j?.();else if(H)U=G(H);else if(B&&(J?DZ(X[0])&&X[0].isPending():X.some((D)=>DZ(D)&&D.isPending())))U=B();else U=q(V)}catch(D){U=G([D instanceof Error?D:Error(String(D))])}if(typeof U==="function")return U;if(U instanceof Promise){let D=T,Q=new AbortController;VZ(D,()=>Q.abort()),U.then((M)=>{if(!Q.signal.aborted&&typeof M==="function")VZ(D,M)}).catch((M)=>{G([M instanceof Error?M:Error(String(M))])})}}function Y4(Z,$){if(m(JZ,Z,YZ),$?.value!==void 0)C(JZ,$.value,$?.guard);let J={value:$?.value,sinks:null,sinksTail:null,equals:$?.equals??p,guard:$?.guard,stop:void 0};return{[Symbol.toStringTag]:JZ,get(){if(N){if(!J.sinks)J.stop=Z((X)=>{C(JZ,X,J.guard),l(J,X)});d(J,N)}return IZ(JZ,J.value),J.value}}}function I4(Z){return w(Z,JZ)}function s4(Z,$){let J={},X={},j={},B=!1,q=Object.keys(Z),G=Object.keys($);for(let H of G)if(H in Z){if(!s(Z[H],$[H]))X[H]=$[H],B=!0}else J[H]=$[H],B=!0;for(let H of q)if(!(H in $))j[H]=void 0,B=!0;return{add:J,change:X,remove:j,changed:B}}function AZ(Z,$){C(jZ,Z,L);let J=new Map,X=(W,V)=>{if(C(`${jZ} for key "${W}"`,V),Array.isArray(V))J.set(W,GZ(V));else if(L(V))J.set(W,AZ(V));else J.set(W,g(V))},j=()=>{let W={};return J.forEach((V,U)=>{W[U]=V.get()}),W},B={fn:j,value:Z,flags:P,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:s,error:void 0},q=(W)=>{let V=!1;for(let U in W.add)X(U,W.add[U]),V=!0;if(Object.keys(W.change).length)k(()=>{for(let U in W.change){let D=W.change[U];C(`${jZ} for key "${U}"`,D);let Q=J.get(U);if(Q)if(L(D)!==TZ(Q))X(U,D),V=!0;else Q.set(D)}});for(let U in W.remove)J.delete(U),V=!0;if(V)B.flags|=b;return W.changed},G=i(B,$?.watched);for(let W of Object.keys(Z))X(W,Z[W]);let H={[Symbol.toStringTag]:jZ,[Symbol.isConcatSpreadable]:!1,*[Symbol.iterator](){for(let W of Array.from(J.keys())){let V=J.get(W);if(V)yield[W,V]}},keys(){return G(),J.keys()},byKey(W){return J.get(W)},get(){if(G(),B.sources){if(B.flags){let W=B.flags&b;if(B.value=f(j),W){if(B.flags=P,R(B),B.error)throw B.error}else B.flags=h}}else if(R(B),B.error)throw B.error;return B.value},set(W){let V=B.flags&P?j():B.value,U=s4(V,W);if(q(U)){B.flags|=P;for(let D=B.sinks;D;D=D.nextSink)Y(D.sink);if(O===0)A()}},update(W){H.set(W(H.get()))},add(W,V){if(J.has(W))throw new UZ(jZ,W,V);X(W,V),B.flags|=P|b;for(let U=B.sinks;U;U=U.nextSink)Y(U.sink);if(O===0)A();return W},remove(W){if(J.delete(W)){B.flags|=P|b;for(let U=B.sinks;U;U=U.nextSink)Y(U.sink);if(O===0)A()}}};return new Proxy(H,{get(W,V){if(V in W)return Reflect.get(W,V);if(typeof V!=="symbol")return W.byKey(V)},has(W,V){if(V in W)return!0;return W.byKey(String(V))!==void 0},ownKeys(W){return Array.from(W.keys())},getOwnPropertyDescriptor(W,V){if(V in W)return Reflect.getOwnPropertyDescriptor(W,V);if(typeof V==="symbol")return;let U=W.byKey(String(V));return U?{enumerable:!0,configurable:!0,writable:!0,value:U}:void 0}})}function TZ(Z){return w(Z,jZ)}var r4=new Set([e,ZZ,$Z,JZ,WZ,S,XZ,jZ]);function wZ(Z,$){return n(Z)?KZ(Z,$):v(Z,$)}function O4(Z){if(QZ(Z))return Z;if(Z==null)throw new qZ("createSignal",Z);if(n(Z))return KZ(Z);if(x(Z))return v(Z);if($4(Z))return GZ(Z);if(L(Z))return AZ(Z);return g(Z)}function x4(Z){if(SZ(Z))return Z;if(Z==null||x(Z)||QZ(Z))throw new qZ("createMutableSignal",Z);if($4(Z))return GZ(Z);if(L(Z))return AZ(Z);return g(Z)}function R4(Z){return BZ(Z)||DZ(Z)}function QZ(Z){return Z!=null&&r4.has(Z[Symbol.toStringTag])}function SZ(Z){return xZ(Z)||TZ(Z)||cZ(Z)}function A4(Z){if(QZ(Z))return!0;return Z!==null&&typeof Z==="object"&&"get"in Z&&typeof Z.get==="function"}function iZ(Z,$){C(WZ,Z,A4);let J=Z,X=$?.guard,j={fn:()=>J.get(),value:void 0,flags:P,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:$?.equals??p,error:void 0},B=()=>{if(N)d(j,N);if(R(j),j.error)throw j.error;return j.value},q=(H)=>{if(PZ(J))return J.set(H);if("set"in J&&typeof J.set==="function")C(WZ,H,X),J.set(H);else throw new _Z(WZ)},G=(H)=>{C(WZ,H,A4),J=H,j.flags|=P;for(let W=j.sinks;W;W=W.nextSink)Y(W.sink);if(O===0)A()};return{[Symbol.toStringTag]:WZ,configurable:!0,enumerable:!0,get:B,set:q,replace:G,current:()=>J}}function PZ(Z){return w(Z,WZ)}var oZ="context-request";class B4 extends Event{context;callback;subscribe;constructor(Z,$,J=!1){super(oZ,{bubbles:!0,composed:!0});this.context=Z,this.callback=$,this.subscribe=J}}var w4=(Z)=>($)=>()=>E(()=>{let J=(X)=>{let{context:j,callback:B}=X;if(typeof j==="string"&&$.includes(j)&&x(B))X.stopImmediatePropagation(),B(()=>Z[j])};return Z.addEventListener(oZ,J),()=>Z.removeEventListener(oZ,J)}),L4=(Z)=>($,J)=>{let X=()=>J;return Z.dispatchEvent(new B4($,(j)=>{X=j})),v(X)};var NZ=typeof process<"u"&&!1,FZ="warn",t4=(Z)=>Z?`#${Z}`:"",n4=(Z)=>Z?.length?`.${Array.from(Z).join(".")}`:"",Q4=(Z)=>Z.localName.includes("-"),z4=(Z)=>Q4(Z)&&Z.matches(":not(:defined)"),o=(Z)=>Z?`<${Z.localName}${t4(Z.id)}${n4(Z.classList)}>`:"<unknown>";class lZ extends TypeError{constructor(Z){super(`Invalid component name "${Z}". Custom element names must contain a hyphen, start with a lowercase letter, and contain only lowercase letters, numbers, and hyphens.`);this.name="InvalidComponentNameError"}}class _4 extends TypeError{constructor(Z,$,J){super(`Invalid property name "${$}" for component <${Z}>. ${J}`);this.name="InvalidPropertyNameError"}}class EZ extends Error{constructor(Z,$,J){super(`Missing required element <${$}> in component ${o(Z)}. ${J}`);this.name="MissingElementError"}}class sZ extends Error{constructor(Z,$){super(`Timeout waiting for: [${$.join(", ")}] in component ${o(Z)}.`);this.name="DependencyTimeoutError"}}class rZ extends TypeError{constructor(Z,$,J){super(`Expected reactives passed from ${o(Z)} to ${o($)} to be a record of signals, reactive property names or functions. Got ${HZ(J)}.`);this.name="InvalidReactivesError"}}class tZ extends TypeError{constructor(Z,$){super(`Target ${o(Z)} is not a custom element in ${$}.`);this.name="InvalidCustomElementError"}}var b4=new WeakMap,yZ=(Z)=>{let $=b4.get(Z);if(!$)$={},b4.set(Z,$);return $};var nZ=(Z)=>{for(let $ of Z)if(Array.isArray($))nZ($);else if($)$()},H4=(Z,$)=>{if(x($))return wZ($);if(typeof $==="string"){let J=yZ(Z)[$];if(J)return J;return v(()=>Z[$])}if($&&typeof $==="object"&&"get"in $&&!(Symbol.toStringTag in $))return $;return $},f4=(Z)=>{function $(J,X){return()=>{if(Array.isArray(J)){let B=J.map((G)=>H4(Z,G)),q=X;return r(()=>RZ(B,{ok:(G)=>f(()=>q(G))}))}let j=H4(Z,J);if(typeof X==="function")return r(()=>RZ(j,{ok:(B)=>f(()=>X(B))}));return r(()=>RZ(j,X))}}return $},T4=(Z)=>{let $=(X,j)=>E(()=>{if(!Q4(X))throw new tZ(X,`pass from ${o(Z)}`);if(!L(j))throw new rZ(Z,X,j);let B=yZ(X),q=o(X),G=[];for(let[H,W]of Object.entries(j)){if(W==null)continue;if(!(H in X)){if(NZ)console[FZ](`pass(): property '${H}' does not exist on ${q}`);continue}let V=H4(Z,W);if(!V)continue;let U=B[H];if(PZ(U)){let D=U.current();U.replace(V),G.push(()=>U.replace(D));continue}if(NZ)console[FZ](`pass(): property '${H}' on ${q} is not Slot-backed — use setProperty() for non-Le Truc elements`)}if(G.length)return()=>{for(let H of G)H()}});function J(X,j){return()=>{if(!X)return;if(BZ(X))r(()=>{for(let B of X.get())E(()=>$(B,j))});else $(X,j)}}return J};function a4(Z,$){return()=>{r(()=>{for(let J of Z.get())E(()=>{let X=$(J);if(Array.isArray(X))nZ(X);else if(typeof X==="function")X()})})}}var q4=new Set,S4=new WeakMap,aZ=new Set,U4,e4=()=>{U4=void 0;let Z=Array.from(q4);q4.clear();for(let J of Z)S4.get(J)?.();let $=Array.from(aZ);aZ.clear();for(let J of $)J()},E4=()=>{if(!U4)U4=requestAnimationFrame(e4)},M4=(Z,$)=>{S4.set(Z,$),q4.add(Z),E4()},eZ=(Z,$)=>{let J=!1,X,j=()=>{J=!1,Z(...X)},B=(...q)=>{if(X=q,J)return;J=!0,aZ.add(j),E4()};return B.cancel=()=>{if(J)aZ.delete(j),J=!1},$?.addEventListener("abort",B.cancel,{once:!0}),B};var Z6=new Set(["scroll","resize","mousewheel","touchstart","touchmove","wheel"]),$6=new Set(["focus","blur","scroll","resize","load","unload","error","toggle","mouseenter","mouseleave","pointerenter","pointerleave","abort","canplay","canplaythrough","durationchange","emptied","ended","loadeddata","loadedmetadata","loadstart","pause","play","playing","progress","ratechange","seeked","seeking","stalled","suspend","timeupdate","volumechange","waiting"]),y4=(Z,$,J,X,j)=>{let B=(G)=>{let H=X(G,$);if(!L(H))return;k(()=>{for(let[W,V]of Object.entries(H))Z[W]=V})},q=j.passive?eZ(B):B;return $.addEventListener(J,q,j),()=>{$.removeEventListener(J,q),q.cancel?.()}},m4=(Z)=>{function $(J,X,j,B={}){return()=>{if(!J)return;if(!("passive"in B))B={...B,passive:Z6.has(X)};if(BZ(J)){if($6.has(X)){if(NZ)console[FZ](`on(): '${X}' does not bubble — prefer each() + on() for per-element listeners in ${o(Z)}`);return r(()=>{for(let W of J.get())E(()=>{return y4(Z,W,X,j,B)})})}let q=Z.shadowRoot??Z,G=(W)=>{let V=W.composedPath();for(let U of J.get())if(V.includes(U)){let D=j(W,U);if(!L(D))break;k(()=>{for(let[Q,M]of Object.entries(D))Z[Q]=M});break}},H=B.passive?eZ(G):G;E(()=>{return q.addEventListener(X,H,B),()=>{q.removeEventListener(X,H),H.cancel?.()}});return}E(()=>y4(Z,J,X,j,B))}}return $};var J6=200,X6=(Z)=>{let $=new Set,J="",X=0;for(let j of Z)if(j==="[")X++;else if(j==="]"){if(X>0)X--}else if(X===0)J+=j;if(J.includes("."))$.add("class");if(J.includes("#"))$.add("id");if(Z.includes("[")){let j=Z.split("[");for(let B=1;B<j.length;B++){let q=j[B];if(!q||!q.includes("]"))continue;let G=q.split("=")[0].split("]")[0].trim().replace(/[^a-zA-Z0-9_-]/g,"");if(G)$.add(G)}}return[...$]};function k4(Z,$){return v(()=>Array.from(Z.querySelectorAll($)),{value:[],equals:(J,X)=>J.length===X.length&&J.every((j,B)=>j===X[B]),watched:(J)=>{let X={childList:!0,subtree:!0},j=X6($);if(j.length)X.attributes=!0,X.attributeFilter=j;let B=(H)=>H instanceof Element&&(H.matches($)||H.querySelector($)),q=(H)=>{if(H.type==="attributes")return!0;if(H.type==="childList")return Array.from(H.addedNodes).some(B)||Array.from(H.removedNodes).some(B);return!1},G=new MutationObserver((H)=>{for(let W of H)if(q(W)){J();return}});return G.observe(Z,X),()=>G.disconnect()}})}var h4=(Z)=>{let $=Z.shadowRoot??Z,J=new Set;function X(q,G){let H=$.querySelector(q);if(G!=null&&!H)throw new EZ(Z,q,G);if(H&&z4(H))J.add(H.localName);return H??void 0}function j(q,G){let H=k4($,q),W=H.get();if(G!=null&&!W.length)throw new EZ(Z,q,G);if(W.length){for(let V of W)if(z4(V))J.add(V.localName)}return H}return[{first:X,all:j},(q)=>{if(J.size)queueMicrotask(()=>{let G=Array.from(J).filter((H)=>!customElements.get(H));if(!G.length){q();return}Promise.race([Promise.all(G.map((H)=>customElements.whenDefined(H))),new Promise((H,W)=>{setTimeout(()=>{W(new sZ(Z,G.filter((V)=>!customElements.get(V))))},J6)})]).then(q).catch((H)=>{if(NZ)console[FZ](H);q()})});else q()}]};var p4=Symbol("parser"),g4=Symbol("method"),v4=(Z)=>x(Z)&&(p4 in Z),d4=(Z)=>x(Z)&&(g4 in Z),c=(Z)=>Object.assign(Z,{[p4]:!0}),j6=(Z)=>Object.assign(Z,{[g4]:!0});function W6(Z,$){if(!Z.includes("-")||!Z.match(/^[a-z][a-z0-9-]*$/))throw new lZ(Z);class J extends HTMLElement{debug;#J=!1;#Z=[];#$;connectedCallback(){let X=()=>{this.#$=E(()=>nZ(this.#Z),{root:!0})};if(this.#J)X();else{let j=this,[B,q]=h4(this),G={expose:this.#j.bind(this),host:j,...B,watch:f4(j),on:m4(j),pass:T4(j),provideContexts:w4(j),requestContext:L4(j)},H=$(G);if(H)this.#Z=H;if(this.#J=!0,!this.#Z.length)return;q(X)}}disconnectedCallback(){if(x(this.#$))this.#$()}#j(X){let j=(B,q)=>{if(v4(q)){let G=q(this.getAttribute(B));if(G!=null)this.#X(B,G)}else if(d4(q))this[B]=q;else{let G=q;if(G!=null)this.#X(B,G)}};for(let[B,q]of Object.entries(X)){if(q==null||B in this)continue;j(B,q)}}#X(X,j){let B=QZ(j)?j:x(j)?wZ(j):g(j),q=yZ(this),G=X,H=q[G];if(PZ(H))H.replace(B);else if(SZ(B)){let W=iZ(B);q[G]=W,Object.defineProperty(this,X,W)}else q[G]=B,Object.defineProperty(this,X,{get:B.get,enumerable:!0})}}return customElements.define(Z,J),customElements.get(Z)}var B6=(Z)=>{if(/^(javascript|data|vbscript):/i.test(Z))return!1;if(/^(mailto|tel):/i.test(Z))return!0;if(Z.includes("://"))try{let $=new URL(Z);return["http:","https:","ftp:"].includes($.protocol)}catch{return!1}return!0},V4=(Z,$,J)=>{if(/^on/i.test($))throw Error(`setAttribute: blocked unsafe attribute name '${$}' on ${Z.localName} — event handler attributes are not allowed`);if(J=String(J).trim(),!B6(J))throw Error(`setAttribute: blocked unsafe value for '${$}' on <${Z.localName}>: '${J}'`);Z.setAttribute($,J)},Q6=(Z)=>Z.replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;").replace(/"/g,"&quot;").replace(/'/g,"&#39;"),G4=(Z,$)=>{Array.from(Z.childNodes).filter((J)=>J.nodeType!==Node.COMMENT_NODE).forEach((J)=>{J.remove()}),Z.append(document.createTextNode($))};var z6=["type","src","async","defer","nomodule","crossorigin","integrity","referrerpolicy","fetchpriority"],H6=(Z,$=!1)=>$?(J)=>G4(Z,String(J)):(J)=>{Z.textContent=String(J)},q6=(Z,$)=>(J)=>{Z[$]=J},U6=(Z,$)=>(J)=>{Z.classList.toggle($,Boolean(J))},M6=(Z)=>($)=>{Z.hidden=!$},V6=(Z,$,J=!1)=>({ok:(X)=>{if(typeof X==="boolean")Z.toggleAttribute($,X);else if(J)Z.setAttribute($,X);else V4(Z,$,X)},nil:()=>{Z.removeAttribute($)}}),G6=(Z,$)=>({ok:(J)=>{Z.style.setProperty($,J)},nil:()=>{Z.style.removeProperty($)}}),K6=(Z,$={})=>{let J=()=>{if(Z.shadowRoot)Z.shadowRoot.innerHTML="<slot></slot>";else Z.innerHTML=""};return{ok:(X)=>{if(!X){J();return}let{shadowRootMode:j,allowScripts:B}=$;if(j&&!Z.shadowRoot)Z.attachShadow({mode:j});let q=Z.shadowRoot||Z;M4(Z,()=>{if(q.innerHTML=X,B)q.querySelectorAll("script").forEach((G)=>{let H=document.createElement("script");for(let W of z6){let V=G.getAttribute(W);if(V!==null)H.setAttribute(W,V)}if(!G.hasAttribute("src"))H.appendChild(document.createTextNode(G.textContent??""));q.appendChild(H),G.remove()})})},nil:J}};var D6=()=>c((Z)=>Z!=null&&Z!=="false");var P6=(Z)=>c(($)=>{if(($??Z)==null)throw TypeError("asJSON: Value and fallback are both null or undefined");if($==null)return Z;if($==="")throw SyntaxError("Empty string is not valid JSON");let J;try{J=JSON.parse($)}catch(X){throw SyntaxError(`Failed to parse JSON: ${String(X)}`,{cause:X})}return J??Z});var mZ=(Z,$)=>{if($==null)return;let J=Z($);return Number.isFinite(J)?J:void 0},N6=(Z=0)=>c(($)=>{if($==null)return Z;let J=$.trim();if(J.toLowerCase().startsWith("0x"))return mZ((j)=>parseInt(j,16),J)??Z;let X=mZ(parseFloat,$);return X!=null?Math.trunc(X):Z}),F6=(Z=0)=>c(($)=>mZ(parseFloat,$)??Z),C6=(Z=0,$=Number.MAX_SAFE_INTEGER)=>c((J)=>{if(J==null)return Z;let X=J.trim(),j=X.toLowerCase().startsWith("0x")?mZ((q)=>parseInt(q,16),X):mZ(parseFloat,J),B=j!=null?Math.trunc(j):Z;return Math.max(Z,Math.min(B,$))});var Y6=(Z="")=>c(($)=>$??Z),I6=(Z)=>c(($)=>{if($==null)return Z[0];let J=$.toLowerCase();return Z.find((j)=>j.toLowerCase()===J)??Z[0]});export{HZ as valueString,f as untrack,N4 as unown,eZ as throttle,G4 as setTextPreservingComments,M4 as schedule,V4 as safeSetAttribute,RZ as match,DZ as isTask,TZ as isStore,xZ as isState,PZ as isSlot,w as isSignalOfType,QZ as isSignal,I4 as isSensor,L as isRecord,v4 as isParser,SZ as isMutableSignal,d4 as isMethodProducer,BZ as isMemo,cZ as isList,x as isFunction,R4 as isComputed,C4 as isCollection,n as isAsyncFunction,Q6 as escapeHTML,a4 as each,j6 as defineMethod,W6 as defineComponent,K6 as dangerouslyBindInnerHTML,KZ as createTask,AZ as createStore,g as createState,iZ as createSlot,O4 as createSignal,Y4 as createSensor,E as createScope,x4 as createMutableSignal,v as createMemo,GZ as createList,k4 as createElementsMemo,r as createEffect,wZ as createComputed,F4 as createCollection,M6 as bindVisible,H6 as bindText,G6 as bindStyle,q6 as bindProperty,U6 as bindClass,V6 as bindAttribute,k as batch,Y6 as asString,c as asParser,F6 as asNumber,P6 as asJSON,N6 as asInteger,I6 as asEnum,C6 as asClampedInteger,D6 as asBoolean,a as UnsetSignalValueError,gZ as SKIP_EQUALITY,bZ as RequiredOwnerError,_Z as ReadonlySignalError,kZ as NullishSignalValueError,EZ as MissingElementError,qZ as InvalidSignalValueError,rZ as InvalidReactivesError,_4 as InvalidPropertyNameError,tZ as InvalidCustomElementError,lZ as InvalidComponentNameError,hZ as InvalidCallbackError,sZ as DependencyTimeoutError,p as DEFAULT_EQUALITY,s as DEEP_EQUALITY,B4 as ContextRequestEvent,LZ as CircularDependencyError,oZ as CONTEXT_REQUEST};
+var K4=Object.getPrototypeOf(async()=>{});function O(Z){return typeof Z==="function"}function n(Z){return O(Z)&&Object.getPrototypeOf(Z)===K4}function YZ(Z){return O(Z)&&Object.getPrototypeOf(Z)!==K4}function w(Z,$){return Z!=null&&Z[Symbol.toStringTag]===$}function L(Z){return Z!==null&&typeof Z==="object"&&Object.getPrototypeOf(Z)===Object.prototype}function $4(Z,$=(J)=>J!=null){return Array.isArray(Z)&&Z.every($)}function zZ(Z){return typeof Z==="string"?`"${Z}"`:!!Z&&typeof Z==="object"?JSON.stringify(Z):String(Z)}class LZ extends Error{constructor(Z){super(`[${Z}] Circular dependency detected`);this.name="CircularDependencyError"}}class hZ extends TypeError{constructor(Z){super(`[${Z}] Signal value cannot be null or undefined`);this.name="NullishSignalValueError"}}class a extends Error{constructor(Z){super(`[${Z}] Signal value is unset`);this.name="UnsetSignalValueError"}}class qZ extends TypeError{constructor(Z,$){super(`[${Z}] Signal value ${zZ($)} is invalid`);this.name="InvalidSignalValueError"}}class kZ extends TypeError{constructor(Z,$){super(`[${Z}] Callback ${zZ($)} is invalid`);this.name="InvalidCallbackError"}}class _Z extends Error{constructor(Z){super(`[${Z}] Signal is read-only`);this.name="ReadonlySignalError"}}class bZ extends Error{constructor(Z){super(`[${Z}] Active owner is required`);this.name="RequiredOwnerError"}}class UZ extends Error{constructor(Z,$,J){super(`[${Z}] Could not add key "${$}"${J?` with value ${JSON.stringify(J)}`:""} because it already exists`);this.name="DuplicateKeyError"}}function C(Z,$,J){if($==null)throw new hZ(Z);if(J&&!J($))throw new qZ(Z,$)}function xZ(Z,$){if($==null)throw new a(Z)}function m(Z,$,J=O){if(!J($))throw new kZ(Z,$)}var e="State",ZZ="Memo",$Z="Task",JZ="Sensor",S="List",XZ="Collection",jZ="Store",WZ="Slot",k=0,MZ=1,P=2,fZ=4,b=8,N=null,T=null,pZ=[],I=0,J4=!1,p=(Z,$)=>Z===$,gZ=(Z,$)=>!1,X4=(Z,$)=>{if(Object.is(Z,$))return!0;if(typeof Z!==typeof $)return!1;if(Z==null||typeof Z!=="object"||$==null||typeof $!=="object")return!1;let J=Array.isArray(Z);if(J!==Array.isArray($))return!1;if(J){let X=Z,j=$;if(X.length!==j.length)return!1;for(let B=0;B<X.length;B++)if(!X4(X[B],j[B]))return!1;return!0}if(L(Z)&&L($)){let X=Object.keys(Z);if(X.length!==Object.keys($).length)return!1;for(let j of X){if(!(j in $))return!1;if(!X4(Z[j],$[j]))return!1}return!0}return!1},s=(Z,$)=>X4(Z,$);function c4(Z,$){let J=$.sourcesTail;if(J){let X=$.sources;while(X){if(X===Z)return!0;if(X===J)break;X=X.nextSource}}return!1}function d(Z,$){let J=$.sourcesTail;if(J?.source===Z)return;let X=null,j=$.flags&fZ;if(j){if(X=J?J.nextSource:$.sources,X?.source===Z){$.sourcesTail=X;return}}let B=Z.sinksTail;if(B?.sink===$&&(!j||c4(B,$)))return;let q={source:Z,sink:$,nextSource:X,prevSink:B,nextSink:null};if($.sourcesTail=Z.sinksTail=q,J)J.nextSource=q;else $.sources=q;if(B)B.nextSink=q;else Z.sinks=q}function u4(Z){let{source:$,nextSource:J,nextSink:X,prevSink:j}=Z;if(X)X.prevSink=j;else $.sinksTail=j;if(j)j.nextSink=X;else $.sinks=X;if(!$.sinks){if($.stop)$.stop(),$.stop=void 0;if("sources"in $&&$.sources){let B=$;B.sourcesTail=null,IZ(B),B.flags|=P}}return J}function IZ(Z){let $=Z.sourcesTail,J=$?$.nextSource:Z.sources;while(J)J=u4(J);if($)$.nextSource=null;else Z.sources=null}function Y(Z,$=P){let J=Z.flags;if("sinks"in Z){if((J&(P|MZ))>=$)return;if(Z.flags=J|$,"controller"in Z&&Z.controller)Z.controller.abort(),Z.controller=void 0;for(let X=Z.sinks;X;X=X.nextSink)Y(X.sink,MZ)}else{if((J&(P|MZ))>=$)return;let X=J&(P|MZ);if(Z.flags=$,!X)pZ.push(Z)}}function l(Z,$){if(Z.equals(Z.value,$))return;Z.value=$;for(let J=Z.sinks;J;J=J.nextSink)Y(J.sink);if(I===0)A()}function VZ(Z,$){if(!Z.cleanup)Z.cleanup=$;else if(Array.isArray(Z.cleanup))Z.cleanup.push($);else Z.cleanup=[Z.cleanup,$]}function vZ(Z){if(!Z.cleanup)return;if(Array.isArray(Z.cleanup))for(let $=0;$<Z.cleanup.length;$++)Z.cleanup[$]();else Z.cleanup();Z.cleanup=null}function i4(Z){let $=N;N=Z,Z.sourcesTail=null,Z.flags=fZ;let J=!1;try{let X=Z.fn(Z.value);if(Z.error||!Z.equals(X,Z.value))Z.value=X,Z.error=void 0,J=!0}catch(X){J=!0,Z.error=X instanceof Error?X:Error(String(X))}finally{N=$,IZ(Z)}if(J){for(let X=Z.sinks;X;X=X.nextSink)if(X.sink.flags&MZ)X.sink.flags|=P}Z.flags=k}function o4(Z){Z.controller?.abort();let $=new AbortController;Z.controller=$,Z.error=void 0;let J=N;N=Z,Z.sourcesTail=null,Z.flags=fZ;let X;try{X=Z.fn(Z.value,$.signal)}catch(j){Z.controller=void 0,Z.error=j instanceof Error?j:Error(String(j));return}finally{N=J,IZ(Z)}l(Z.pendingNode,!0),X.then((j)=>{if($.signal.aborted)return;Z.controller=void 0,h(()=>{if(Z.error||!Z.equals(j,Z.value)){Z.value=j,Z.error=void 0;for(let B=Z.sinks;B;B=B.nextSink)Y(B.sink)}l(Z.pendingNode,!1)})},(j)=>{if($.signal.aborted)return;Z.controller=void 0;let B=j instanceof Error?j:Error(String(j));h(()=>{if(!Z.error||B.name!==Z.error.name||B.message!==Z.error.message){Z.error=B;for(let q=Z.sinks;q;q=q.nextSink)Y(q.sink)}l(Z.pendingNode,!1)})}),Z.flags=k}function j4(Z){vZ(Z);let $=N,J=T;N=T=Z,Z.sourcesTail=null,Z.flags=fZ;try{let X=Z.fn();if(typeof X==="function")VZ(Z,X)}finally{N=$,T=J,IZ(Z)}Z.flags=k}function R(Z){if(Z.flags&MZ)for(let $=Z.sources;$;$=$.nextSource){if("fn"in $.source)R($.source);if(Z.flags&P)break}if(Z.flags&fZ)throw new LZ("controller"in Z?$Z:("value"in Z)?ZZ:"Effect");if(Z.flags&P)if("controller"in Z)o4(Z);else if("value"in Z)i4(Z);else j4(Z);else Z.flags=k}function A(){if(J4)return;J4=!0;try{for(let Z=0;Z<pZ.length;Z++){let $=pZ[Z];if($.flags&(P|MZ))R($)}pZ.length=0}finally{J4=!1}}function h(Z){I++;try{Z()}finally{if(I--,I===0)A()}}function f(Z){let $=N;N=null;try{return Z()}finally{N=$}}function E(Z,$){let J=T,X={cleanup:null};T=X;let j=()=>vZ(X);try{let B=Z();if(typeof B==="function")VZ(X,B);return j}finally{if(T=J,!$?.root&&J)VZ(J,j)}}function D4(Z){let $=T;T=null;try{return Z()}finally{T=$}}function i(Z,$){return $?()=>{if(N){if(!Z.sinks)Z.stop=$();d(Z,N)}}:()=>{if(N)d(Z,N)}}function g(Z,$){C(e,Z,$?.guard);let J={value:Z,sinks:null,sinksTail:null,equals:$?.equals??p,guard:$?.guard};return{[Symbol.toStringTag]:e,get(){if(N)d(J,N);return J.value},set(X){C(e,X,J.guard),l(J,X)},update(X){m(e,X);let j=X(J.value);C(e,j,J.guard),l(J,j)}}}function OZ(Z){return w(Z,e)}function dZ(Z,$){if(Z.length!==$.length)return!1;for(let J=0;J<Z.length;J++)if(Z[J]!==$[J])return!1;return!0}function W4(Z){let $=0,J=typeof Z==="function";return[typeof Z==="string"?()=>`${Z}${$++}`:J?(X)=>Z(X)||String($++):()=>String($++),J]}function l4(Z,$,J,X,j,B){let q={},G={},z={},W=[],V=!1,U=new Map;for(let Q=0;Q<Z.length;Q++){let M=J[Q],H=Z[Q];if(M&&H!==void 0)U.set(M,H)}let D=new Set;for(let Q=0;Q<$.length;Q++){let M=$[Q];if(M===void 0)continue;let H=j?X(M):J[Q]??X(M);if(D.has(H))throw new UZ(S,H,M);if(W.push(H),D.add(H),!U.has(H))q[H]=M,V=!0;else if(!B(U.get(H),M))G[H]=M,V=!0}for(let[Q]of U)if(!D.has(Q))z[Q]=null,V=!0;if(!V&&!dZ(J,W))V=!0;return{add:q,change:G,remove:z,newKeys:W,changed:V}}function GZ(Z,$){C(S,Z,Array.isArray);let J=new Map,X=[],[j,B]=W4($?.keyConfig),q=$?.itemEquals??s,G=$?.createItem??((Q)=>g(Q,{equals:q})),z=()=>X.map((Q)=>J.get(Q)?.get()).filter((Q)=>Q!==void 0),W={fn:z,value:Z,flags:P,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:s,error:void 0},V=(Q)=>{let M=!1;for(let H in Q.add){let K=Q.add[H];C(`${S} item for key "${H}"`,K),J.set(H,G(K)),M=!0}if(Object.keys(Q.change).length)h(()=>{for(let H in Q.change){let K=Q.change[H];C(`${S} item for key "${H}"`,K);let F=J.get(H);if(F)F.set(K)}});for(let H in Q.remove){J.delete(H);let K=X.indexOf(H);if(K!==-1)X.splice(K,1);M=!0}if(M)W.flags|=b;return Q.changed},U=i(W,$?.watched);for(let Q=0;Q<Z.length;Q++){let M=Z[Q];if(M===void 0)continue;let H=X[Q];if(!H)H=j(M),X[Q]=H;C(`${S} item for key "${H}"`,M),J.set(H,G(M))}W.value=Z,W.flags=0;let D={[Symbol.toStringTag]:S,[Symbol.isConcatSpreadable]:!0,*[Symbol.iterator](){for(let Q of X){let M=J.get(Q);if(M)yield M}},get length(){return U(),X.length},get(){if(U(),W.sources){if(W.flags){let Q=W.flags&b;if(W.value=f(z),Q){if(W.flags=P,R(W),W.error)throw W.error}else W.flags=k}}else if(R(W),W.error)throw W.error;return W.value},set(Q){let M=W.flags&P?z():W.value,H=l4(M,Q,X,j,B,q);if(H.changed){X=H.newKeys,V(H),W.flags|=P;for(let K=W.sinks;K;K=K.nextSink)Y(K.sink);if(I===0)A()}},update(Q){D.set(Q(D.get()))},at(Q){let M=X[Q];return M!==void 0?J.get(M):void 0},keys(){return U(),X.values()},byKey(Q){return J.get(Q)},keyAt(Q){return X[Q]},indexOfKey(Q){return X.indexOf(Q)},add(Q){let M=j(Q);if(J.has(M))throw new UZ(S,M,Q);if(!X.includes(M))X.push(M);C(`${S} item for key "${M}"`,Q),J.set(M,G(Q)),W.flags|=P|b;for(let H=W.sinks;H;H=H.nextSink)Y(H.sink);if(I===0)A();return M},remove(Q){let M=typeof Q==="number"?X[Q]:Q;if(M===void 0)return;if(J.delete(M)){let K=typeof Q==="number"?Q:X.indexOf(M);if(K>=0)X.splice(K,1);W.flags|=P|b;for(let F=W.sinks;F;F=F.nextSink)Y(F.sink);if(I===0)A()}},replace(Q,M){let H=J.get(Q);if(!H)return;if(C(`${S} item for key "${Q}"`,M),q(f(()=>H.get()),M))return;H.set(M),W.flags|=P;for(let K=W.sinks;K;K=K.nextSink)Y(K.sink);if(I===0)A()},sort(Q){let H=X.map((K)=>[K,J.get(K)?.get()]).sort(O(Q)?(K,F)=>Q(K[1],F[1]):(K,F)=>String(K[1]).localeCompare(String(F[1]))).map(([K])=>K);if(!dZ(X,H)){X=H,W.flags|=P;for(let K=W.sinks;K;K=K.nextSink)Y(K.sink);if(I===0)A()}},splice(Q,M,...H){let K=X.length,F=Q<0?Math.max(0,K+Q):Math.min(Q,K),CZ=Math.max(0,Math.min(M??Math.max(0,K-Math.max(0,F)),K-F)),HZ={},x={};for(let y=0;y<CZ;y++){let t=F+y,Z4=X[t];if(Z4){let G4=J.get(Z4);if(G4)x[Z4]=G4.get()}}let _=X.slice(0,F),u={};for(let y of H){let t=j(y);if(t in x)delete x[t],u[t]=y;else if(J.has(t))throw new UZ(S,t,y);else HZ[t]=y;_.push(t)}_.push(...X.slice(F+CZ));let V4=!!(Object.keys(HZ).length||Object.keys(x).length||Object.keys(u).length);if(V4){V({add:HZ,change:u,remove:x,changed:V4}),X=_,W.flags|=P;for(let y=W.sinks;y;y=y.nextSink)Y(y.sink);if(I===0)A()}return Object.values(x)},deriveCollection(Q){return uZ(D,Q)}};return D}function cZ(Z){return w(Z,S)}function v(Z,$){if(m(ZZ,Z,YZ),$?.value!==void 0)C(ZZ,$.value,$?.guard);let J={fn:Z,value:$?.value,flags:P,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:$?.equals??p,error:void 0,stop:void 0},X=$?.watched,j=i(J,X?()=>X(()=>{if(Y(J),I===0)A()}):void 0);return{[Symbol.toStringTag]:ZZ,get(){if(j(),R(J),J.error)throw J.error;return xZ(ZZ,J.value),J.value}}}function BZ(Z){return w(Z,ZZ)}function KZ(Z,$){if(m($Z,Z,n),$?.value!==void 0)C($Z,$.value,$?.guard);let J={value:!1,sinks:null,sinksTail:null,equals:p},X={fn:Z,value:$?.value,sources:null,sourcesTail:null,sinks:null,sinksTail:null,flags:P,equals:$?.equals??p,controller:void 0,error:void 0,stop:void 0,pendingNode:J},j=$?.watched,B=i(X,j?()=>j(()=>{if(Y(X),I===0)A()}):void 0),q=i(J);return{[Symbol.toStringTag]:$Z,get(){if(B(),R(X),X.error)throw X.error;return xZ($Z,X.value),X.value},isPending(){return q(),X.pendingNode.value},abort(){X.controller?.abort(),X.controller=void 0,l(X.pendingNode,!1)}}}function DZ(Z){return w(Z,$Z)}function uZ(Z,$){m(XZ,$);let J=n($),X=new Map,j=[],B=(Q)=>{let M=J?KZ(async(H,K)=>{let F=Z.byKey(Q)?.get();if(F==null)return H;return $(F,K)}):v(()=>{let H=Z.byKey(Q)?.get();if(H==null)return;return $(H)});X.set(Q,M)};function q(Q){if(!dZ(j,Q)){let M=new Set(j),H=new Set(Q);for(let K of j)if(!H.has(K))X.delete(K);for(let K of Q)if(!M.has(K))B(K);j=Q,W.flags|=b}}function G(){q(Array.from(Z.keys()));let Q=[];for(let M of j)try{let H=X.get(M)?.get();if(H!=null)Q.push(H)}catch(H){if(!(H instanceof a))throw H}return Q}let W={fn:G,value:[],flags:P,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:(Q,M)=>{if(Q.length!==M.length)return!1;for(let H=0;H<Q.length;H++)if(Q[H]!==M[H])return!1;return!0},error:void 0};function V(){if(W.sources){if(W.flags)if(W.value=f(G),W.flags&b){if(W.flags=P,R(W),W.error)throw W.error}else W.flags=k}else if(W.sinks){if(R(W),W.error)throw W.error}else W.value=f(G)}let U=Array.from(f(()=>Z.keys()));for(let Q of U)B(Q);j=U;let D={[Symbol.toStringTag]:XZ,[Symbol.isConcatSpreadable]:!0,*[Symbol.iterator](){for(let Q of j){let M=X.get(Q);if(M)yield M}},get length(){if(N)d(W,N);return V(),j.length},keys(){if(N)d(W,N);return V(),j.values()},get(){if(N)d(W,N);return V(),W.value},at(Q){let M=j[Q];return M!==void 0?X.get(M):void 0},byKey(Q){return X.get(Q)},keyAt(Q){return j[Q]},indexOfKey(Q){return j.indexOf(Q)},deriveCollection(Q){return uZ(D,Q)}};return D}function P4(Z,$){let J=$?.value??[];if(J.length)C(XZ,J,Array.isArray);m(XZ,Z,YZ);let X=new Map,j=[],B=new Map,[q,G]=W4($?.keyConfig),z=(H)=>B.get(H)??(G?q(H):void 0),W=$?.createItem??((H)=>g(H,{equals:$?.itemEquals??s}));function V(){let H=[];for(let K of j)try{let F=X.get(K)?.get();if(F!=null)H.push(F)}catch(F){if(!(F instanceof a))throw F}return H}let U={fn:V,value:J,flags:P,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:gZ,error:void 0};for(let H of J){let K=q(H);X.set(K,W(H)),B.set(H,K),j.push(K)}U.value=J,U.flags=P;let D=(H)=>{let{add:K,change:F,remove:CZ}=H;if(!K?.length&&!F?.length&&!CZ?.length)return;let HZ=!1;h(()=>{if(K)for(let x of K){let _=q(x);if(X.set(_,W(x)),B.set(x,_),!j.includes(_))j.push(_);HZ=!0}if(F)for(let x of F){let _=z(x);if(!_)continue;let u=X.get(_);if(u&&OZ(u))B.delete(u.get()),u.set(x),B.set(x,_)}if(CZ)for(let x of CZ){let _=z(x);if(!_)continue;B.delete(x),X.delete(_);let u=j.indexOf(_);if(u!==-1)j.splice(u,1);HZ=!0}U.flags=P|(HZ?b:0);for(let x=U.sinks;x;x=x.nextSink)Y(x.sink)})},Q=i(U,()=>Z(D)),M={[Symbol.toStringTag]:XZ,[Symbol.isConcatSpreadable]:!0,*[Symbol.iterator](){for(let H of j){let K=X.get(H);if(K)yield K}},get length(){return Q(),j.length},keys(){return Q(),j.values()},get(){if(Q(),U.sources){if(U.flags){let H=U.flags&b;if(U.value=f(V),H){if(U.flags=P,R(U),U.error)throw U.error}else U.flags=k}}else if(R(U),U.error)throw U.error;return U.value},at(H){let K=j[H];return K!==void 0?X.get(K):void 0},byKey(H){return X.get(H)},keyAt(H){return j[H]},indexOfKey(H){return j.indexOf(H)},deriveCollection(H){return uZ(M,H)}};return M}function N4(Z){return w(Z,XZ)}function r(Z){m("Effect",Z);let $={fn:Z,flags:P,sources:null,sourcesTail:null,cleanup:null},J=()=>{vZ($),$.fn=void 0,$.flags=k,$.sourcesTail=null,IZ($)};if(T)VZ(T,J);return j4($),J}function RZ(Z,$){if(!T)throw new bZ("match");let J=!Array.isArray(Z),X=J?[Z]:Z,{nil:j,stale:B}=$,q=J?(D)=>$.ok(D[0]):(D)=>$.ok(D),G=J&&$.err?(D)=>$.err(D[0]):$.err??console.error,z,W=!1,V=Array(X.length);for(let D=0;D<X.length;D++)try{V[D]=X[D].get()}catch(Q){if(Q instanceof a){W=!0;continue}if(!z)z=[];z.push(Q instanceof Error?Q:Error(String(Q)))}let U;try{if(W)U=j?.();else if(z)U=G(z);else if(B&&(J?DZ(X[0])&&X[0].isPending():X.some((D)=>DZ(D)&&D.isPending())))U=B();else U=q(V)}catch(D){U=G([D instanceof Error?D:Error(String(D))])}if(typeof U==="function")return U;if(U instanceof Promise){let D=T,Q=new AbortController;VZ(D,()=>Q.abort()),U.then((M)=>{if(!Q.signal.aborted&&typeof M==="function")VZ(D,M)}).catch((M)=>{G([M instanceof Error?M:Error(String(M))])})}}function F4(Z,$){if(m(JZ,Z,YZ),$?.value!==void 0)C(JZ,$.value,$?.guard);let J={value:$?.value,sinks:null,sinksTail:null,equals:$?.equals??p,guard:$?.guard,stop:void 0};return{[Symbol.toStringTag]:JZ,get(){if(N){if(!J.sinks)J.stop=Z((X)=>{C(JZ,X,J.guard),l(J,X)});d(J,N)}return xZ(JZ,J.value),J.value}}}function C4(Z){return w(Z,JZ)}function s4(Z,$){let J={},X={},j={},B=!1,q=Object.keys(Z),G=Object.keys($);for(let z of G)if(z in Z){if(!s(Z[z],$[z]))X[z]=$[z],B=!0}else J[z]=$[z],B=!0;for(let z of q)if(!(z in $))j[z]=void 0,B=!0;return{add:J,change:X,remove:j,changed:B}}function AZ(Z,$){C(jZ,Z,L);let J=new Map,X=(W,V)=>{if(C(`${jZ} for key "${W}"`,V),Array.isArray(V))J.set(W,GZ(V));else if(L(V))J.set(W,AZ(V));else J.set(W,g(V))},j=()=>{let W={};return J.forEach((V,U)=>{W[U]=V.get()}),W},B={fn:j,value:Z,flags:P,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:s,error:void 0},q=(W)=>{let V=!1;for(let U in W.add)X(U,W.add[U]),V=!0;if(Object.keys(W.change).length)h(()=>{for(let U in W.change){let D=W.change[U];C(`${jZ} for key "${U}"`,D);let Q=J.get(U);if(Q)if(L(D)!==TZ(Q))X(U,D),V=!0;else Q.set(D)}});for(let U in W.remove)J.delete(U),V=!0;if(V)B.flags|=b;return W.changed},G=i(B,$?.watched);for(let W of Object.keys(Z))X(W,Z[W]);let z={[Symbol.toStringTag]:jZ,[Symbol.isConcatSpreadable]:!1,*[Symbol.iterator](){for(let W of Array.from(J.keys())){let V=J.get(W);if(V)yield[W,V]}},keys(){return G(),J.keys()},byKey(W){return J.get(W)},get(){if(G(),B.sources){if(B.flags){let W=B.flags&b;if(B.value=f(j),W){if(B.flags=P,R(B),B.error)throw B.error}else B.flags=k}}else if(R(B),B.error)throw B.error;return B.value},set(W){let V=B.flags&P?j():B.value,U=s4(V,W);if(q(U)){B.flags|=P;for(let D=B.sinks;D;D=D.nextSink)Y(D.sink);if(I===0)A()}},update(W){z.set(W(z.get()))},add(W,V){if(J.has(W))throw new UZ(jZ,W,V);X(W,V),B.flags|=P|b;for(let U=B.sinks;U;U=U.nextSink)Y(U.sink);if(I===0)A();return W},remove(W){if(J.delete(W)){B.flags|=P|b;for(let U=B.sinks;U;U=U.nextSink)Y(U.sink);if(I===0)A()}}};return new Proxy(z,{get(W,V){if(V in W)return Reflect.get(W,V);if(typeof V!=="symbol")return W.byKey(V)},has(W,V){if(V in W)return!0;return W.byKey(String(V))!==void 0},ownKeys(W){return Array.from(W.keys())},getOwnPropertyDescriptor(W,V){if(V in W)return Reflect.getOwnPropertyDescriptor(W,V);if(typeof V==="symbol")return;let U=W.byKey(String(V));return U?{enumerable:!0,configurable:!0,writable:!0,value:U}:void 0}})}function TZ(Z){return w(Z,jZ)}var r4=new Set([e,ZZ,$Z,JZ,WZ,S,XZ,jZ]);function wZ(Z,$){return n(Z)?KZ(Z,$):v(Z,$)}function Y4(Z){if(QZ(Z))return Z;if(Z==null)throw new qZ("createSignal",Z);if(n(Z))return KZ(Z);if(O(Z))return v(Z);if($4(Z))return GZ(Z);if(L(Z))return AZ(Z);return g(Z)}function x4(Z){if(SZ(Z))return Z;if(Z==null||O(Z)||QZ(Z))throw new qZ("createMutableSignal",Z);if($4(Z))return GZ(Z);if(L(Z))return AZ(Z);return g(Z)}function I4(Z){return BZ(Z)||DZ(Z)}function QZ(Z){return Z!=null&&r4.has(Z[Symbol.toStringTag])}function SZ(Z){return OZ(Z)||TZ(Z)||cZ(Z)}function O4(Z){if(QZ(Z))return!0;return Z!==null&&typeof Z==="object"&&"get"in Z&&typeof Z.get==="function"}function iZ(Z,$){C(WZ,Z,O4);let J=Z,X=$?.guard,j={fn:()=>J.get(),value:void 0,flags:P,sources:null,sourcesTail:null,sinks:null,sinksTail:null,equals:$?.equals??p,error:void 0},B=()=>{if(N)d(j,N);if(R(j),j.error)throw j.error;return j.value},q=(z)=>{if(PZ(J))return J.set(z);if("set"in J&&typeof J.set==="function")C(WZ,z,X),J.set(z);else throw new _Z(WZ)},G=(z)=>{C(WZ,z,O4),J=z,j.flags|=P;for(let W=j.sinks;W;W=W.nextSink)Y(W.sink);if(I===0)A()};return{[Symbol.toStringTag]:WZ,configurable:!0,enumerable:!0,get:B,set:q,replace:G,current:()=>J}}function PZ(Z){return w(Z,WZ)}var oZ="context-request";class B4 extends Event{context;callback;subscribe;constructor(Z,$,J=!1){super(oZ,{bubbles:!0,composed:!0});this.context=Z,this.callback=$,this.subscribe=J}}var R4=(Z)=>($)=>()=>E(()=>{let J=(X)=>{let{context:j,callback:B}=X;if(typeof j==="string"&&$.includes(j)&&O(B))X.stopImmediatePropagation(),B(()=>Z[j])};return Z.addEventListener(oZ,J),()=>Z.removeEventListener(oZ,J)}),A4=(Z)=>($,J)=>{let X=()=>J;return Z.dispatchEvent(new B4($,(j)=>{X=j})),v(X)};var NZ=typeof process<"u"&&!1,FZ="warn",t4=(Z)=>Z?`#${Z}`:"",n4=(Z)=>Z?.length?`.${Array.from(Z).join(".")}`:"",Q4=(Z)=>Z.localName.includes("-"),H4=(Z)=>Q4(Z)&&Z.matches(":not(:defined)"),o=(Z)=>Z?`<${Z.localName}${t4(Z.id)}${n4(Z.classList)}>`:"<unknown>";class lZ extends TypeError{constructor(Z){super(`Invalid component name "${Z}". Custom element names must contain a hyphen, start with a lowercase letter, and contain only lowercase letters, numbers, and hyphens.`);this.name="InvalidComponentNameError"}}class w4 extends TypeError{constructor(Z,$,J){super(`Invalid property name "${$}" for component <${Z}>. ${J}`);this.name="InvalidPropertyNameError"}}class EZ extends Error{constructor(Z,$,J){super(`Missing required element <${$}> in component ${o(Z)}. ${J}`);this.name="MissingElementError"}}class sZ extends Error{constructor(Z,$){super(`Timeout waiting for: [${$.join(", ")}] in component ${o(Z)}.`);this.name="DependencyTimeoutError"}}class rZ extends TypeError{constructor(Z,$,J){super(`Expected reactives passed from ${o(Z)} to ${o($)} to be a record of signals, reactive property names or functions. Got ${zZ(J)}.`);this.name="InvalidReactivesError"}}class tZ extends TypeError{constructor(Z,$){super(`Target ${o(Z)} is not a custom element in ${$}.`);this.name="InvalidCustomElementError"}}var L4=new WeakMap,yZ=(Z)=>{let $=L4.get(Z);if(!$)$={},L4.set(Z,$);return $};var nZ=(Z)=>{for(let $ of Z)if(Array.isArray($))nZ($);else if($)$()},z4=(Z,$)=>{if(O($))return wZ($);if(typeof $==="string"){let J=yZ(Z)[$];if(J)return J;return v(()=>Z[$])}if($&&typeof $==="object"&&"get"in $&&!(Symbol.toStringTag in $))return $;return $},_4=(Z)=>{function $(J,X){return()=>{if(Array.isArray(J)){let B=J.map((G)=>z4(Z,G)),q=X;return r(()=>RZ(B,{ok:(G)=>f(()=>q(G))}))}let j=z4(Z,J);if(typeof X==="function")return r(()=>RZ(j,{ok:(B)=>f(()=>X(B))}));return r(()=>RZ(j,X))}}return $},b4=(Z)=>{let $=(X,j)=>E(()=>{if(!Q4(X))throw new tZ(X,`pass from ${o(Z)}`);if(!L(j))throw new rZ(Z,X,j);let B=yZ(X),q=o(X),G=[];for(let[z,W]of Object.entries(j)){if(W==null)continue;if(!(z in X)){if(NZ)console[FZ](`pass(): property '${z}' does not exist on ${q}`);continue}let V=z4(Z,W);if(!V)continue;let U=B[z];if(PZ(U)){let D=U.current();U.replace(V),G.push(()=>U.replace(D));continue}if(NZ)console[FZ](`pass(): property '${z}' on ${q} is not Slot-backed — use setProperty() for non-Le Truc elements`)}if(G.length)return()=>{for(let z of G)z()}});function J(X,j){return()=>{if(!X)return;if(BZ(X))r(()=>{for(let B of X.get())E(()=>$(B,j))});else $(X,j)}}return J};function a4(Z,$){return()=>{r(()=>{for(let J of Z.get())E(()=>{let X=$(J);if(Array.isArray(X))nZ(X);else if(typeof X==="function")X()})})}}var q4=new Set,f4=new WeakMap,aZ=new Set,U4,e4=()=>{U4=void 0;let Z=Array.from(q4);q4.clear();for(let J of Z)f4.get(J)?.();let $=Array.from(aZ);aZ.clear();for(let J of $)J()},T4=()=>{if(!U4)U4=requestAnimationFrame(e4)},M4=(Z,$)=>{f4.set(Z,$),q4.add(Z),T4()},eZ=(Z,$)=>{let J=!1,X,j=()=>{J=!1,Z(...X)},B=(...q)=>{if(X=q,J)return;J=!0,aZ.add(j),T4()};return B.cancel=()=>{if(J)aZ.delete(j),J=!1},$?.addEventListener("abort",B.cancel,{once:!0}),B};var Z6=new Set(["scroll","resize","mousewheel","touchstart","touchmove","wheel"]),$6=new Set(["focus","blur","scroll","resize","load","unload","error","toggle","mouseenter","mouseleave","pointerenter","pointerleave","abort","canplay","canplaythrough","durationchange","emptied","ended","loadeddata","loadedmetadata","loadstart","pause","play","playing","progress","ratechange","seeked","seeking","stalled","suspend","timeupdate","volumechange","waiting"]),S4=(Z,$,J,X,j)=>{let B=(G)=>{let z=X(G,$);if(!L(z))return;h(()=>{for(let[W,V]of Object.entries(z))Z[W]=V})},q=j.passive?eZ(B):B;return $.addEventListener(J,q,j),()=>{$.removeEventListener(J,q),q.cancel?.()}},E4=(Z)=>{function $(J,X,j,B={}){return()=>{if(!J)return;if(!("passive"in B))B={...B,passive:Z6.has(X)};if(BZ(J)){if($6.has(X)){if(NZ)console[FZ](`on(): '${X}' does not bubble — prefer each() + on() for per-element listeners in ${o(Z)}`);return r(()=>{for(let W of J.get())E(()=>{return S4(Z,W,X,j,B)})})}let q=Z.shadowRoot??Z,G=(W)=>{let V=W.composedPath();for(let U of J.get())if(V.includes(U)){let D=j(W,U);if(!L(D))break;h(()=>{for(let[Q,M]of Object.entries(D))Z[Q]=M});break}},z=B.passive?eZ(G):G;E(()=>{return q.addEventListener(X,z,B),()=>{q.removeEventListener(X,z),z.cancel?.()}});return}E(()=>S4(Z,J,X,j,B))}}return $};var J6=200,X6=(Z)=>{let $=new Set,J="",X=0;for(let j of Z)if(j==="[")X++;else if(j==="]"){if(X>0)X--}else if(X===0)J+=j;if(J.includes("."))$.add("class");if(J.includes("#"))$.add("id");if(Z.includes("[")){let j=Z.split("[");for(let B=1;B<j.length;B++){let q=j[B];if(!q||!q.includes("]"))continue;let G=q.split("=")[0].split("]")[0].trim().replace(/[^a-zA-Z0-9_-]/g,"");if(G)$.add(G)}}return[...$]};function y4(Z,$){return v(()=>Array.from(Z.querySelectorAll($)),{value:[],equals:(J,X)=>J.length===X.length&&J.every((j,B)=>j===X[B]),watched:(J)=>{let X={childList:!0,subtree:!0},j=X6($);if(j.length)X.attributes=!0,X.attributeFilter=j;let B=(z)=>z instanceof Element&&(z.matches($)||z.querySelector($)),q=(z)=>{if(z.type==="attributes")return!0;if(z.type==="childList")return Array.from(z.addedNodes).some(B)||Array.from(z.removedNodes).some(B);return!1},G=new MutationObserver((z)=>{for(let W of z)if(q(W)){J();return}});return G.observe(Z,X),()=>G.disconnect()}})}var m4=(Z)=>{let $=Z.shadowRoot??Z,J=new Set;function X(q,G){let z=$.querySelector(q);if(G!=null&&!z)throw new EZ(Z,q,G);if(z&&H4(z))J.add(z.localName);return z??void 0}function j(q,G){let z=y4($,q),W=z.get();if(G!=null&&!W.length)throw new EZ(Z,q,G);if(W.length){for(let V of W)if(H4(V))J.add(V.localName)}return z}return[{first:X,all:j},(q)=>{if(J.size)queueMicrotask(()=>{let G=Array.from(J).filter((z)=>!customElements.get(z));if(!G.length){q();return}Promise.race([Promise.all(G.map((z)=>customElements.whenDefined(z))),new Promise((z,W)=>{setTimeout(()=>{W(new sZ(Z,G.filter((V)=>!customElements.get(V))))},J6)})]).then(q).catch((z)=>{if(NZ)console[FZ](z);q()})});else q()}]};var h4=Symbol("parser"),k4=Symbol("method"),p4=(Z)=>O(Z)&&(h4 in Z),g4=(Z)=>O(Z)&&(k4 in Z),c=(Z)=>Object.assign(Z,{[h4]:!0}),j6=(Z)=>Object.assign(Z,{[k4]:!0});function W6(Z,$){if(!Z.includes("-")||!Z.match(/^[a-z][a-z0-9-]*$/))throw new lZ(Z);class J extends HTMLElement{debug;#J=!1;#Z=[];#$;connectedCallback(){let X=()=>{this.#$=E(()=>nZ(this.#Z),{root:!0})};if(this.#J)X();else{let j=this,[B,q]=m4(this),G={expose:this.#j.bind(this),host:j,...B,watch:_4(j),on:E4(j),pass:b4(j),provideContexts:R4(j),requestContext:A4(j)},z=$(G);if(z)this.#Z=z;if(this.#J=!0,!this.#Z.length)return;q(X)}}disconnectedCallback(){if(O(this.#$))this.#$()}#j(X){let j=(B,q)=>{if(p4(q)){let G=q(this.getAttribute(B));if(G!=null)this.#X(B,G)}else if(g4(q))this[B]=q;else{let G=q;if(G!=null)this.#X(B,G)}};for(let[B,q]of Object.entries(X)){if(q==null||B in this)continue;j(B,q)}}#X(X,j){let B=QZ(j)?j:O(j)?wZ(j):g(j),q=yZ(this),G=X,z=q[G];if(PZ(z))z.replace(B);else if(SZ(B)){let W=iZ(B);q[G]=W,Object.defineProperty(this,X,W)}else q[G]=B,Object.defineProperty(this,X,{get:B.get,enumerable:!0})}}return customElements.define(Z,J),customElements.get(Z)}var B6=["type","src","async","defer","nomodule","crossorigin","integrity","referrerpolicy","fetchpriority"],Q6=(Z)=>{if(/^(javascript|data|vbscript):/i.test(Z))return!1;if(/^(mailto|tel):/i.test(Z))return!0;if(Z.includes("://"))try{let $=new URL(Z);return["http:","https:","ftp:"].includes($.protocol)}catch{return!1}return!0},v4=(Z,$,J)=>{if(/^on/i.test($))throw Error(`setAttribute: blocked unsafe attribute name '${$}' on ${Z.localName} — event handler attributes are not allowed`);if(J=String(J).trim(),!Q6(J))throw Error(`setAttribute: blocked unsafe value for '${$}' on <${Z.localName}>: '${J}'`);Z.setAttribute($,J)},H6=(Z)=>Z.replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;").replace(/"/g,"&quot;").replace(/'/g,"&#39;"),d4=(Z,$)=>{Array.from(Z.childNodes).filter((J)=>J.nodeType!==Node.COMMENT_NODE).forEach((J)=>{J.remove()}),Z.append(document.createTextNode($))},z6=(Z,$=!1)=>$?(J)=>d4(Z,String(J)):(J)=>{Z.textContent=String(J)},q6=(Z,$)=>(J)=>{Z[$]=J},U6=(Z,$)=>(J)=>{Z.classList.toggle($,Boolean(J))},M6=(Z)=>($)=>{Z.hidden=!$},V6=(Z,$,J=!1)=>({ok:(X)=>{if(typeof X==="boolean")Z.toggleAttribute($,X);else if(J)Z.setAttribute($,X);else v4(Z,$,X)},nil:()=>{Z.removeAttribute($)}}),G6=(Z,$)=>({ok:(J)=>{Z.style.setProperty($,J)},nil:()=>{Z.style.removeProperty($)}}),K6=(Z,$={})=>{let J=()=>{if(Z.shadowRoot)Z.shadowRoot.innerHTML="<slot></slot>";else Z.innerHTML=""};return{ok:(X)=>{if(!X){J();return}let{shadowRootMode:j,allowScripts:B}=$;if(j&&!Z.shadowRoot)Z.attachShadow({mode:j});let q=Z.shadowRoot||Z;M4(Z,()=>{if(q.innerHTML=X,B)q.querySelectorAll("script").forEach((G)=>{let z=document.createElement("script");for(let W of B6){let V=G.getAttribute(W);if(V!==null)z.setAttribute(W,V)}if(!G.hasAttribute("src"))z.appendChild(document.createTextNode(G.textContent??""));q.appendChild(z),G.remove()})})},nil:J}};var D6=()=>c((Z)=>Z!=null&&Z!=="false");var P6=(Z)=>c(($)=>{if(($??Z)==null)throw TypeError("asJSON: Value and fallback are both null or undefined");if($==null)return Z;if($==="")throw SyntaxError("Empty string is not valid JSON");let J;try{J=JSON.parse($)}catch(X){throw SyntaxError(`Failed to parse JSON: ${String(X)}`,{cause:X})}return J??Z});var mZ=(Z,$)=>{if($==null)return;let J=Z($);return Number.isFinite(J)?J:void 0},N6=(Z=0)=>c(($)=>{if($==null)return Z;let J=$.trim();if(J.toLowerCase().startsWith("0x"))return mZ((j)=>parseInt(j,16),J)??Z;let X=mZ(parseFloat,$);return X!=null?Math.trunc(X):Z}),F6=(Z=0)=>c(($)=>mZ(parseFloat,$)??Z),C6=(Z=0,$=Number.MAX_SAFE_INTEGER)=>c((J)=>{if(J==null)return Z;let X=J.trim(),j=X.toLowerCase().startsWith("0x")?mZ((q)=>parseInt(q,16),X):mZ(parseFloat,J),B=j!=null?Math.trunc(j):Z;return Math.max(Z,Math.min(B,$))});var Y6=(Z="")=>c(($)=>$??Z),x6=(Z)=>c(($)=>{if($==null)return Z[0];let J=$.toLowerCase();return Z.find((j)=>j.toLowerCase()===J)??Z[0]});export{zZ as valueString,f as untrack,D4 as unown,eZ as throttle,d4 as setTextPreservingComments,M4 as schedule,v4 as safeSetAttribute,RZ as match,DZ as isTask,TZ as isStore,OZ as isState,PZ as isSlot,w as isSignalOfType,QZ as isSignal,C4 as isSensor,L as isRecord,p4 as isParser,SZ as isMutableSignal,g4 as isMethodProducer,BZ as isMemo,cZ as isList,O as isFunction,I4 as isComputed,N4 as isCollection,n as isAsyncFunction,H6 as escapeHTML,a4 as each,j6 as defineMethod,W6 as defineComponent,K6 as dangerouslyBindInnerHTML,KZ as createTask,AZ as createStore,g as createState,iZ as createSlot,Y4 as createSignal,F4 as createSensor,E as createScope,x4 as createMutableSignal,v as createMemo,GZ as createList,y4 as createElementsMemo,r as createEffect,wZ as createComputed,P4 as createCollection,M6 as bindVisible,z6 as bindText,G6 as bindStyle,q6 as bindProperty,U6 as bindClass,V6 as bindAttribute,h as batch,Y6 as asString,c as asParser,F6 as asNumber,P6 as asJSON,N6 as asInteger,x6 as asEnum,C6 as asClampedInteger,D6 as asBoolean,a as UnsetSignalValueError,gZ as SKIP_EQUALITY,bZ as RequiredOwnerError,_Z as ReadonlySignalError,hZ as NullishSignalValueError,EZ as MissingElementError,qZ as InvalidSignalValueError,rZ as InvalidReactivesError,w4 as InvalidPropertyNameError,tZ as InvalidCustomElementError,lZ as InvalidComponentNameError,kZ as InvalidCallbackError,sZ as DependencyTimeoutError,p as DEFAULT_EQUALITY,s as DEEP_EQUALITY,B4 as ContextRequestEvent,LZ as CircularDependencyError,oZ as CONTEXT_REQUEST};
 
-//# debugId=B601AA3F770D926264756E2164756E21
+//# debugId=934CAE2C68EE3C0E64756E2164756E21
