+++ src/tests/parsers.test.ts
@@ -1,5 +1,5 @@
 /**
- * Unit tests for src/parsers/* and src/safety.ts
+ * Unit tests for src/parsers/*.ts
  *
  * Pure functions only — no DOM required.
  */
@@ -8,63 +8,6 @@ import { describe, expect, test } from 'bun:test'
 import { asJSON } from '../parsers/json'
 import { asClampedInteger, asInteger, asNumber } from '../parsers/number'
 import { asEnum, asString } from '../parsers/string'
-import { safeSetAttribute } from '../safety'
-
-/* === safety.ts === */
-
-describe('safeSetAttribute', () => {
-	const makeEl = () => {
-		const attrs: Record<string, string> = {}
-		return {
-			localName: 'a',
-			setAttribute: (attr: string, val: string) => {
-				attrs[attr] = val
-			},
-			_attrs: attrs,
-		} as unknown as Element & { _attrs: Record<string, string> }
-	}
-
-	test('blocks javascript: URIs', () => {
-		expect(() =>
-			safeSetAttribute(makeEl(), 'href', 'javascript:alert(1)'),
-		).toThrow()
-	})
-
-	test('blocks data: URIs', () => {
-		expect(() =>
-			safeSetAttribute(makeEl(), 'href', 'data:text/html,<h1>XSS</h1>'),
-		).toThrow()
-	})
-
-	test('blocks vbscript: URIs', () => {
-		expect(() =>
-			safeSetAttribute(makeEl(), 'href', 'vbscript:MsgBox(1)'),
-		).toThrow()
-	})
-
-	test('blocks on* event handler attributes', () => {
-		expect(() => safeSetAttribute(makeEl(), 'onclick', 'alert(1)')).toThrow()
-	})
-
-	test('allows https: URIs', () => {
-		const el = makeEl()
-		expect(() =>
-			safeSetAttribute(el, 'href', 'https://example.com'),
-		).not.toThrow()
-	})
-
-	test('allows mailto: URIs', () => {
-		const el = makeEl()
-		expect(() =>
-			safeSetAttribute(el, 'href', 'mailto:foo@example.com'),
-		).not.toThrow()
-	})
-
-	test('allows relative paths', () => {
-		const el = makeEl()
-		expect(() => safeSetAttribute(el, 'href', '/page')).not.toThrow()
-	})
-})
 
 /* === parsers/string.ts === */
