--- types/src/safety.d.ts
@@ -1,35 +1,0 @@
-/**
- * Set an attribute on an element with security validation.
- *
- * Blocks `on*` event handler attributes and validates URL-like values against
- * a safe-protocol allowlist (`http:`, `https:`, `ftp:`, `mailto:`, `tel:`).
- * Violations throw a descriptive error — they are never silent.
- *
- * @since 1.1
- * @param {Element} element - Target element
- * @param {string} attr - Attribute name to set
- * @param {string} value - Attribute value to set
- */
-declare const safeSetAttribute: (element: Element, attr: string, value: string) => void;
-/**
- * Escape HTML entities to prevent XSS when inserting user-supplied text as HTML.
- *
- * Escapes `&`, `<`, `>`, `"`, and `'`.
- *
- * @since 1.1
- * @param {string} text - Plain text to escape
- * @returns {string} HTML-safe string
- */
-declare const escapeHTML: (text: string) => string;
-/**
- * Set the text content of an element while preserving comment nodes.
- *
- * Removes all child nodes except comments, then appends a new text node.
- * Useful when HTML comments are used as markers or server-rendered annotations.
- *
- * @since 1.1
- * @param {Element} element - Target element
- * @param {string} text - Text content to set
- */
-declare const setTextPreservingComments: (element: Element, text: string) => void;
-export { escapeHTML, safeSetAttribute, setTextPreservingComments };
